### 2024-02-17 #6364 @editorialbot
[PRE REVIEW]: ATHENA: A Fortran package for neural networks
**Submitting author:** @nedtaylor (<a href="http://orcid.org/0000-0002-9134-9712">Ned Thaddeus Taylor</a>)
**Repository:** https://github.com/nedtaylor/athena
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.1
