### 2024-02-17 #6364 @arfon
@nedtaylor – I note that this submission appears to be on an institutionally-hosted GitLab instance. Can you confirm that this environment meets this condition around account creation to provide feedback? https://joss.readthedocs.io/en/latest/review_criteria.html#where-can-i-host-the-repository
