### 2024-02-17 #6364 @nedtaylor
> @nedtaylor – I note that this submission appears to be on an institutionally-hosted GitLab instance. Can you confirm that this environment meets this condition around account creation to provide feedback? https://joss.readthedocs.io/en/latest/review_criteria.html#where-can-i-host-the-repository

Hi Arfon,

Thank you for bringing this to my attention. Unfortunately, you are correct that my institutionally-hosted GitLab instance does not allow user creation or setting up issues without first logging in (I thought that it had and believed I had achieved this successfully earlier, but was clearly mistaken).

Am I allowed to make a fork of my code on the general GitHub and then use that here instead? Will there be an issue with that? I assume that may require making a whole new submission, but as long as that's allowed, I will go ahead and do that.
