+++ fpm.toml
@@ -1,5 +1,5 @@
 name = "athena"
-version = "1.2.1"
+version = "1.2.2"
 license = "MIT"
 author = "Ned Thaddeus Taylor"
 maintainer = "n.t.taylor@exeter.ac.uk"
