### 2024-02-18 2a4f52d
Update version to 1.2.2
