+++ paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
  - name: Department of Physics and Astronomy, University of Exeter, United Kingdom, EX4 4QL
    index: 1
-date: 8 February 2024
+date: 18 February 2024
 bibliography: paper.bib
 
 
