+++ paper.bib
@@ -1,28 +1,49 @@
 @article{curcic2019parallel,
-      title={A parallel Fortran framework for neural networks and deep learning}, 
-      author={Milan Curcic},
-      year={2019},
-      eprint={1902.06714},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG}
+      author = {Curcic, Milan},
+      title = {A parallel Fortran framework for neural networks and deep learning},
+      year = {2019},
+      issue_date = {December 2018},
+      publisher = {Association for Computing Machinery},
+      address = {New York, NY, USA},
+      volume = {38},
+      number = {1},
+      issn = {1061-7264},
+      url = {https://doi.org/10.1145/3323057.3323059},
+      doi = {10.1145/3323057.3323059},
+      abstract = {This paper describes neural-fortran, a parallel Fortran framework for neural networks and deep learning. It features a simple interface to construct feed-forward neural networks of arbitrary structure and size, several activation functions, and stochastic gradient descent as the default optimization algorithm. Neural-fortran also leverages the Fortran 2018 standard collective subroutines to achieve data-based parallelism on shared- or distributed-memory machines. First, I describe the implementation of neural networks with Fortran derived types, whole-array arithmetic, and collective sum and broadcast operations to achieve parallelism. Second, I demonstrate the use of neural-fortran in an example of recognizing hand-written digits from images. Finally, I evaluate the computational performance in both serial and parallel modes. Ease of use and computational performance are similar to an existing popular machine learning framework, making neural-fortran a viable candidate for further development and use in production.},
+      journal = {SIGPLAN Fortran Forum},
+      month = {mar},
+      pages = {4–21},
+      numpages = {18}
 }
 
-@article{ghiasi2018dropblock,
-      title={DropBlock: A regularization method for convolutional networks}, 
-      author={Golnaz Ghiasi and Tsung-Yi Lin and Quoc V. Le},
-      year={2018},
-      eprint={1810.12890},
-      archivePrefix={arXiv},
-      primaryClass={cs.CV}
+@inproceedings{ghiasi2018dropblock,
+      author = {Ghiasi, Golnaz and Lin, Tsung-Yi and Le, Quoc V.},
+      title = {DropBlock: a regularization method for convolutional networks},
+      year = {2018},
+      publisher = {Curran Associates Inc.},
+      address = {Red Hook, NY, USA},
+      abstract = {Deep neural networks often work well when they are over-parameterized and trained with a massive amount of noise and regularization, such as weight decay and dropout. Although dropout is widely used as a regularization technique for fully connected layers, it is often less effective for convolutional layers. This lack of success of dropout for convolutional layers is perhaps due to the fact that activation units in convolutional layers are spatially correlated so information can still flow through convolutional networks despite dropout. Thus a structured form of dropout is needed to regularize convolutional networks. In this paper, we introduce DropBlock, a form of structured dropout, where units in a contiguous region of a feature map are dropped together. We found that applying DropbBlock in skip connections in addition to the convolution layers increases the accuracy. Also, gradually increasing number of dropped units during training leads to better accuracy and more robust to hyperparameter choices. Extensive experiments show that DropBlock works better than dropout in regularizing convolutional networks. On ImageNet classification, ResNet-50 architecture with DropBlock achieves 78.13\% accuracy, which is more than 1.6\% improvement on the baseline. On COCO detection, DropBlock improves Average Precision of RetinaNet from 36.8\% to 38.4\%.},
+      booktitle = {Proceedings of the 32nd International Conference on Neural Information Processing Systems},
+      pages = {10750–10760},
+      numpages = {11},
+      location = {Montr\'{e}al, Canada},
+      doi = {10.5555/3327546.3327732},
+      series = {NIPS'18}
 }
 
-@article{ioffe2015batch,
[…]
