### 2024-02-18 #6364 @nedtaylor
Hi @arfon,

I have moved my repository to GitHub. The new repository link is: https://github.com/nedtaylor/athena

I have also updated the version number to: 1.2.2

The branch name has not changed though, that is still: paper

Are these changes ones that you are allowed to implement? Hopefully the changes now allow the community to engage and provide feedback.
