### 2024-02-18 b4124e2
Add JOSS badge
