### 2024-02-26 #6364 @nedtaylor
> @nedtaylor – I note that this submission appears to be on an institutionally-hosted GitLab instance. Can you confirm that this environment meets this condition around account creation to provide feedback? https://joss.readthedocs.io/en/latest/review_criteria.html#where-can-i-host-the-repository

Hi @arfon. Just wanted to check if there is anything else I need to do at this moment in time?
