### 2024-02-29 c695687
Add avgpool3d unit test
