+++ test/test_batchnorm3d_layer.f90
@@ -1,6 +1,6 @@
-program test_batchnorm2d_layer
+program test_batchnorm3d_layer
   use athena, only: &
-     batchnorm2d_layer_type, &
+     batchnorm3d_layer_type, &
      base_layer_type, &
      learnable_layer_type
   implicit none
@@ -8,7 +8,7 @@ program test_batchnorm2d_layer
   class(base_layer_type), allocatable :: bn_layer
   integer, parameter :: num_channels = 3, width = 8, batch_size = 1
   real, parameter :: gamma  = 0.5, beta = 0.3
-  real, allocatable, dimension(:,:,:,:) :: input_data, output, gradient
+  real, allocatable, dimension(:,:,:,:,:) :: input_data, output, gradient
   real, parameter :: tol = 0.5E-3
   logical :: success = .true.
 
@@ -23,9 +23,9 @@ program test_batchnorm2d_layer
   allocate(seed(seed_size), source=0)
   call random_seed(put = seed)
 
-  !! set up batchnorm2d layer
-  bn_layer = batchnorm2d_layer_type( &
-     input_shape = [width, width, num_channels], &
+  !! set up batchnorm3d layer
+  bn_layer = batchnorm3d_layer_type( &
+     input_shape = [width, width, width, num_channels], &
      batch_size = batch_size, &
      momentum = 0.0, &
      epsilon = 1e-5, &
@@ -40,40 +40,40 @@ program test_batchnorm2d_layer
      )
 
   !! check layer name
-  if(.not. bn_layer%name .eq. 'batchnorm2d')then
+  if(.not. bn_layer%name .eq. 'batchnorm3d')then
     success = .false.
-    write(0,*) 'batchnorm2d layer has wrong name'
+    write(0,*) 'batchnorm3d layer has wrong name'
   end if
 
   !! check layer type
   select type(bn_layer)
-  type is(batchnorm2d_layer_type)
+  type is(batchnorm3d_layer_type)
     !! check input shape
-    if(any(bn_layer%input_shape .ne. [width,width,num_channels]))then
+    if(any(bn_layer%input_shape .ne. [width,width,width,num_channels]))then
       success = .false.
-      write(0,*) 'batchnorm2d layer has wrong input_shape'
+      write(0,*) 'batchnorm3d layer has wrong input_shape'
     end if
 
     !! check output shape
-    if(any(bn_layer%output_shape .ne. [width,width,num_channels]))then
+    if(any(bn_layer%output_shape .ne. [width,width,width,num_channels]))then
       success = .false.
-      write(0,*) 'batchnorm2d layer has wrong output_shape'
+      write(0,*) 'batchnorm3d layer has wrong output_shape'
     end if
 
     !! check batch size
     if(bn_layer%batch_size .ne. 1)then
       success = .false.
-      write(0,*) 'batchnorm2d layer has wrong batch size'
+      write(0,*) 'batchnorm3d layer has wrong batch size'
     end if
   class default
     success = .false.
-    write(0,*) 'batchnorm2d layer has wrong type'
+    write(0,*) 'batchnorm3d layer has wrong type'
   end select
 
 !!!-----------------------------------------------------------------------------
 
   !! initialise sample input
-  allocate(input_data(width, width, num_channels, batch_size), source = 0.0)
+  allocate(input_data(width,width,width,num_channels,batch_size), source = 0.0)
   
    input_data = max_value
 
@@ -84,7 +84,7 @@ program test_batchnorm2d_layer
   !! check outputs all get normalised to zero
   if (any(output-beta.gt. tol)) then
     success = .false.
-    write(0,*) 'batchnorm2d layer forward pass failed: &
+    write(0,*) 'batchnorm3d layer forward pass failed: &
[…]
