### 2024-03-02 1b68647
Add shuffle unit test
