### 2024-03-02 b40f29b
Fix kernel size and seed
