+++ test/test_shuffle.f90
@@ -2,18 +2,29 @@ program test_shuffle
   use misc_ml, only: shuffle
   implicit none
 
-  integer :: i, j, seed_size
+  integer :: i, j, k, l, seed_size
   integer, allocatable, dimension(:) :: iseed
   real :: r
   
   integer, parameter :: n = 3
   integer, parameter :: m = 5
+  integer, parameter :: p = 5
+  integer, parameter :: q = 5
 
 
   integer, dimension(n) :: array, original_array, array_tmp
 
   real, dimension(n, m) :: array_2d_original, array_2d_shuffled, &
        array_2d_shuffled_tmp
+
+  real :: array_3d_original(n, m, p)
+  real :: array_3d_shuffled(n, m, p)
+  real :: array_3d_shuffled_tmp(n, m, p)
+
+  real :: array_4d_original(n, m, p, q)
+  real :: array_4d_shuffled(n, m, p, q)
+  real :: array_4d_shuffled_tmp(n, m, p, q)
+
   logical :: success = .true.
   
 
@@ -99,11 +110,118 @@ program test_shuffle
 
   !! Check if array_2d_shuffled and array_2d_shuffled_tmp are the same
   if (any(abs(array_2d_shuffled_tmp - array_2d_shuffled).gt.1.E-6)) then
-     write(*,*) array_2d_shuffled_tmp
-       write(*,*) array_2d_shuffled
      write(*,*) '2D Array shuffle seed does not work as intended'
      success = .false.
   end if
+
+
+!!!-----------------------------------------------------------------------------
+!!! 3D array shuffle tests
+!!!-----------------------------------------------------------------------------
+
+  !! test 3D array shuffling
+  do i = 1, n
+     do j = 1, m
+        do k = 1, p
+           array_3d_original(i, j, k) = i * m * p + j * p + k
+        end do
+     end do
+  end do
+
+  !! Shuffle the 3D array along the third dimension
+  array_3d_shuffled = array_3d_original
+  call shuffle(array_3d_shuffled, dim = 3, seed = 1)
+
+  !! Check if the 3D array is shuffled along the third dimension
+  if (all(array_3d_shuffled .eq. array_3d_original)) then
+     write(*,*) '3D Array is not shuffled along the third dimension'
+     success = .false.
+  end if
+
+  !! Check if all original elements are still in the shuffled 3D array
+  do i = 1, n
+    do j = 1, m
+       do k = 1, p
+          if ( all(abs(array_3d_shuffled(i, j, :) - &
+               array_3d_original(i, j, k)).gt.1.E-6) ) then
+             write(*,*) 'Original element', array_3d_original(i, j, k), &
+                  'is missing in the shuffled 3D array'
+             write(*,*) array_3d_shuffled(i, j, :)
+             write(*,*) array_3d_original(i, j, k)
+             success = .false.
+          end if
+       end do
+    end do
+  end do
+
+  !! Check that seed works for 3D array shuffling
+  array_3d_shuffled_tmp = array_3d_original
+  call shuffle(array_3d_shuffled_tmp, dim = 3, seed = 4)
+  array_3d_shuffled = array_3d_original
+  call shuffle(array_3d_shuffled, dim = 3, seed = 4)
+
+  !! Check if array_3d_shuffled and array_3d_shuffled_tmp are the same
+  if (any(abs(array_3d_shuffled_tmp - array_3d_shuffled).gt.1.E-6)) then
+     write(*,*) '3D Array shuffle seed does not work as intended'
+     success = .false.
+  end if
+
+
+  !!!-----------------------------------------------------------------------------
+  !!! 4D array shuffle tests
+  !!!-----------------------------------------------------------------------------
+  
+    !! test 3D array shuffling
+    do i = 1, n
+       do j = 1, m
+          do k = 1, p
[…]
