+++ src/lib/mod_misc_ml.f90
@@ -17,20 +17,47 @@ module misc_ml
   implicit none
 
 
+!!!-----------------------------------------------------
+!!! shuffle an array along one dimension
+!!! optional index array is also shuffled
+!!!-----------------------------------------------------
+!!! data  = (I/R, io) input array to be shuffled
+!!! dim   = (I, in) dimension along which to shuffle
+!!! label = (I, io, opt) index array to be shuffled
+!!! seed  = (I, in, opt) random seed
   interface shuffle
-  procedure shuffle_1Dlist, &
-       shuffle_2Ddata, shuffle_3Didata, shuffle_3Drdata, &
-       shuffle_4Ddata, shuffle_5Ddata, &
+  procedure shuffle_1Dilist, &
+       shuffle_2Drdata, shuffle_3Didata, shuffle_3Drdata, &
+       shuffle_4Drdata, shuffle_5Drdata, &
        shuffle_3Didata_1Dilist, shuffle_3Didata_1Drlist, &
-       shuffle_4Ddata_1Dlist, shuffle_5Ddata_1Dilist, shuffle_5Ddata_1Drlist
+       shuffle_4Drdata_1Dlist, shuffle_5Drdata_1Dilist, shuffle_5Drdata_1Drlist
   end interface shuffle
-  
+!!!=====================================================
+
+
+!!!-----------------------------------------------------
+!!! split an array into train and test sets
+!!! optional index array is also split
+!!!-----------------------------------------------------
+!!! data        = (I/R, in ) input array to be shuffled
+!!! left_data   = (I/R, out) left split of data
+!!! right_data  = (I/R, out) right split of data
+!!! dim         = (I, in ) dimension along which to split
+!!! label       = (I, in , opt) index array to be shuffled
+!!! left_label  = (I, out, opt) left split of label
+!!! right_label = (I, out, opt) right split of label
+!!! left_size   = (I, in , opt) size of left split
+!!! right_size  = (I, in , opt) size of right split
+!!! shuffle     = (B, in , opt) shuffle data before splitting
+!!! seed        = (I, in , opt) random seed
+!!! split_list  = (I, out, opt) index array of split
   interface split
   procedure &
        split_3Didata_1Dilist, split_3Didata_1Drlist, &
        split_5Drdata, &
        split_5Drdata_1Drlist
   end interface split
+!!!=====================================================
 
 
   private
@@ -46,7 +73,7 @@ contains
 !!!#####################################################
 !!! shuffle an array along one dimension
 !!!#####################################################
-subroutine shuffle_1Dlist(data,seed)
+subroutine shuffle_1Dilist(data,seed)
 implicit none
 integer :: istart, num_data, seed_size
 integer :: itmp1, i, j
@@ -66,6 +93,7 @@ else
    call random_seed(get=iseed)
 end if
 
+!! shuffle the data
 num_data = size(data,dim=1)
 istart=1
 do i=1,num_data
@@ -77,10 +105,10 @@ do i=1,num_data
   data(i) = itmp1
 end do
 
-end subroutine shuffle_1Dlist
+end subroutine shuffle_1Dilist
 !!!-----------------------------------------------------
 !!!-----------------------------------------------------
-subroutine shuffle_2Ddata(data,dim,seed)
+subroutine shuffle_2Drdata(data,dim,seed)
 implicit none
 integer :: istart, seed_size
 integer :: i,j,n_data,iother
@@ -104,6 +132,7 @@ else
[…]
