+++ src/lib/mod_misc_ml.f90
@@ -110,7 +110,7 @@ end subroutine shuffle_1Dilist
 !!!-----------------------------------------------------
 subroutine shuffle_2Drdata(data,dim,seed)
 implicit none
-integer :: istart, seed_size
+integer :: istart,seed_size
 integer :: i,j,n_data,iother
 integer :: i1s,i2s,i1e,i2e,j1s,j2s,j1e,j2e
 real(real12) :: r
@@ -167,12 +167,12 @@ end subroutine shuffle_2Drdata
 !!!-----------------------------------------------------
 subroutine shuffle_3Drdata(data,dim,seed)
 implicit none
-integer :: istart, seed_size
+integer :: istart,seed_size
 integer :: i,j,n_data
 real(real12) :: r
-integer, allocatable, dimension(:) :: iseed
 integer, dimension(3) :: idx_s,idx_e,jdx_s,jdx_e
 integer, dimension(3,2) :: t_size
+integer, allocatable, dimension(:) :: iseed
 real(real12), allocatable, dimension(:,:,:) :: tlist
 
 integer, intent(in) :: dim
@@ -243,11 +243,12 @@ end subroutine shuffle_3Drdata
 !!!-----------------------------------------------------
 subroutine shuffle_3Didata(data,dim,seed)
  implicit none
- integer :: iseed,istart
+ integer :: istart,seed_size
  integer :: i,j,n_data
  real(real12) :: r
  integer, dimension(3) :: idx_s,idx_e,jdx_s,jdx_e
  integer, dimension(3,2) :: t_size
+ integer, allocatable, dimension(:) :: iseed
  integer, allocatable, dimension(:,:,:) :: tlist
 
  integer, intent(in) :: dim
@@ -255,9 +256,18 @@ subroutine shuffle_3Didata(data,dim,seed)
 
  integer, optional, intent(in) :: seed
 
- if(present(seed)) iseed = seed
 
- call random_seed(iseed)
+ !! set or get random seed
+ call random_seed(size=seed_size)
+ allocate(iseed(seed_size))
+ if(present(seed))then
+    iseed = seed
+    call random_seed(put=iseed)
+ else
+    call random_seed(get=iseed)
+ end if
+
+ !! get the size of the data
  n_data = size(data,dim=dim)
  do i=1,3
     t_size(i,1) = 1
@@ -271,9 +281,9 @@ subroutine shuffle_3Didata(data,dim,seed)
        t_size(i,2) = size(data,dim=i)
     end if
  end do
-
  allocate(tlist(t_size(1,2),t_size(2,2),t_size(3,2)))
 
+ !! shuffle the data
  istart=1
  do i=1,n_data
     call random_number(r)
@@ -311,18 +321,19 @@ end subroutine shuffle_3Didata
 !!!-----------------------------------------------------
 subroutine shuffle_4Drdata(data,dim,seed)
 implicit none
-integer :: istart, seed_size
+integer :: istart,seed_size
 integer :: i,j,n_data
 real(real12) :: r
-integer, allocatable, dimension(:) :: iseed
 integer, dimension(4) :: idx_s,idx_e,jdx_s,jdx_e
 integer, dimension(4,2) :: t_size
+integer, allocatable, dimension(:) :: iseed
 real(real12), allocatable, dimension(:,:,:,:) :: tlist
 
 integer, intent(in) :: dim
 real(real12), dimension(:,:,:,:), intent(inout) :: data
 
 integer, optional, intent(in) :: seed
+
 
 !! set or get random seed
 call random_seed(size=seed_size)
@@ -334,6 +345,7 @@ else
    call random_seed(get=iseed)
 end if
 
+!! get the size of the data
 n_data = size(data,dim=dim)
 do i=1,4
   t_size(i,1) = 1
[…]
