+++ test/test_shuffle.f90
@@ -2,20 +2,24 @@ program test_shuffle
   use misc_ml, only: shuffle
   implicit none
 
-  integer :: i, j, k, l, seed_size
+  integer :: i, j, k, l, o, seed_size
   integer, allocatable, dimension(:) :: iseed
   real :: r
   
   integer, parameter :: n = 3
-  integer, parameter :: m = 5
-  integer, parameter :: p = 5
-  integer, parameter :: q = 5
+  integer, parameter :: m = 3
+  integer, parameter :: p = 3
+  integer, parameter :: q = 3
+  integer, parameter :: s = 3
 
 
-  integer, dimension(n) :: array, original_array, array_tmp
+  integer :: original_array(n)
+  integer :: array(n)
+  integer :: array_tmp(n)
 
-  real, dimension(n, m) :: array_2d_original, array_2d_shuffled, &
-       array_2d_shuffled_tmp
+  real :: array_2d_original(n, m)
+  real :: array_2d_shuffled(n, m)
+  real :: array_2d_shuffled_tmp(n, m)
 
   real :: array_3d_original(n, m, p)
   real :: array_3d_shuffled(n, m, p)
@@ -24,6 +28,10 @@ program test_shuffle
   real :: array_4d_original(n, m, p, q)
   real :: array_4d_shuffled(n, m, p, q)
   real :: array_4d_shuffled_tmp(n, m, p, q)
+
+  real :: array_5d_original(n, m, p, q, s)
+  real :: array_5d_shuffled(n, m, p, q, s)
+  real :: array_5d_shuffled_tmp(n, m, p, q, s)
 
   logical :: success = .true.
   
@@ -95,8 +103,6 @@ program test_shuffle
             array_2d_original(i, j)).gt.1.E-6) ) then
           write(*,*) 'Original element', array_2d_original(i, j), &
                'is missing in the shuffled 2D array'
-          write(*,*) array_2d_shuffled(i,:)
-          write(*,*) array_2d_original(i,j)
           success = .false.
        end if
     end do
@@ -146,8 +152,6 @@ program test_shuffle
                array_3d_original(i, j, k)).gt.1.E-6) ) then
              write(*,*) 'Original element', array_3d_original(i, j, k), &
                   'is missing in the shuffled 3D array'
-             write(*,*) array_3d_shuffled(i, j, :)
-             write(*,*) array_3d_original(i, j, k)
              success = .false.
           end if
        end do
@@ -171,7 +175,7 @@ program test_shuffle
   !!! 4D array shuffle tests
   !!!-----------------------------------------------------------------------------
   
-    !! test 3D array shuffling
+    !! test 4D array shuffling
     do i = 1, n
        do j = 1, m
           do k = 1, p
@@ -182,11 +186,11 @@ program test_shuffle
        end do
     end do
   
-    !! Shuffle the 3D array along the third dimension
+    !! Shuffle the 4D array along the third dimension
     array_4d_shuffled = array_4d_original
     call shuffle(array_4d_shuffled, dim = 4, seed = 1)
   
-    !! Check if the 3D array is shuffled along the third dimension
+    !! Check if the 4D array is shuffled along the third dimension
     if (all(array_4d_shuffled .eq. array_4d_original)) then
        write(*,*) '4D Array is not shuffled along the third dimension'
        success = .false.
@@ -201,9 +205,7 @@ program test_shuffle
                   array_4d_original(i, j, k, l)).gt.1.E-6) ) then
                   write(*,*) 'Original element', &
                        array_4d_original(i, j, k, l), &
-                       'is missing in the shuffled 3D array'
-                  write(*,*) array_4d_shuffled(i, j, k, :)
-                  write(*,*) array_4d_original(i, j, k, l)
[…]
