### 2024-03-02 693050e
Remove superfluous variables
