+++ test/test_shuffle.f90
@@ -2,9 +2,7 @@ program test_shuffle
   use misc_ml, only: shuffle
   implicit none
 
-  integer :: i, j, k, l, o, seed_size
-  integer, allocatable, dimension(:) :: iseed
-  real :: r
+  integer :: i, j, k, l, o
   
   integer, parameter :: n = 3
   integer, parameter :: m = 3
@@ -171,9 +169,9 @@ program test_shuffle
   end if
 
 
-  !!!-----------------------------------------------------------------------------
-  !!! 4D array shuffle tests
-  !!!-----------------------------------------------------------------------------
+!!!-----------------------------------------------------------------------------
+!!! 4D array shuffle tests
+!!!-----------------------------------------------------------------------------
   
     !! test 4D array shuffling
     do i = 1, n
@@ -226,10 +224,9 @@ program test_shuffle
     end if
 
 
-
-  !!!-----------------------------------------------------------------------------
-  !!! 5D array shuffle tests
-  !!!-----------------------------------------------------------------------------
+!!!-----------------------------------------------------------------------------
+!!! 5D array shuffle tests
+!!!-----------------------------------------------------------------------------
   
     !! test 5D array shuffling
     do i = 1, n
