### 2024-03-02 7e5f479
Set seed for 1D shuffle
