### 2024-03-02 c1858e5
Add symmetric padding test
