### 2024-03-02 fe6560c
Fix left_size and right_size present in split
