### 2024-03-02 55e25c8
Add learning rate decay unit test
