+++ test/test_shuffle.f90
@@ -3,6 +3,7 @@ program test_shuffle
   implicit none
 
   integer :: i, j, k, l, o
+  integer :: idim
   
   integer, parameter :: n = 10
   integer, parameter :: m = 3
@@ -10,18 +11,26 @@ program test_shuffle
   integer, parameter :: q = 3
   integer, parameter :: s = 3
 
+  integer :: ilabel_original(n)
+  integer :: ilabel_shuffled(n)
+  real :: rlabel_original(n)
+  real :: rlabel_shuffled(n)
 
-  integer :: original_array(n)
-  integer :: array(n)
-  integer :: array_tmp(n)
+  integer :: array_1d_original(n)
+  integer :: array_1d_shuffled(n)
+  integer :: array_1d_shuffled_tmp(n)
 
   real :: array_2d_original(n, m)
   real :: array_2d_shuffled(n, m)
   real :: array_2d_shuffled_tmp(n, m)
 
-  real :: array_3d_original(n, m, p)
-  real :: array_3d_shuffled(n, m, p)
-  real :: array_3d_shuffled_tmp(n, m, p)
+  integer :: iarray_3d_original(n, m, p)
+  integer :: iarray_3d_shuffled(n, m, p)
+  integer :: iarray_3d_shuffled_tmp(n, m, p)
+
+  real :: rarray_3d_original(n, m, p)
+  real :: rarray_3d_shuffled(n, m, p)
+  real :: rarray_3d_shuffled_tmp(n, m, p)
 
   real :: array_4d_original(n, m, p, q)
   real :: array_4d_shuffled(n, m, p, q)
@@ -34,40 +43,43 @@ program test_shuffle
   logical :: success = .true.
   
 
+   ilabel_original = (/ (i, i = 1, n) /)
+   rlabel_original = (/ (i, i = 1, n) /)
+
 !!!-----------------------------------------------------------------------------
 !!! 1D array shuffle tests
 !!!-----------------------------------------------------------------------------
 
   !! initialize the array
-  array = (/ (i, i = 1, n) /)
-  original_array = array ! Make a copy of the original array
+  array_1d_original = (/ (i, i = 1, n) /)
+  array_1d_shuffled = array_1d_original ! Make a copy of the original array
 
   !! shuffle the array
-  call shuffle(array, seed = 1)
+  call shuffle(array_1d_shuffled, seed = 1)
   
   !! check if the array is shuffled
-  if (all(array .eq. original_array)) then
+  if (all(array_1d_shuffled .eq. array_1d_original)) then
      write(*,*) '1D Array is not shuffled'
      success = .false.
   end if
   
   !! check if all original elements are still in the shuffled array
   do i = 1, n
-     if (.not. any(array .eq. original_array(i))) then
-        write(*,*) 'Original element', original_array(i), &
-             'is missing in the shuffled array'
+     if (all(array_1d_shuffled .ne. array_1d_original(i))) then
+        write(*,*) 'Original element', array_1d_original(i), &
+             'is missing in the shuffled 1D array'
         success = .false.
      end if
   end do
 
   !! check that seed works
-  array_tmp = original_array
-  call shuffle(array_tmp, seed = 1)
-  array = original_array
-  call shuffle(array, seed = 1)
+  array_1d_shuffled_tmp = array_1d_original
+  call shuffle(array_1d_shuffled_tmp, seed = 1)
+  array_1d_shuffled = array_1d_original
+  call shuffle(array_1d_shuffled, seed = 1)
 
-  !! check if array and array_tmp are the same
-  if (all(array_tmp .ne. array)) then
+  !! check if array and array_1d_shuffled_tmp are the same
[…]
