### 2024-03-02 4f5b269
Add regulariser unit test
