### 2024-03-02 fc0c4c1
Print to zero
