+++ test/test_conv3d_network.f90
@@ -49,7 +49,7 @@ program test_conv3d_network
   if(success)then
      write(*,*) 'test_conv3d_network passed all tests'
   else
-     write(*,*) 'test_conv3d_network failed one or more tests'
+     write(0,*) 'test_conv3d_network failed one or more tests'
      stop 1
   end if
 
