+++ test/test_conv2d_network.f90
@@ -49,7 +49,7 @@ program test_conv2d_network
   if(success)then
      write(*,*) 'test_conv2d_network passed all tests'
   else
-     write(*,*) 'test_conv2d_network failed one or more tests'
+     write(0,*) 'test_conv2d_network failed one or more tests'
      stop 1
   end if
 
