+++ test/test_flatten4d_layer.f90
@@ -82,7 +82,7 @@ program test_flatten4d_layer
   if(success)then
      write(*,*) 'test_flatten4d_layer passed all tests'
   else
-     write(*,*) 'test_flatten4d_layer failed one or more tests'
+     write(0,*) 'test_flatten4d_layer failed one or more tests'
      stop 1
   end if
 
