+++ test/test_maxpool2d_layer.f90
@@ -146,7 +146,7 @@ program test_maxpool2d_layer
   if(success)then
      write(*,*) 'test_maxpool2d_layer passed all tests'
   else
-     write(*,*) 'test_maxpool2d_layer failed one or more tests'
+     write(0,*) 'test_maxpool2d_layer failed one or more tests'
      stop 1
   end if
 
