+++ test/test_flatten3d_layer.f90
@@ -80,7 +80,7 @@ program test_flatten3d_layer
   if(success)then
      write(*,*) 'test_flatten3d_layer passed all tests'
   else
-     write(*,*) 'test_flatten3d_layer failed one or more tests'
+     write(0,*) 'test_flatten3d_layer failed one or more tests'
      stop 1
   end if
 
