+++ test/test_batchnorm3d_layer.f90
@@ -146,7 +146,7 @@ program test_batchnorm3d_layer
   if(success)then
      write(*,*) 'test_batchnorm3d_layer passed all tests'
   else
-     write(*,*) 'test_batchnorm3d_layer failed one or more tests'
+     write(0,*) 'test_batchnorm3d_layer failed one or more tests'
      stop 1
   end if
 
