+++ test/test_avgpool2d_layer.f90
@@ -169,7 +169,7 @@ program test_avgpool2d_layer
   if(success)then
      write(*,*) 'test_avgpool2d_layer passed all tests'
   else
-     write(*,*) 'test_avgpool2d_layer failed one or more tests'
+     write(0,*) 'test_avgpool2d_layer failed one or more tests'
      stop 1
   end if
 
