+++ test/test_full_layer.f90
@@ -58,7 +58,7 @@ program test_full_layer
   if(success)then
      write(*,*) 'test_full_layer passed all tests'
   else
-     write(*,*) 'test_full_layer failed one or more tests'
+     write(0,*) 'test_full_layer failed one or more tests'
      stop 1
   end if
 
