### 2024-03-02 55ac766
Add misc unit test
