### 2024-03-03 c1e2f94
Add normalisation unit test
