### 2024-03-03 6e69d91
Change optional magnitude
