### 2024-03-03 c00009a
Add more activation tests
