### 2024-03-03 a151165
Add more initialiser unit tests
