### 2024-03-04 17aa304
Move coverage flags to exclude unit tests
