### 2024-03-04 2a4e83d
Add accuracy unit test
