+++ src/lib/mod_tools_infile.f90
@@ -5,16 +5,9 @@
 !!!#############################################################################
 !!! module contains customn input file reading functions and subroutines.
 !!! module includes the following functionsand subroutines:
-!!! val              (outputs contents of string occuring after "=")
-!!! getline          (gets the line using grep and goes back to start of line)
-!!! assignI          (assign an integer to variable)
-!!! assignD          (assign a double precision to variable)
-!!! assignIvec       (assign an arbitrary length vector of integers to variable)
-!!! assignDvec       (assign an arbitrary length vector of DP to variable)
-!!! assignS          (assign a string to variable)
-!!! assignL          (assign a logical to variable (T/t/1 and F/f/0 accepted)
-!!! assign_list      (assign output to a user-defined value in a list)
-!!! assign_listvec   (assign output vector to a user-defined value in a list)
+!!! assign_val       (assign a value to a variable)
+!!! assign_vec       (assign a vector to a variable)
+!!! getline          (return line using grep and goes back to start of line)
 !!! rm_comments      (remove comments from a string (anything after ! or #))
 !!! cat              (cat lines until user-defined end string is encountered)
 !!! stop_check       (check for <STOP> file and LSTOP or LABORT tags inside)
@@ -25,6 +18,16 @@ module infile_tools
 
   private
   
+!!!-----------------------------------------------------
+!!! assign a value to variable
+!!!-----------------------------------------------------
+!!! buffer   = (S, io) sacrifical input character string
+!!! variable = (*, in) variable to assign data to
+!!! found    = (I, io) count for finding variable
+!!! keyword  = (S, in, opt) keyword to start from
+!!! num      = (I, in, opt) number of tags in tag_list
+!!! tag_list = (S, in, opt) list of tags to search for
+!!!-----------------------------------------------------
   interface assign_val
      procedure assignI,assignR,assignD,assignS,assignL
   end interface assign_val
@@ -38,7 +41,7 @@ module infile_tools
   public :: stop_check
 
   
-!!!updated 2023/06/22
+!!!updated 2024/03/04
 
 
 contains
@@ -46,8 +49,9 @@ contains
 !!! val outputs the section of buffer that occurs after an "="
 !!!#############################################################################
   function val(buffer)
-    character(*) :: buffer
+    character(*), intent(in) :: buffer
     character(100) :: val
+
     val=trim(adjustl(buffer((scan(buffer,"=",back=.false.)+1):)))
     return
   end function val
@@ -57,10 +61,16 @@ contains
 !!!#############################################################################
 !!! gets the line from a grep and assigns it to buffer
 !!!#############################################################################
+!!! unit    = (I, in) unit to read from
+!!! pattern = (S, in) pattern to grep for
+!!! buffer  = (S, io) buffer to assign line to
   subroutine getline(unit,pattern,buffer)
-    integer :: unit,Reason
-    character(*) :: pattern,buffer
+    integer, intent(in) :: unit
+    character(*), intent(in) :: pattern
+    character(*), intent(out) :: buffer
     
+    integer :: Reason
+
     call grep(unit,pattern)
     backspace(unit);read(unit,'(A)',iostat=Reason) buffer
     
@@ -69,14 +79,16 @@ contains
 
 
 !!!#############################################################################
-!!! assigns an integer to variable
+!!! assign an integer to variable
 !!!#############################################################################
   subroutine assignI(buffer,variable,found,keyword)
-    integer :: found
-    character(1024) :: buffer2
-    character(*) :: buffer
-    integer :: variable
+    integer, intent(inout) :: found
+    character(*), intent(inout) :: buffer
+    integer, intent(out) :: variable
[…]
