### 2024-03-04 5cb8649
Add operation checks
