+++ test/test_conv2d_layer.f90
@@ -6,11 +6,15 @@ program test_conv2d_layer
        learnable_layer_type
   implicit none
 
-  class(base_layer_type), allocatable :: conv_layer, input_layer
+  class(base_layer_type), allocatable :: conv_layer, conv_layer1, conv_layer2
+  class(base_layer_type), allocatable :: input_layer
   integer, parameter :: num_filters = 32, kernel_size = 3
   real, allocatable, dimension(:,:,:,:) :: input_data, output
   real, parameter :: tol = 1.E-7
   logical :: success = .true.
+
+  real, allocatable, dimension(:) :: params
+  real, allocatable, dimension(:,:) :: outputs
 
 
   !! set up conv2d layer
@@ -103,7 +107,114 @@ program test_conv2d_layer
     write(*,*) output
   end if
 
-  !! check for any fails
+
+!!!-----------------------------------------------------------------------------
+!!! handle layer parameters gradients, and outputs
+!!!-----------------------------------------------------------------------------
+  select type(conv_layer)
+  class is(learnable_layer_type)
+     !! check layer parameter handling
+     params = conv_layer%get_params()
+     if(size(params) .eq. 0)then
+        success = .false.
+        write(0,*) 'conv2d layer has wrong number of parameters'
+     end if
+     params = 1.E0
+     call conv_layer%set_params(params)
+     params = conv_layer%get_params()
+     if(any(abs(params - 1.E0).gt.tol))then
+        success = .false.
+        write(0,*) 'conv2d layer has wrong parameters'
+     end if
+
+     !! check layer gradient handling
+     params = conv_layer%get_gradients() 
+     if(size(params) .eq. 0)then
+        success = .false.
+        write(0,*) 'conv2d layer has wrong number of gradients'
+     end if
+     params = 1.E0
+     call conv_layer%set_gradients(params)
+     params = conv_layer%get_gradients()
+     if(any(abs(params - 1.E0).gt.tol))then
+        success = .false.
+        write(0,*) 'conv2d layer has wrong gradients'
+     end if
+     call conv_layer%set_gradients(10.E0)
+     params = conv_layer%get_gradients()
+     if(any(abs(params - 10.E0).gt.tol))then
+        success = .false.
+        write(0,*) 'conv2d layer has wrong gradients'
+     end if
+
+     !! check layer output handling
+     call conv_layer%get_output(params)
+     if(size(params) .ne. product(conv_layer%output_shape))then
+        success = .false.
+        write(0,*) 'conv2d layer has wrong number of outputs'
+     end if
+     call conv_layer%get_output(outputs)
+     if(any(shape(outputs) .ne. conv_layer%output_shape))then
+        success = .false.
+        write(0,*) 'conv2d layer has wrong number of outputs'
+     end if
+  end select
+
+
+!!!-----------------------------------------------------------------------------
+!!! check layer operations
+!!!-----------------------------------------------------------------------------
+  conv_layer1 = conv2d_layer_type( &
+       num_filters = 1, &
+       kernel_size = 3, &
+       activation_function = 'sigmoid' &
+       )
+  call conv_layer1%init(input_layer%input_shape, batch_size=1)
+  conv_layer2 = conv2d_layer_type( &
+       num_filters = 1, &
+       kernel_size = 3, &
+       activation_function = 'sigmoid' &
+       )
+  call conv_layer2%init(input_layer%input_shape, batch_size=1)
+  select type(conv_layer1)
+  type is(conv2d_layer_type)
+     select type(conv_layer2)
+     type is(conv2d_layer_type)
+        conv_layer = conv_layer1 + conv_layer2
+        select type(conv_layer)
+        type is(conv2d_layer_type)
+           !! check layer addition
[…]
