### 2024-03-04 69f21d3
Add gradient tests
