+++ test/test_conv3d_network.f90
@@ -7,19 +7,33 @@ program test_conv3d_network
 
   type(network_type) :: network
 
-  real, allocatable, dimension(:,:,:,:,:) :: input_data, output
+  integer :: i
+  integer, parameter :: num_channels = 3
+  integer, parameter :: kernel_size = 3
+  integer, parameter :: num_filters1 = 4
+  integer, parameter :: num_filters2 = 8
+  integer, parameter :: width = 7
+
+  real, allocatable, dimension(:,:) :: output_reshaped
+  real, allocatable, dimension(:,:,:,:,:) :: input_data, output, &
+       gradients_weight
+  real, allocatable, dimension(:) :: gradients, gradients_bias
   logical :: success = .true.
 
   write(*,*) "test_conv3d_network"
   !! create network
   call network%add(conv3d_layer_type( &
-       input_shape=[12, 12, 12, 3], &
-       num_filters = 16, &
-       kernel_size = 3 &
+       input_shape=[width, width, width, num_channels], &
+       num_filters = num_filters1, &
+       kernel_size = kernel_size, &
+       kernel_initialiser = "ones", &
+       activation_function = "linear" &
        ))
   call network%add(conv3d_layer_type( &
-       num_filters = 32, &
-       kernel_size = 3 &
+       num_filters = num_filters2, &
+       kernel_size = kernel_size, &
+       kernel_initialiser = "ones", &
+       activation_function = "linear" &
        ))
     
   call network%compile( &
@@ -33,18 +47,74 @@ program test_conv3d_network
   end if
 
   call network%set_batch_size(1)
-  allocate(input_data(12, 12, 12, 3, 1))
+  allocate(input_data(width, width, width, num_channels, 1))
   input_data = 0.0
 
   call network%forward(input_data)
   call network%model(3)%layer%get_output(output)
 
-  if(any(shape(output).ne.[8,8,8,32,1]))then
+  if(any(shape(output).ne.[width-4,width-4,width-4,num_filters2,1]))then
      success = .false.
-     write(*,*) "conv3d network output shape should be [8,8,8,32]"
+     write(*,*) "conv3d network output shape should be [28,28,28,32]"
   end if
 
-  !! check for any fails
+
+!!!-----------------------------------------------------------------------------
+!!! check gradients
+!!!-----------------------------------------------------------------------------
+  output = 0.E0
+  output(:(width-4)/2,:,:,:,:) = 1.E0
+  input_data = 0.E0
+  input_data(:(width)/2,:,:,:,:) = 1.E0
+  call network%forward(input_data)
+  output_reshaped = reshape(output, [(width-4)**3*num_filters2,1])
+  call network%backward(output_reshaped)
+  select type(current => network%model(3)%layer)
+  type is(conv3d_layer_type)
+     gradients = current%get_gradients()
+     gradients_weight = &
+          reshape(&
+          gradients(:kernel_size**3*num_filters1*num_filters2), &
+          [kernel_size,kernel_size,kernel_size,num_filters1,num_filters2])
+     gradients_bias = &
+          gradients(kernel_size**3*num_filters1*num_filters2+1:)
+     if(size(gradients).ne.&
+          (kernel_size**3*num_filters1 + 1) * num_filters2)then
+        success = .false.
+        write(*,*) "conv3d network gradients size should be ", &
+             ( kernel_size**3 * num_filters1 + 1 ) * num_filters2
+     end if
+     if(any(abs(gradients).lt.1.E-6))then
+        success = .false.
+        write(*,*) "conv3d network gradients should not be zero"
+     end if
+     if(any(abs(gradients_weight(1,:,:,:,:) - &
+          gradients_weight(1,1,1,1,1)).gt.1.E-6))then
+        success = .false.
[…]
