+++ test/test_conv2d_network.f90
@@ -7,14 +7,12 @@ program test_conv2d_network
 
   type(network_type) :: network
 
-  integer :: i,j,k
+  integer :: i
   integer, parameter :: num_channels = 3
   integer, parameter :: kernel_size = 3
   integer, parameter :: num_filters1 = 4
   integer, parameter :: num_filters2 = 8
   integer, parameter :: width = 7
-
-  integer, dimension(2) :: lim_w1, lim_w2, lim_g1, lim_g2
 
   real, allocatable, dimension(:,:) :: output_reshaped
   real, allocatable, dimension(:,:,:,:) :: input_data, output, gradients_weight
@@ -68,22 +66,22 @@ program test_conv2d_network
   input_data = 0.E0
   input_data(:(width)/2,:,:,:) = 1.E0
   call network%forward(input_data)
-  output_reshaped = reshape(output, [(width-4)*(width-4)*num_filters2,1])
+  output_reshaped = reshape(output, [(width-4)**2*num_filters2,1])
   call network%backward(output_reshaped)
   select type(current => network%model(3)%layer)
   type is(conv2d_layer_type)
      gradients = current%get_gradients()
      gradients_weight = &
           reshape(&
-          gradients(:kernel_size*kernel_size*num_filters1*num_filters2), &
+          gradients(:kernel_size**2*num_filters1*num_filters2), &
           [kernel_size,kernel_size,num_filters1,num_filters2])
      gradients_bias = &
-          gradients(kernel_size*kernel_size*num_filters1*num_filters2+1:)
+          gradients(kernel_size**2*num_filters1*num_filters2+1:)
      if(size(gradients).ne.&
-          (kernel_size*kernel_size*num_filters1 + 1) * num_filters2)then
+          (kernel_size**2*num_filters1 + 1) * num_filters2)then
         success = .false.
         write(*,*) "conv2d network gradients size should be ", &
-             ( kernel_size * kernel_size * num_filters1 + 1 ) * num_filters2
+             ( kernel_size**2 * num_filters1 + 1 ) * num_filters2
      end if
      if(any(abs(gradients).lt.1.E-6))then
         success = .false.
@@ -114,7 +112,7 @@ program test_conv2d_network
   end select
 
 !!!-----------------------------------------------------------------------------
-!!! check for any fails
+!!! check for any failed tests
 !!!-----------------------------------------------------------------------------
   write(*,*) "----------------------------------------"
   if(success)then
