+++ test/test_maxpool3d_layer.f90
@@ -149,7 +149,9 @@ program test_maxpool3d_layer
   end select
 
 
-  !! check for any fails
+!!!-----------------------------------------------------------------------------
+!!! check for any failed tests
+!!!-----------------------------------------------------------------------------
   write(*,*) "----------------------------------------"
   if(success)then
      write(*,*) 'test_maxpool3d_layer passed all tests'
