+++ test/test_avgpool3d_layer.f90
@@ -177,9 +177,10 @@ program test_avgpool3d_layer
 
   end if
 
-!!!-----------------------------------------------------------------------------
 
-  !! check for any fails
+!!!-----------------------------------------------------------------------------
+!!! check for any failed tests
+!!!-----------------------------------------------------------------------------
   write(*,*) "----------------------------------------"
   if(success)then
      write(*,*) 'test_avgpool3d_layer passed all tests'
