+++ test/test_dropblock2d_layer.f90
@@ -103,7 +103,9 @@ program test_dropblock2d_layer
   end select
 
 
-  !! check for any fails
+!!!-----------------------------------------------------------------------------
+!!! check for any failed tests
+!!!-----------------------------------------------------------------------------
   write(*,*) "----------------------------------------"
   if(success)then
      write(*,*) 'test_dropblock2d_layer passed all tests'
