+++ test/test_shuffle.f90
@@ -472,10 +472,8 @@ program test_shuffle
 
 
 !!!-----------------------------------------------------------------------------
-!!! Final printing array shuffle tests
+!!! check for any failed tests
 !!!-----------------------------------------------------------------------------
-  
-  !! check for any fails
   write(*,*) "----------------------------------------"
   if(success)then
      write(*,*) 'test_shuffle passed all tests'
