+++ test/test_normalisation.f90
@@ -70,11 +70,10 @@ program test_normalisation
      success = .false.
   end if
   
+
 !!!-----------------------------------------------------------------------------
-!!! Final printing array shuffle tests
+!!! check for any failed tests
 !!!-----------------------------------------------------------------------------
-  
-  !! check for any fails
   write(*,*) "----------------------------------------"
   if(success)then
      write(*,*) 'test_shuffle passed all tests'
