### 2024-03-06 9c89a31
Fix negative_log_likelihood
