### 2024-03-06 49cd199
Make copy and reduce structure procedures
