### 2024-03-06 e702bc4
Add default value for already_initialised
