### 2024-03-06 8ffdb03
Add random and optimiser unit tests
