### 2024-03-06 366abcc
Add more learning rate decay tests
