+++ test/test_lr_decay.f90
@@ -1,5 +1,6 @@
 program test_lr_decay
   use misc_ml, only: step_decay, reduce_lr_on_plateau
+  use learning_rate_decay
   implicit none
 
   real :: learning_rate
@@ -15,6 +16,8 @@ program test_lr_decay
   real :: best_metric_value = 0.0E0
   real :: expected_learning_rate
   logical :: success = .true.
+
+  class(base_lr_decay_type), allocatable :: lr_decay
 
 
   !! test step_decay
@@ -46,6 +49,100 @@ program test_lr_decay
      end if
   end do
 
+  !! test base learning rate decay
+  lr_decay = base_lr_decay_type()
+  select type(lr_decay)
+  type is (base_lr_decay_type)
+     if(abs(lr_decay%decay_rate).gt.1.E-6)then
+        write(0,*) "base decay rate failed to initialise"
+        success = .false.
+     end if
+  class default
+     write(0,*) "base_lr_decay_type failed"
+     success = .false.
+  end select
+  if(abs(learning_rate - &
+       lr_decay%get_lr(learning_rate, 1)).gt.1.E-6)then
+     write(0,*) "base learning rate decay failed"
+     success = .false.
+  end if
+
+  !! test exponential learning rate decay
+  lr_decay = exp_lr_decay_type()
+  if(abs(lr_decay%decay_rate-0.9E0).gt.1.E-6)then
+     write(0,*) "step decay rate failed to initialise"
+     success = .false.
+  end if
+  lr_decay = exp_lr_decay_type(decay_rate=0.2E0)
+  select type(lr_decay)
+  type is (exp_lr_decay_type)
+     if(abs(lr_decay%decay_rate-0.2E0).gt.1.E-6)then
+        write(0,*) "exponential decay rate failed to initialise"
+        success = .false.
+     end if
+  class default
+     write(0,*) "exp_lr_decay_type failed"
+     success = .false.
+  end select
+  if(abs(learning_rate * exp( -1 * 0.2E0) - &
+       lr_decay%get_lr(learning_rate, 1)) .gt. 1.E-6)then
+     write(0,*) "expoential learning rate decay failed"
+     success = .false.
+  end if
+
+  !! test step learning rate decay
+  lr_decay = step_lr_decay_type()
+  if(abs(lr_decay%decay_rate-0.1E0).gt.1.E-6)then
+     write(0,*) "step decay rate failed to initialise"
+     success = .false.
+  end if
+  lr_decay = step_lr_decay_type(decay_rate=0.2E0, decay_steps=10)
+  select type(lr_decay)
+  type is (step_lr_decay_type)
+     if(abs(lr_decay%decay_rate-0.2E0).gt.1.E-6)then
+        write(0,*) "step decay rate failed to initialise"
+        success = .false.
+     end if
+     if(lr_decay%decay_steps.ne.10)then
+        write(0,*) "number of decay steps failed to initialise"
+        success = .false.
+     end if
+  class default
+     write(0,*) "step_lr_decay_type failed"
+     success = .false.
+  end select
+  if(abs(learning_rate * 0.2E0 ** (11/10) - &
+       lr_decay%get_lr(learning_rate, 11)) .gt. 1.E-6)then
+     write(0,*) "step learning rate decay failed"
+     success = .false.
+  end if
+
+  !! test inverse learning rate decay
+  lr_decay = inv_lr_decay_type()
+  if(abs(lr_decay%decay_rate-0.001E0).gt.1.E-6)then
+     write(0,*) "inv decay rate failed to initialise"
+     success = .false.
+  end if
+  lr_decay = inv_lr_decay_type(decay_rate=0.2E0, decay_power=2.E0)
+  select type(lr_decay)
[…]
