### 2024-03-06 08cb395
Add further setup tests
