+++ test/test_avgpool2d_layer.f90
@@ -7,7 +7,10 @@ program test_avgpool2d_layer
 
   class(base_layer_type), allocatable :: pool_layer
   integer, parameter :: num_channels = 3, pool = 3, stride = 3, width = 9
-  real, allocatable, dimension(:,:,:,:) :: input_data, output, gradient, di_compare
+  real, allocatable, dimension(:) :: output_1d
+  real, allocatable, dimension(:,:) :: output_2d
+  real, allocatable, dimension(:,:,:,:) :: input_data, output, gradient, &
+       di_compare
   real, parameter :: tol = 1.E-7
   logical :: success = .true.
 
@@ -108,6 +111,23 @@ program test_avgpool2d_layer
      end do
   end do
 
+  !! check 1d and 2d output are the same
+  call pool_layer%get_output(output_1d)
+  call pool_layer%get_output(output_2d)
+  if(any(abs(output_1d - &
+       reshape(output, [output_width*output_width*num_channels])) &
+       .gt. 1.E-6))then
+     success = .false.
+     write(*,*) 'avgpool2d layer output pass failed'
+  end if
+  if(any(abs(&
+       reshape(output_2d, [output_width*output_width*num_channels]) - &
+       reshape(output, [output_width*output_width*num_channels])) &
+       .gt. 1.E-6))then
+     success = .false.
+     write(*,*) 'avgpool2d layer output pass failed'
+  end if
+
 !!!-----------------------------------------------------------------------------
 
   !! run backward pass
@@ -162,6 +182,39 @@ program test_avgpool2d_layer
 
   end if
 
+  !! check expected initialisation of pool and stride
+  pool_layer = avgpool2d_layer_type( &
+       pool_size = [2, 2], &
+       stride = [2, 2] &
+       )
+  select type(pool_layer)
+  type is (avgpool2d_layer_type)
+     if(any(pool_layer%pool .ne. [2, 2]))then
+        success = .false.
+        write(0,*) 'avgpool2d layer has wrong pool size'
+     end if
+     if(any(pool_layer%strd .ne. [2, 2]))then
+        success = .false.
+        write(0,*) 'avgpool2d layer has wrong stride size'
+     end if
+  end select
+
+  !! check expected initialisation of pool and stride
+  pool_layer = avgpool2d_layer_type( &
+       pool_size = [4], &
+       stride = [4] &
+       )
+  select type(pool_layer)
+  type is (avgpool2d_layer_type)
+     if(any(pool_layer%pool .ne. 4))then
+        success = .false.
+        write(0,*) 'avgpool2d layer has wrong pool size'
+     end if
+     if(any(pool_layer%strd .ne. 4))then
+        success = .false.
+        write(0,*) 'avgpool2d layer has wrong stride size'
+     end if
+  end select
 
 !!!-----------------------------------------------------------------------------
 !!! check for any failed tests
