+++ test/test_maxpool2d_layer.f90
@@ -140,6 +140,39 @@ program test_maxpool2d_layer
      end do
   end select
 
+  !! check expected initialisation of pool and stride
+  pool_layer = maxpool2d_layer_type( &
+       pool_size = [2, 2], &
+       stride = [2, 2] &
+       )
+  select type(pool_layer)
+  type is (maxpool2d_layer_type)
+     if(any(pool_layer%pool .ne. [2, 2]))then
+        success = .false.
+        write(0,*) 'maxpool2d layer has wrong pool size'
+     end if
+     if(any(pool_layer%strd .ne. [2, 2]))then
+        success = .false.
+        write(0,*) 'maxpool2d layer has wrong stride size'
+     end if
+  end select
+
+  !! check expected initialisation of pool and stride
+  pool_layer = maxpool2d_layer_type( &
+       pool_size = [4], &
+       stride = [4] &
+       )
+  select type(pool_layer)
+  type is (maxpool2d_layer_type)
+     if(any(pool_layer%pool .ne. 4))then
+        success = .false.
+        write(0,*) 'maxpool2d layer has wrong pool size'
+     end if
+     if(any(pool_layer%strd .ne. 4))then
+        success = .false.
+        write(0,*) 'maxpool2d layer has wrong stride size'
+     end if
+  end select
 
 !!!-----------------------------------------------------------------------------
 !!! check for any failed tests
