### 2024-03-06 65639b4
Add operation tests
