### 2024-03-06 0a56e56
Fix total_loss_mae reference
