+++ src/lib/mod_loss.f90
@@ -131,7 +131,7 @@ contains
     real(real12), dimension(:,:), intent(in) :: predicted, expected
     real(real12), dimension(size(predicted,2)) :: output
 
-    output = sum(compute_loss_mse(predicted,expected),dim=1)
+    output = sum(compute_loss_mae(predicted,expected),dim=1)
     
   end function total_loss_mae
 !!!#############################################################################
