### 2024-03-06 b8e71ca
Add operation tests
