+++ test/test_batchnorm2d_layer.f90
@@ -5,7 +5,7 @@ program test_batchnorm2d_layer
      learnable_layer_type
   implicit none
 
-  class(base_layer_type), allocatable :: bn_layer
+  class(base_layer_type), allocatable :: bn_layer, bn_layer1, bn_layer2
   integer, parameter :: num_channels = 3, width = 8, batch_size = 1
   real, parameter :: gamma  = 0.5, beta = 0.3
   real, allocatable, dimension(:,:,:,:) :: input_data, output, gradient
@@ -143,6 +143,51 @@ program test_batchnorm2d_layer
 
 
 !!!-----------------------------------------------------------------------------
+!!! check layer operations
+!!!-----------------------------------------------------------------------------
+  bn_layer1 = batchnorm2d_layer_type(input_shape=[2,2,1], batch_size=1)
+  bn_layer2 = batchnorm2d_layer_type(input_shape=[2,2,1], batch_size=1)
+  select type(bn_layer1)
+  type is(batchnorm2d_layer_type)
+     bn_layer1%dg = 1.E0
+     bn_layer1%db = 1.E0
+     select type(bn_layer2)
+     type is(batchnorm2d_layer_type)
+        bn_layer2%dg = 2.E0
+        bn_layer2%db = 2.E0
+        bn_layer = bn_layer1 + bn_layer2
+        select type(bn_layer)
+        type is(batchnorm2d_layer_type)
+           !! check layer addition
+           call compare_batchnorm2d_layers(&
+                bn_layer, bn_layer1, success, bn_layer2)
+
+           !! check layer reduction
+           bn_layer = bn_layer1
+           call bn_layer%reduce(bn_layer2)
+           call compare_batchnorm2d_layers(&
+                bn_layer, bn_layer1, success, bn_layer2)
+
+           !! check layer merge
+           bn_layer = bn_layer1
+           call bn_layer%merge(bn_layer2)
+           call compare_batchnorm2d_layers(&
+                bn_layer, bn_layer1, success, bn_layer2)
+        class default
+            success = .false.
+            write(0,*) 'batchnorm2d layer has wrong type'
+        end select
+     class default
+        success = .false.
+        write(0,*) 'batchnorm2d layer has wrong type'
+     end select
+  class default
+     success = .false.
+     write(0,*) 'batchnorm2d layer has wrong type'
+  end select
+
+
+!!!-----------------------------------------------------------------------------
 !!! check for any failed tests
 !!!-----------------------------------------------------------------------------
   write(*,*) "----------------------------------------"
@@ -152,5 +197,25 @@ program test_batchnorm2d_layer
      write(0,*) 'test_batchnorm2d_layer failed one or more tests'
      stop 1
   end if
+
+contains
+
+  subroutine compare_batchnorm2d_layers(layer1, layer2, success, layer3)
+     type(batchnorm2d_layer_type), intent(in) :: layer1, layer2
+     logical, intent(inout) :: success
+     type(batchnorm2d_layer_type), optional, intent(in) :: layer3
+
+     if(present(layer3))then
+        if(any(abs(layer1%dg-layer2%dg-layer3%dg).gt.tol))then
+           success = .false.
+           write(0,*) 'batchnorm2d layer has wrong gradients'
+        end if
+        if(any(abs(layer1%db-layer2%db-layer3%db).gt.tol))then
+           success = .false.
+           write(0,*) 'batchnorm2d layer has wrong gradients'
+        end if
+     end if
+
+  end subroutine compare_batchnorm2d_layers
 
 end program test_batchnorm2d_layer
