### 2024-03-06 6d4820d
Add output 1d and 2d comparison
