### 2024-03-06 1d06e1d
Add output 1d and 2d comparison
