+++ test/test_maxpool2d_layer.f90
@@ -8,6 +8,8 @@ program test_maxpool2d_layer
   class(base_layer_type), allocatable :: pool_layer
   integer, parameter :: num_channels = 3, pool = 3, stride = 2, width = 18
   real, allocatable, dimension(:,:,:,:) :: input_data, output, gradient
+  real, allocatable, dimension(:) :: output_1d
+  real, allocatable, dimension(:,:) :: output_2d
   real, parameter :: tol = 1.E-7
   logical :: success = .true.
 
@@ -109,6 +111,14 @@ program test_maxpool2d_layer
        end if
      end do
   end do
+
+  !! check 1d and 2d output are consistent
+  call pool_layer%get_output(output_1d)
+  call pool_layer%get_output(output_2d)
+  if(any(abs(output_1d - reshape(output_2d, [size(output_2d)])) .gt. 1.E-6))then
+     success = .false.
+     write(0,*) 'output_1d and output_2d are not consistent'
+  end if
 
 !!!-----------------------------------------------------------------------------
 
