+++ test/test_dropblock2d_layer.f90
@@ -7,6 +7,8 @@ program test_dropblock2d_layer
   class(base_layer_type), allocatable :: db_layer
   integer, parameter :: num_channels = 3, width = 6
   real, allocatable, dimension(:,:,:,:) :: input_data, output, gradient
+  real, allocatable, dimension(:) :: output_1d
+  real, allocatable, dimension(:,:) :: output_2d
   real, parameter :: tol = 1.E-7
   logical :: success = .true.
 
@@ -102,6 +104,13 @@ program test_dropblock2d_layer
     end if
   end select
 
+  !! check 1d and 2d output are consistent
+  call db_layer%get_output(output_1d)
+  call db_layer%get_output(output_2d)
+  if(any(abs(output_1d - reshape(output_2d, [size(output_2d)])) .gt. 1.E-6))then
+     success = .false.
+     write(0,*) 'output_1d and output_2d are not consistent'
+  end if
 
 !!!-----------------------------------------------------------------------------
 !!! check for any failed tests
