### 2024-03-06 aab109b
Add default initialiser name test
