### 2024-03-06 116369c
Add output 1d and 2d comparison
