### 2024-03-06 b554733
Remove unnecessary gfortran calls
