+++ .github/workflows/coverage.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: actions-setup-cmake
         uses: jwlawson/actions-setup-cmake@v2.0.1
         with:
-          cmake-version: '3.3.x'
+          cmake-version: '3.16.x'
             
       - uses: fortran-lang/setup-fortran@v1
         id: setup-fortran
@@ -42,10 +42,13 @@ jobs:
           version: ${{ matrix.version }}
       
       - run: |
+          sudo apt-get update
+          sudo apt-get install -y cmake make
+          sudo apt-get install -y cmake make libidn11
           cmake --version
           mkdir -p build
           cd build
-          cmake3 -DCMAKE_BUILD_TYPE=Debug ..
+          cmake -DCMAKE_BUILD_TYPE=Debug ..
           make
           make coverage
 
