+++ .github/workflows/coverage.yaml
@@ -44,7 +44,6 @@ jobs:
       - run: |
           sudo apt-get update
           sudo apt-get install -y cmake make
-          sudo apt-get install -y cmake make libidn11
           cmake --version
           mkdir -p build
           cd build
