+++ .github/workflows/coverage.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: actions-setup-cmake
         uses: jwlawson/actions-setup-cmake@v2.0.1
         with:
-          cmake-version: '3.16.x'
+          cmake-version: '3.17.5'
             
       - uses: fortran-lang/setup-fortran@v1
         id: setup-fortran
