### 2024-03-06 f36260e
Add gcovr installation
