### 2024-03-06 0654a33
Initialise param
