### 2024-03-06 99cc836
Create coverage_badge.py
