### 2024-03-06 9ece3c9
Print only coverage integer
