+++ .github/workflows/coverage.yaml
@@ -57,6 +57,22 @@ jobs:
           ./test/test_regulariser
           make coverage
 
+      - name: Get coverage percentage
+        run: |
+          pip install bs4
+          echo "COVERAGE_PERCENTAGE="$(python ./tools/coverage_badge.py) >> $GITHUB_ENV
+  
+      - name: Create coverage badge
+        uses: schneegans/dynamic-badges-action@v1.7.0
+        with:
+          auth: ${{ secrets.GIST_SECRET }}
+          gistID: <gist-ID>
+          filename: test.json # Use test.svg if you want to use the SVG mode.
+          label: Coverage
+          message: ${{ env.COVERAGE_PERCENTAGE }}%
+          valColorRange: ${{ env.COVERAGE_PERCENTAGE }}
+          maxColorRange: 100
+          minColorRange: 0
 
       - name: upload artifact
         uses: actions/upload-pages-artifact@v1
