+++ .github/workflows/coverage.yaml
@@ -66,7 +66,7 @@ jobs:
         uses: schneegans/dynamic-badges-action@v1.7.0
         with:
           auth: ${{ secrets.GIST_SECRET }}
-          gistID: <gist-ID>
+          gistID: fd7c07c046ecc92d92eaf7bdcb78c4b5
           filename: test.json # Use test.svg if you want to use the SVG mode.
           label: Coverage
           message: ${{ env.COVERAGE_PERCENTAGE }}%
