### 2024-03-07 388b80e
Add parameter and gradient tests
