### 2024-03-07 7deb2f1
Add network metrics setup test
