+++ .github/workflows/coverage.yaml
@@ -3,6 +3,17 @@ name: run-code-coverage
 on:
   push:
     branches: ["main"]
+    paths:
+      - ".github/workflows/coverage.yaml"
+      - "CMakeLists.txt"
+      - "fpm.toml"
+      - "**.f90"
+  pull_request:
+    paths:
+      - ".github/workflows/coverage.yaml"
+      - "CMakeLists.txt"
+      - "fpm.toml"
+      - "**.f90"
   workflow_dispatch:
 
 permissions:
