+++ .github/workflows/coverage.yml
@@ -4,15 +4,13 @@ on:
   push:
     branches: ["main"]
     paths:
-      - ".github/workflows/coverage.yaml"
+      - ".github/workflows/coverage.yml"
       - "CMakeLists.txt"
-      - "fpm.toml"
       - "**.f90"
   pull_request:
     paths:
-      - ".github/workflows/coverage.yaml"
+      - ".github/workflows/coverage.yml"
       - "CMakeLists.txt"
-      - "fpm.toml"
       - "**.f90"
   workflow_dispatch:
 
