### 2024-03-07 f803d9f
Add gcc version to both build and test
