+++ .github/workflows/fpm.yml
@@ -37,11 +37,6 @@ jobs:
     - name: checkout repo
       uses: actions/checkout@v2
 
-    - name: Set up GCC
-      uses: actions/setup-gcc@v0
-      with:
-        gcc-version: latest
-      
     - uses: fortran-lang/setup-fpm@v5
       id: setup-fortran
       with:
@@ -51,13 +46,15 @@ jobs:
 
     - name: Compile
       run: |
+        sudo apt-get update
         gfortran --version
-        fpm build --profile debug
+        fpm build --profile debug --compiler gfortran
 
     - name: Test
       run: |
+        sudo apt-get update
         gfortran --version
-        fpm test --profile debug
+        fpm test --profile debug --compiler gfortran
 
   build-and-test-fpm-release:
     name: Build and test in release mode
@@ -73,11 +70,6 @@ jobs:
     steps:
     - name: checkout repo
       uses: actions/checkout@v2
-
-    - name: Set up GCC
-      uses: actions/setup-gcc@v0
-      with:
-        gcc-version: latest
         
     - uses: fortran-lang/setup-fpm@v5
       id: setup-fortran
@@ -88,10 +80,12 @@ jobs:
 
     - name: Compile
       run: |
+        sudo apt-get update
         gfortran --version
-        fpm build --profile release  --compiler 
+        fpm build --profile release --compiler gfortran
 
     - name: Test
       run: |
+        sudo apt-get update
         gfortran --version
-        fpm test --profile release
+        fpm test --profile release --compiler gfortran
