+++ fpm.toml
@@ -1,5 +1,5 @@
 name = "athena"
-version = "1.2.3"
+version = "1.2.4"
 license = "MIT"
 author = "Ned Thaddeus Taylor"
 maintainer = "n.t.taylor@exeter.ac.uk"
