### 2024-03-07 db993e1
Remove reference to doc directory
A doc directory will be added at a later date. For now, the wiki should provide all the documentation on functionality as needed.
