### 2024-03-07 0bc0865
Move coverage compiler flag
