### 2024-03-07 6936d8f
Add batchnorm1d layer
