+++ src/lib/mod_batchnorm1d_layer.f90
@@ -1,0 +1,708 @@
+!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ARTEMIS group (Hepplestone research group)
+!!! Think Hepplestone, think HRG
+!!!#############################################################################
+module batchnorm1d_layer
+  use constants, only: real12
+  use base_layer, only: batch_layer_type, learnable_layer_type
+  use custom_types, only: initialiser_type
+  implicit none
+  
+  
+  type, extends(batch_layer_type) :: batchnorm1d_layer_type
+     real(real12), allocatable, dimension(:,:) :: output
+     real(real12), allocatable, dimension(:,:) :: di ! gradient of input (i.e. delta)
+   contains
+     procedure, pass(this) :: get_output => get_output_batchnorm1d
+     procedure, pass(this) :: init => init_batchnorm1d
+     procedure, pass(this) :: set_batch_size => set_batch_size_batchnorm1d
+     procedure, pass(this) :: print => print_batchnorm1d
+
+     procedure, pass(this) :: forward  => forward_rank
+     procedure, pass(this) :: backward => backward_rank
+     procedure, private, pass(this) :: forward_2d
+     procedure, private, pass(this) :: backward_2d
+
+     procedure, pass(this) :: reduce => layer_reduction
+     procedure, pass(this) :: merge => layer_merge
+     procedure :: add_t_t => layer_add  !t = type, r = real, i = int
+     generic :: operator(+) => add_t_t !, public
+  end type batchnorm1d_layer_type
+
+  
+  interface batchnorm1d_layer_type
+     module function layer_setup( &
+          input_shape, batch_size, &
+          momentum, epsilon, &
+          gamma_init_mean, gamma_init_std, &
+          beta_init_mean, beta_init_std, &
+          kernel_initialiser, bias_initialiser, &
+          moving_mean_initialiser, moving_variance_initialiser &
+          ) result(layer)
+       integer, dimension(:), optional, intent(in) :: input_shape
+       integer, optional, intent(in) :: batch_size
+       real(real12), optional, intent(in) :: momentum, epsilon
+       real(real12), optional, intent(in) :: gamma_init_mean, gamma_init_std
+       real(real12), optional, intent(in) :: beta_init_mean, beta_init_std
+       character(*), optional, intent(in) :: &
+            kernel_initialiser, bias_initialiser, &
+            moving_mean_initialiser, moving_variance_initialiser
+       type(batchnorm1d_layer_type) :: layer
+     end function layer_setup
+  end interface batchnorm1d_layer_type
+
+
+  private
+  public :: batchnorm1d_layer_type
+  public :: read_batchnorm1d_layer
+
+
+contains
+
+!!!#############################################################################
+!!! layer reduction
+!!!#############################################################################
+  subroutine layer_reduction(this, rhs)
+    implicit none
+    class(batchnorm1d_layer_type), intent(inout) :: this
+    class(learnable_layer_type), intent(in) :: rhs
+
+    select type(rhs)
+    class is(batchnorm1d_layer_type)
+       this%dg = this%dg + rhs%dg
+       this%db = this%db + rhs%db
+    end select
+
+  end subroutine  layer_reduction
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! layer addition
+!!!#############################################################################
+  function layer_add(a, b) result(output)
+    implicit none
+    class(batchnorm1d_layer_type), intent(in) :: a, b
+    type(batchnorm1d_layer_type) :: output
+
+    output = a
+    output%dg = output%dg + b%dg
+    output%db = output%db + b%db
+
+  end function layer_add
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! layer merge
[…]
