+++ test/test_batchnorm1d_layer.f90
@@ -1,0 +1,263 @@
+program test_batchnorm1d_layer
+  use athena, only: &
+     batchnorm1d_layer_type, &
+     base_layer_type, &
+     learnable_layer_type
+  implicit none
+
+  class(base_layer_type), allocatable :: bn_layer, bn_layer1, bn_layer2
+  integer, parameter :: num_channels = 1, width = 8, batch_size = 1
+  real, parameter :: gamma  = 0.5, beta = 0.3
+  real, allocatable, dimension(:,:) :: input_data, output, gradient
+  real, allocatable, dimension(:) :: output_1d, params1, params2
+  real, parameter :: tol = 0.5E-3
+  logical :: success = .true.
+
+  integer :: i, j, output_width, num_params
+  integer :: seed_size = 1
+  real :: mean, std
+  integer, allocatable, dimension(:) :: seed
+  real, parameter :: max_value = 3.0
+
+  !! Initialize random number generator with a seed
+  call random_seed(size = seed_size)
+  allocate(seed(seed_size), source=0)
+  call random_seed(put = seed)
+
+  !! set up batchnorm1d layer
+  bn_layer = batchnorm1d_layer_type( &
+     input_shape = [width], &
+     batch_size = batch_size, &
+     momentum = 0.0, &
+     epsilon = 1e-5, &
+     gamma_init_mean = (gamma), &
+     gamma_init_std = 0.0, &
+     beta_init_mean = beta, &
+     beta_init_std = 0.0, &
+     kernel_initialiser = 'gaussian', &
+     bias_initialiser = 'gaussian', &
+     moving_mean_initialiser = 'zeros', &
+      moving_variance_initialiser = 'zeros' &
+     )
+
+  !! check layer name
+  if(.not. bn_layer%name .eq. 'batchnorm1d')then
+    success = .false.
+    write(0,*) 'batchnorm1d layer has wrong name'
+  end if
+
+  !! check layer type
+  select type(bn_layer)
+  type is(batchnorm1d_layer_type)
+    !! check input shape
+    if(any(bn_layer%input_shape .ne. [width]))then
+      success = .false.
+      write(0,*) 'batchnorm1d layer has wrong input_shape'
+    end if
+
+    !! check output shape
+    if(any(bn_layer%output_shape .ne. [width]))then
+      success = .false.
+      write(0,*) 'batchnorm1d layer has wrong output_shape'
+    end if
+
+    !! check batch size
+    if(bn_layer%batch_size .ne. 1)then
+      success = .false.
+      write(0,*) 'batchnorm1d layer has wrong batch size'
+    end if
+  class default
+    success = .false.
+    write(0,*) 'batchnorm1d layer has wrong type'
+  end select
+
+!!!-----------------------------------------------------------------------------
+
+  !! initialise sample input
+  allocate(input_data(width, batch_size), source = 0.0)
+  
+   input_data = max_value
+
+  !! run forward pass
+  call bn_layer%forward(input_data)
+  call bn_layer%get_output(output)
+
+  !! check outputs all get normalised to zero
+  if (any(output-beta.gt. tol)) then
+    success = .false.
+    write(0,*) 'batchnorm1d layer forward pass failed: &
+         &output should all equal beta'
+  end if
+
+!!!-----------------------------------------------------------------------------
+
+  !! initialise sample input
+  call random_number(input_data)
+
+  !! run forward pass
+  call bn_layer%forward(input_data)
+  call bn_layer%get_output(output)
+
+  !! check outputs all get normalised to zero
+  mean = sum(output(:,:))/(width*batch_size)
+  std = sqrt(sum((output(:,:) - mean)**2)/(width*batch_size))
[…]
