+++ src/lib/mod_autodiff.f90
@@ -1,0 +1,123 @@
+module autodiff
+  implicit none
+
+  type :: ValueAndPartial
+    real :: value
+    real :: partial = 0.0
+  end type ValueAndPartial
+
+  type, abstract :: Expression
+    contains
+      procedure(evaluateAndDerive_gen), deferred, pass(this) :: evaluateAndDerive
+  end type Expression
+
+   
+  abstract interface
+     function evaluateAndDerive_gen(this, variable) result(output)
+       import :: Expression, ValueAndPartial
+       class(Expression), intent(inout) :: this
+       class(Expression), intent(inout) :: variable
+       type(ValueAndPartial) :: output
+     end function evaluateAndDerive_gen
+  end interface
+
+  type, extends(Expression) :: Variable_expr
+    real :: value
+  contains
+    procedure :: evaluateAndDerive => evaluateAndDerive_Variable
+  end type Variable_expr
+
+  type, extends(Expression) :: Plus_expr
+    class(Expression), pointer :: a, b
+  contains
+    procedure :: evaluateAndDerive => evaluateAndDerive_Plus
+  end type Plus_expr
+
+  type, extends(Expression) :: Multiply_expr
+    class(Expression), pointer :: a, b
+  contains
+    procedure :: evaluateAndDerive => evaluateAndDerive_Multiply
+  end type Multiply_expr
+
+  
+contains
+
+  ! Implementation of evaluateAndDerive for Variable
+  type(ValueAndPartial) function evaluateAndDerive_Variable(this, variable)
+    class(Variable_expr), intent(inout) :: this
+    class(Expression), intent(inout) :: variable
+    if(loc(this).eq.loc(variable))then
+       evaluateAndDerive_Variable = ValueAndPartial(this%value, 1.0)
+    else
+       evaluateAndDerive_Variable = ValueAndPartial(this%value, 0.0)
+    end if
+    !write(*,*) "variable", this%value, evaluateAndDerive_Variable
+  end function evaluateAndDerive_Variable
+
+  ! Implementation of evaluateAndDerive for Plus
+  type(ValueAndPartial) function evaluateAndDerive_Plus(this, variable)
+    class(Plus_expr), intent(inout) :: this
+    class(Expression), intent(inout) :: variable
+    type(ValueAndPartial) :: vpA, vpB
+    vpA = this%a%evaluateAndDerive(variable)
+    vpB = this%b%evaluateAndDerive(variable)
+    evaluateAndDerive_Plus = ValueAndPartial(vpA%value + vpB%value, vpA%partial + vpB%partial)
+    !write(*,*) "plus", vpA, vpB, evaluateAndDerive_Plus
+  end function evaluateAndDerive_Plus
+
+  ! Implementation of evaluateAndDerive for Multiply
+  type(ValueAndPartial) function evaluateAndDerive_Multiply(this, variable)
+    class(Multiply_expr), intent(inout) :: this
+    class(Expression), intent(inout) :: variable
+    type(ValueAndPartial) :: vpA, vpB
+    vpA = this%a%evaluateAndDerive(variable)
+    vpB = this%b%evaluateAndDerive(variable)
+    evaluateAndDerive_Multiply = ValueAndPartial(vpA%value * vpB%value, vpB%value * vpA%partial + vpA%value * vpB%partial)
+    !write(*,*) "multiply", vpA, vpB, evaluateAndDerive_Multiply
+  end function evaluateAndDerive_Multiply
+
[…]
