### 2024-03-07 1bfa121
Correctly handle batchnorm gradient setting
