+++ src/lib/mod_batchnorm1d_layer.f90
@@ -34,6 +34,7 @@ module batchnorm1d_layer
   interface batchnorm1d_layer_type
      module function layer_setup( &
           input_shape, batch_size, &
+          num_channels, num_features, &
           momentum, epsilon, &
           gamma_init_mean, gamma_init_std, &
           beta_init_mean, beta_init_std, &
@@ -42,6 +43,7 @@ module batchnorm1d_layer
           ) result(layer)
        integer, dimension(:), optional, intent(in) :: input_shape
        integer, optional, intent(in) :: batch_size
+       integer, optional, intent(in) :: num_channels, num_features
        real(real12), optional, intent(in) :: momentum, epsilon
        real(real12), optional, intent(in) :: gamma_init_mean, gamma_init_std
        real(real12), optional, intent(in) :: beta_init_mean, beta_init_std
@@ -184,6 +186,7 @@ contains
 !!!#############################################################################
   module function layer_setup( &
        input_shape, batch_size, &
+       num_channels, num_features, &
        momentum, epsilon, &
        gamma_init_mean, gamma_init_std, &
        beta_init_mean, beta_init_std, &
@@ -194,6 +197,7 @@ contains
     implicit none
     integer, dimension(:), optional, intent(in) :: input_shape
     integer, optional, intent(in) :: batch_size
+    integer, optional, intent(in) :: num_channels, num_features
     real(real12), optional, intent(in) :: momentum, epsilon
     real(real12), optional, intent(in) :: gamma_init_mean, gamma_init_std
     real(real12), optional, intent(in) :: beta_init_mean, beta_init_std
@@ -210,6 +214,25 @@ contains
     !! initialise batch size
     !!--------------------------------------------------------------------------
     if(present(batch_size)) layer%batch_size = batch_size
+
+
+    !!--------------------------------------------------------------------------
+    !! set up number of channels (alt. number of features)
+    !!--------------------------------------------------------------------------
+      if(present(num_channels))then
+         if(present(num_features))then
+            write(0,*) "ERROR: both num_channels and num_features present"
+            write(0,*) "These represent the same parameter, so are conflicting"
+            stop 1
+         end if
+         layer%num_channels = num_channels
+      else
+         if(present(num_features))then
+            layer%num_channels = num_features
+         else
+            layer%num_channels = -1
+         end if
+      end if
 
 
     !!--------------------------------------------------------------------------
@@ -271,7 +294,13 @@ contains
     !!--------------------------------------------------------------------------
     !! initialise layer shape
     !!--------------------------------------------------------------------------
-    if(present(input_shape)) call layer%init(input_shape=input_shape)
+    if(present(input_shape))then
+       call layer%init(input_shape=input_shape)
+    elseif(present(num_channels))then
+       call layer%init(input_shape=[num_channels])
+    elseif(present(num_features))then
+       call layer%init(input_shape=[num_features])
+    end if
 
   end function layer_setup
 !!!#############################################################################
@@ -310,10 +339,14 @@ contains
 
 
     !!-----------------------------------------------------------------------
-    !! set up number of channels, width, height
+    !! set up number of channels
     !!-----------------------------------------------------------------------
[…]
