### 2024-03-07 dd229c5
Fix batch size check
