### 2024-03-07 0e8e5db
Add number of features or channels handling
