### 2024-03-07 592b0b8
Fix erronous stop when providing num_channels or num_features
