### 2024-03-07 2ad5a2c
Add tests for num_channels and num_features
