### 2024-03-10 4e7d24f
Change batchnorm1d to handle rank 1 and 2 data
