+++ src/lib/mod_batchnorm1d_layer.f90
@@ -11,8 +11,9 @@ module batchnorm1d_layer
   
   
   type, extends(batch_layer_type) :: batchnorm1d_layer_type
-     real(real12), allocatable, dimension(:,:) :: output
-     real(real12), allocatable, dimension(:,:) :: di ! gradient of input (i.e. delta)
+     integer :: num_inputs = 1
+     real(real12), allocatable, dimension(:,:,:) :: output
+     real(real12), allocatable, dimension(:,:,:) :: di ! gradient of input (i.e. delta)
    contains
      procedure, pass(this) :: get_output => get_output_batchnorm1d
      procedure, pass(this) :: init => init_batchnorm1d
@@ -21,8 +22,8 @@ module batchnorm1d_layer
 
      procedure, pass(this) :: forward  => forward_rank
      procedure, pass(this) :: backward => backward_rank
-     procedure, private, pass(this) :: forward_2d
-     procedure, private, pass(this) :: backward_2d
+     procedure, private, pass(this) :: forward_3d
+     procedure, private, pass(this) :: backward_3d
 
      procedure, pass(this) :: reduce => layer_reduction
      procedure, pass(this) :: merge => layer_merge
@@ -34,7 +35,7 @@ module batchnorm1d_layer
   interface batchnorm1d_layer_type
      module function layer_setup( &
           input_shape, batch_size, &
-          num_channels, num_features, &
+          num_channels, num_inputs, &
           momentum, epsilon, &
           gamma_init_mean, gamma_init_std, &
           beta_init_mean, beta_init_std, &
@@ -43,7 +44,7 @@ module batchnorm1d_layer
           ) result(layer)
        integer, dimension(:), optional, intent(in) :: input_shape
        integer, optional, intent(in) :: batch_size
-       integer, optional, intent(in) :: num_channels, num_features
+       integer, optional, intent(in) :: num_channels, num_inputs
        real(real12), optional, intent(in) :: momentum, epsilon
        real(real12), optional, intent(in) :: gamma_init_mean, gamma_init_std
        real(real12), optional, intent(in) :: beta_init_mean, beta_init_std
@@ -131,6 +132,9 @@ contains
     rank(1)
        output = reshape(this%output, [size(this%output)])
     rank(2)
+       output = reshape(this%output, &
+            [product(this%output_shape(:)), this%batch_size])
+    rank(3)
        output = this%output
     end select
   
@@ -151,8 +155,11 @@ contains
     class(batchnorm1d_layer_type), intent(inout) :: this
     real(real12), dimension(..), intent(in) :: input
 
-    select rank(input); rank(2)
-       call forward_2d(this, input)
+    select rank(input)
+    rank(2)
+       call forward_3d(this, input)
+    rank(3)
+       call forward_3d(this, input)
     end select
   end subroutine forward_rank
 !!!#############################################################################
@@ -169,7 +176,12 @@ contains
 
     select rank(input); rank(2)
     select rank(gradient); rank(2)
-      call backward_2d(this, input, gradient)
+      call backward_3d(this, input, gradient)
+    end select
+    end select
+    select rank(input); rank(3)
+    select rank(gradient); rank(3)
+      call backward_3d(this, input, gradient)
     end select
     end select
   end subroutine backward_rank
@@ -186,7 +198,7 @@ contains
 !!!#############################################################################
   module function layer_setup( &
        input_shape, batch_size, &
-       num_channels, num_features, &
+       num_channels, num_inputs, &
        momentum, epsilon, &
        gamma_init_mean, gamma_init_std, &
        beta_init_mean, beta_init_std, &
@@ -197,7 +209,7 @@ contains
     implicit none
[…]
