+++ test/test_batchnorm1d_layer.f90
@@ -19,12 +19,18 @@ program test_batchnorm1d_layer
   integer, allocatable, dimension(:) :: seed
   real, parameter :: max_value = 3.0
 
-  !! Initialize random number generator with a seed
+
+!!!-----------------------------------------------------------------------------
+!!! Initialize random number generator with a seed
+!!!-----------------------------------------------------------------------------
   call random_seed(size = seed_size)
   allocate(seed(seed_size), source=0)
   call random_seed(put = seed)
 
-  !! test num_channels and num_inputs
+
+!!!-----------------------------------------------------------------------------
+!!! test num_channels and num_inputs
+!!!-----------------------------------------------------------------------------
   bn_layer = batchnorm1d_layer_type(input_shape=[width], batch_size=batch_size)
   select type(bn_layer)
   type is(batchnorm1d_layer_type)
@@ -84,7 +90,10 @@ program test_batchnorm1d_layer
   deallocate(bn_layer)
 
 
-  !! set up batchnorm1d layer
+!!!-----------------------------------------------------------------------------
+!!! set up batchnorm1d layer
+!!!-----------------------------------------------------------------------------
+  deallocate(bn_layer)
   bn_layer = batchnorm1d_layer_type( &
      input_shape = [width], &
      batch_size = batch_size, &
@@ -131,11 +140,13 @@ program test_batchnorm1d_layer
     write(0,*) 'batchnorm1d layer has wrong type'
   end select
 
-!!!-----------------------------------------------------------------------------
 
+!!!-----------------------------------------------------------------------------
+!!! test forward pass and check expected output for single-valued input
+!!! use existing layer
+!!!-----------------------------------------------------------------------------
   !! initialise sample input
   allocate(input_data(width, batch_size), source = 0.0)
-  
   input_data = max_value
 
   !! run forward pass
@@ -149,8 +160,11 @@ program test_batchnorm1d_layer
          &output should all equal beta'
   end if
 
-!!!-----------------------------------------------------------------------------
 
+!!!-----------------------------------------------------------------------------
+!!! test forward pass and check expected output for randomised input
+!!! use existing layer
+!!!-----------------------------------------------------------------------------
   !! initialise sample input
   call random_number(input_data)
 
@@ -176,6 +190,11 @@ program test_batchnorm1d_layer
      end if
   end do
 
+
+!!!-----------------------------------------------------------------------------
+!!! test backward pass and check expected output for randomised input
+!!! use existing layer
+!!!-----------------------------------------------------------------------------
   !! run backward pass
   allocate(gradient, source = output)
   call bn_layer%backward(input_data, gradient)
@@ -205,7 +224,10 @@ program test_batchnorm1d_layer
      end do
   end select
 
-  !! handle layer parameters and gradients
+
+!!!-----------------------------------------------------------------------------
+!!! check handling of layer parameters and gradients
+!!!-----------------------------------------------------------------------------
   select type(bn_layer)
   class is(learnable_layer_type)
      !! check parameters
@@ -262,19 +284,19 @@ program test_batchnorm1d_layer
         type is(batchnorm1d_layer_type)
            !! check layer addition
            call compare_batchnorm1d_layers(&
-                bn_layer, bn_layer1, success, bn_layer2)
+                bn_layer, bn_layer1, bn_layer2, success)
 
            !! check layer reduction
            bn_layer = bn_layer1
            call bn_layer%reduce(bn_layer2)
            call compare_batchnorm1d_layers(&
[…]
