+++ src/lib/mod_maxpool2d_layer.f90
@@ -203,17 +203,13 @@ contains
     integer, optional, intent(in) :: batch_size
     integer, optional, intent(in) :: verbose
 
-    integer :: t_verb
+    integer :: verbose_ = 0
 
 
     !!--------------------------------------------------------------------------
     !! initialise optional arguments
     !!--------------------------------------------------------------------------
-    if(present(verbose))then
-       t_verb = verbose
-    else
-       t_verb = 0
-    end if
+    if(present(verbose)) verbose_ = verbose
     if(present(batch_size)) this%batch_size = batch_size
 
 
@@ -251,17 +247,13 @@ contains
    integer, intent(in) :: batch_size
    integer, optional, intent(in) :: verbose
 
-   integer :: t_verb
+   integer :: verbose_ = 0
 
 
    !!--------------------------------------------------------------------------
    !! initialise optional arguments
    !!--------------------------------------------------------------------------
-   if(present(verbose))then
-      t_verb = verbose
-   else
-      t_verb = 0
-   end if
+   if(present(verbose)) verbose_ = verbose
    this%batch_size = batch_size
 
 
