+++ src/lib/mod_network.f90
@@ -512,7 +512,7 @@ contains
     integer, optional, intent(in) :: verbose
     
     integer :: i
-    integer :: t_verb, num_addit_inputs
+    integer :: verbose_, num_addit_inputs
     class(base_layer_type), allocatable :: t_input_layer, t_flatten_layer
 
 
@@ -520,9 +520,9 @@ contains
 !!! initialise optional arguments
 !!!-----------------------------------------------------------------------------
     if(present(verbose))then
-       t_verb = verbose
+       verbose_ = verbose
     else
-       t_verb = 0
+       verbose_ = 0
     end if
 
     
@@ -535,7 +535,7 @@ contains
 !!!-----------------------------------------------------------------------------
 !!! initialise loss method
 !!!-----------------------------------------------------------------------------
-    if(present(loss_method)) call this%set_loss(loss_method, t_verb)
+    if(present(loss_method)) call this%set_loss(loss_method, verbose_)
 
 
 !!!-----------------------------------------------------------------------------
@@ -601,7 +601,7 @@ contains
 !!!-----------------------------------------------------------------------------
 !!! initialise layers
 !!!-----------------------------------------------------------------------------
-    if(t_verb.gt.0)then
+    if(verbose_.gt.0)then
        write(*,*) "layer:",1, this%model(1)%name
        write(*,*) this%model(1)%layer%input_shape
        write(*,*) this%model(1)%layer%output_shape
@@ -610,7 +610,7 @@ contains
        if(.not.allocated(this%model(i)%layer%output_shape)) &
             call this%model(i)%layer%init(this%model(i-1)%layer%output_shape, &
                  this%batch_size)
-       if(t_verb.gt.0)then
+       if(verbose_.gt.0)then
           write(*,*) "layer:",i, this%model(i)%name
           write(*,*) this%model(i)%layer%input_shape
           write(*,*) this%model(i)%layer%output_shape
@@ -1113,10 +1113,10 @@ end function get_gradients
     integer :: num_batches
     integer :: converged
     integer :: history_length
-    integer :: t_verb
-    integer :: t_batch_print
-    real(real12) :: t_plateau
-    logical :: t_shuffle
+    integer :: verbose_ = 0
+    integer :: batch_print_step_ = 20
+    real(real12) :: plateau_threshold_ = 1.E-2_real12
+    logical :: shuffle_batches_ = .true.
 
     !! training loop variables
     integer :: epoch, batch, start_index, end_index
@@ -1134,26 +1134,10 @@ end function get_gradients
 !!!-----------------------------------------------------------------------------
 !!! initialise optional arguments
 !!!-----------------------------------------------------------------------------
-    if(present(plateau_threshold))then
-       t_plateau = plateau_threshold
-    else
-       t_plateau = 1.E-2_real12
-    end if
-    if(present(shuffle_batches))then
-       t_shuffle = shuffle_batches
-    else
-       t_shuffle = .true.
-    end if
-    if(present(batch_print_step))then
-       t_batch_print = batch_print_step
-    else
-       t_batch_print = 20
-    end if
-    if(present(verbose))then
-       t_verb = verbose
-    else
-       t_verb = 0
-    end if
+    if(present(plateau_threshold)) plateau_threshold_ = plateau_threshold
+    if(present(shuffle_batches)) shuffle_batches_ = shuffle_batches
+    if(present(batch_print_step)) batch_print_step_ = batch_print_step
+    if(present(verbose)) verbose_ = verbose
     if(present(batch_size)) this%batch_size = batch_size
 
 
@@ -1231,7 +1215,7 @@ end function get_gradients
        !!-----------------------------------------------------------------------
        !! shuffle batch order at the start of each epoch
[…]
