+++ test/test_batchnorm2d_layer.f90
@@ -20,12 +20,18 @@ program test_batchnorm2d_layer
   integer, allocatable, dimension(:) :: seed
   real, parameter :: max_value = 3.0
 
-  !! Initialize random number generator with a seed
+
+!!!-----------------------------------------------------------------------------
+!!! Initialize random number generator with a seed
+!!!-----------------------------------------------------------------------------
   call random_seed(size = seed_size)
   allocate(seed(seed_size), source=0)
   call random_seed(put = seed)
 
-  !! set up batchnorm2d layer
+
+!!!-----------------------------------------------------------------------------
+!!! set up layer
+!!!-----------------------------------------------------------------------------
   bn_layer = batchnorm2d_layer_type( &
      input_shape = [width, width, num_channels], &
      batch_size = batch_size, &
@@ -72,8 +78,11 @@ program test_batchnorm2d_layer
     write(0,*) 'batchnorm2d layer has wrong type'
   end select
 
-!!!-----------------------------------------------------------------------------
 
+!!!-----------------------------------------------------------------------------
+!!! test forward pass and check expected output for single-valued input
+!!! use existing layer
+!!!-----------------------------------------------------------------------------
   !! initialise sample input
   allocate(input_data(width, width, num_channels, batch_size), source = 0.0)
   
@@ -90,8 +99,11 @@ program test_batchnorm2d_layer
          &output should all equal beta'
   end if
 
-!!!-----------------------------------------------------------------------------
 
+!!!-----------------------------------------------------------------------------
+!!! test forward pass and check expected output for randomised input
+!!! use existing layer
+!!!-----------------------------------------------------------------------------
   !! initialise sample input
   call random_number(input_data)
 
@@ -115,7 +127,11 @@ program test_batchnorm2d_layer
      end if
   end do
 
-  !! run backward pass
+
+!!!-----------------------------------------------------------------------------
+!!! test backward pass and check expected output for randomised input
+!!! use existing layer
+!!!-----------------------------------------------------------------------------
   allocate(gradient, source = output)
   call bn_layer%backward(input_data, gradient)
 
@@ -143,7 +159,10 @@ program test_batchnorm2d_layer
     end do
   end select
 
-  !! handle layer parameters and gradients
+
+!!!-----------------------------------------------------------------------------
+!!! check handling of layer parameters and gradients
+!!!-----------------------------------------------------------------------------
   select type(bn_layer)
   class is(learnable_layer_type)
      !! check parameters
@@ -199,19 +218,19 @@ program test_batchnorm2d_layer
         type is(batchnorm2d_layer_type)
            !! check layer addition
            call compare_batchnorm2d_layers(&
-                bn_layer, bn_layer1, success, bn_layer2)
+                bn_layer, bn_layer1, bn_layer2, success)
 
            !! check layer reduction
            bn_layer = bn_layer1
            call bn_layer%reduce(bn_layer2)
            call compare_batchnorm2d_layers(&
-                bn_layer, bn_layer1, success, bn_layer2)
+                bn_layer, bn_layer1, bn_layer2, success)
 
            !! check layer merge
            bn_layer = bn_layer1
            call bn_layer%merge(bn_layer2)
            call compare_batchnorm2d_layers(&
-                bn_layer, bn_layer1, success, bn_layer2)
+                bn_layer, bn_layer1, bn_layer2, success)
         class default
             success = .false.
             write(0,*) 'batchnorm2d layer has wrong type'
[…]

[…]
