### 2024-03-10 fb236fa
Fix initialisation of argument
