### 2024-03-10 eea898d
Fix error handling
