### 2024-03-10 2a84dd6
Remove redundant learning rate decay procedures
