+++ test/test_misc.f90
@@ -39,37 +39,37 @@ contains
 !!!-----------------------------------------------------------------------------
 !!! compare two strings
 !!!-----------------------------------------------------------------------------
-subroutine check_strings(actual, expected)
-  character(*), intent(in) :: actual
-  character(*), intent(in) :: expected
-
-  if (actual .ne. expected) then
-    write(0,*) "Error: Strings are not equal."
-    write(0,*) "Actual: ", actual
-    write(0,*) "Expected: ", expected
-    success = .false.
-  end if
-end subroutine check_strings
+  subroutine check_strings(actual, expected)
+    character(*), intent(in) :: actual
+    character(*), intent(in) :: expected
+  
+    if (actual .ne. expected) then
+      write(0,*) "Error: Strings are not equal."
+      write(0,*) "Actual: ", actual
+      write(0,*) "Expected: ", expected
+      success = .false.
+    end if
+  end subroutine check_strings
 
 !!!-----------------------------------------------------------------------------
 !!! check counter returns correct number of words in string
 !!!-----------------------------------------------------------------------------
-subroutine check_icount(full_line, tmpchar, expected)
-  character(*), intent(in) :: full_line
-  character(*), intent(in) :: tmpchar
-  integer, intent(in) :: expected
-  integer :: actual
-
-  actual = Icount(full_line, tmpchar)
-
-  if (actual .ne. expected) then
-    write(0,*) "Error: Word counts are not equal."
-    write(0,*) "Line: ", full_line
-    write(0,*) "Char: ", tmpchar
-    write(0,*) "Actual: ", actual
-    write(0,*) "Expected: ", expected
-    success = .false.
-  end if
-end subroutine check_icount
+  subroutine check_icount(full_line, tmpchar, expected)
+    character(*), intent(in) :: full_line
+    character(*), intent(in) :: tmpchar
+    integer, intent(in) :: expected
+    integer :: actual
+  
+    actual = Icount(full_line, tmpchar)
+  
+    if (actual .ne. expected) then
+      write(0,*) "Error: Word counts are not equal."
+      write(0,*) "Line: ", full_line
+      write(0,*) "Char: ", tmpchar
+      write(0,*) "Actual: ", actual
+      write(0,*) "Expected: ", expected
+      success = .false.
+    end if
+  end subroutine check_icount
 
 end program test_misc
