+++ src/lib/mod_accuracy.f90
@@ -3,6 +3,13 @@
 !!! Code part of the ARTEMIS group (Hepplestone research group)
 !!! Think Hepplestone, think HRG
 !!!#############################################################################
+!!! module contains functions to compute the accuracy of a model
+!!! module includes the following procedures:
+!!! categorical_score - computes the accuracy of a categorisation model
+!!! mae_score         - computes the mean absolute error of a continuous model
+!!! mse_score         - computes the mean squared error of a continuous model
+!!! r2_score          - computes the R^2 score of a continuous model
+!!!#############################################################################
 module accuracy
   use constants, only: real12
   implicit none
