+++ src/lib/mod_base_layer.f90
@@ -3,11 +3,55 @@
 !!! Code part of the ARTEMIS group (Hepplestone research group)
 !!! Think Hepplestone, think HRG
 !!!#############################################################################
+!!! definition of the abstract base layer type, from which all other layers ...
+!!! ... are derived
+!!! module includes the following public abstract types:
+!!! base_layer_type      - abstract type for all layers
+!!! input_layer_type     - abstract type for input layers
+!!! flatten_layer_type   - abstract type for flatten (rehsape) layers
+!!! pool_layer_type      - abstract type for spatial pooling layers
+!!! drop_layer_type      - abstract type for dropout layers
+!!! learnable_layer_type - abstract type for layers with learnable parameters
+!!! conv_layer_type      - abstract type for spatial convolutional layers
+!!! batch_layer_type     - abstract type for batch normalisation layers
+!!!##################
+!!! base_layer_type includes the following procedures:
+!!! set_shape            - set the input shape of the layer
+!!! get_num_params       - get the number of parameters in the layer
+!!! print                - print the layer to a file
+!!! get_output           - get the output of the layer
+!!! init                 - initialise the layer
+!!! set_batch_size       - set the batch size of the layer
+!!! forward              - forward pass of layer
+!!! backward             - backward pass of layer
+!!!##################
+!!! input_layer_type includes the following unique procedures:
+!!! set                  - set the input of the layer
+!!!##################
+!!! learnable_layer_type includes the following unique procedures:
+!!! layer_reduction      - reduce the layer to a single value
+!!! layer_merge          - merge the layer with another layer
+!!! get_params           - get the learnable parameters of the layer
+!!! set_params           - set the learnable parameters of the layer
+!!! get_gradients        - get the gradients of the layer
+!!! set_gradients        - set the gradients of the layer
+!!!#############################################################################
 module base_layer
   use constants, only: real12
   use clipper, only: clip_type
   use custom_types, only: activation_type
   implicit none
+
+  private
+
+  public :: base_layer_type
+  public :: input_layer_type
+  public :: flatten_layer_type
+  public :: pool_layer_type
+  public :: drop_layer_type
+  public :: learnable_layer_type
+  public :: conv_layer_type
+  public :: batch_layer_type
 
 !!!------------------------------------------------------------------------
 !!! layer abstract type
@@ -38,17 +82,53 @@ module base_layer
      !! In each layer container, you know what size you are expecting for the input, so just take that based on a select type (or of a previous?)
   end type base_layer_type
 
-  abstract interface
-     subroutine initialise(this, input_shape, batch_size, verbose)
-       import :: base_layer_type
-       class(base_layer_type), intent(inout) :: this
-       integer, dimension(:), intent(in) :: input_shape
-       integer, optional, intent(in) :: batch_size
-       integer, optional, intent(in) :: verbose
-     end subroutine initialise
+  interface
+    !!--------------------------------------------------------------------------
+    !! print layer to file (do nothing for a base layer)
+    !!--------------------------------------------------------------------------
+    !! this = (T, in) base_layer_type
+    !! file = (I, in) file name
+    module subroutine print_base(this, file)
+      class(base_layer_type), intent(in) :: this
+      character(*), intent(in) :: file
+    end subroutine print_base
+  end interface
 
-    subroutine set_batch_size(this, batch_size, verbose)
-      import :: base_layer_type
+  interface
+    !!--------------------------------------------------------------------------
+    !! setup input layer shape
+    !!--------------------------------------------------------------------------
+    !! this        = (T, inout) base_layer_type
+    !! input_shape = (I, in) input shape
[…]
