+++ src/lib/mod_base_layer_sub.f90
@@ -1,0 +1,214 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ARTEMIS group (Hepplestone research group)
+!!! Think Hepplestone, think HRG
+!!!#############################################################################
+!!! definition of the abstract base layer type, from which all other layers ...
+!!! ... are derived
+!!! module includes the following public abstract types:
+!!! base_layer_type      - abstract type for all layers
+!!! input_layer_type     - abstract type for input layers
+!!! flatten_layer_type   - abstract type for flatten (rehsape) layers
+!!! pool_layer_type      - abstract type for spatial pooling layers
+!!! drop_layer_type      - abstract type for dropout layers
+!!! learnable_layer_type - abstract type for layers with learnable parameters
+!!! conv_layer_type      - abstract type for spatial convolutional layers
+!!! batch_layer_type     - abstract type for batch normalisation layers
+!!!##################
+!!! base_layer_type includes the following procedures:
+!!! set_shape            - set the input shape of the layer
+!!! get_num_params       - get the number of parameters in the layer
+!!! print                - print the layer to a file
+!!! get_output           - get the output of the layer
+!!! init                 - initialise the layer
+!!! set_batch_size       - set the batch size of the layer
+!!! forward              - forward pass of layer
+!!! backward             - backward pass of layer
+!!!##################
+!!! input_layer_type includes the following unique procedures:
+!!! set                  - set the input of the layer
+!!!##################
+!!! learnable_layer_type includes the following unique procedures:
+!!! layer_reduction      - reduce the layer to a single value
+!!! layer_merge          - merge the layer with another layer
+!!! get_params           - get the learnable parameters of the layer
+!!! set_params           - set the learnable parameters of the layer
+!!! get_gradients        - get the gradients of the layer
+!!! set_gradients        - set the gradients of the layer
+!!!#############################################################################
+submodule(base_layer) base_layer_submodule
+  implicit none
+
+contains
+
+!!!#############################################################################
+!!! print layer to file (do nothing for a base layer)
+!!!#############################################################################
+!!! this = (T, in) base_layer_type
+!!! file = (I, in) file name
+  module subroutine print_base(this, file)
+    implicit none
+    class(base_layer_type), intent(in) :: this
+    character(*), intent(in) :: file
+
+    !! NO NEED TO WRITE ANYTHING FOR A DEFAULT LAYER
+    return
+  end subroutine print_base
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! setup input layer shape
+!!!#############################################################################
+!!! this        = (T, inout) base_layer_type
+!!! input_shape = (I, in) input shape
+  module subroutine set_shape_base(this, input_shape)
+    implicit none
+    class(base_layer_type), intent(inout) :: this
+    integer, dimension(:), intent(in) :: input_shape
+    character(len=100) :: err_msg
+
+    !!--------------------------------------------------------------------------
+    !! initialise input shape
+    !!--------------------------------------------------------------------------
+    if(size(input_shape,dim=1).eq.this%input_rank)then
+       this%input_shape = input_shape
+    else
+       write(err_msg,'("ERROR: invalid size of input_shape in ",A,&
+            &" expected (",I0,"), got (",I0")")')  &
+            trim(this%name), this%input_rank, size(input_shape,dim=1)
+       stop trim(err_msg)
+    end if
+ 
+  end subroutine set_shape_base
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! get number of parameters in layer
+!!!#############################################################################
+!!! this       = (T, in) layer_type
+!!! num_params = (I, out) number of parameters
+  pure module function get_num_params_base(this) result(num_params)
[…]
