### 2024-03-10 2654b0e
Fix deallocation statement
