### 2024-03-11 806cf5e
Fix erronous duplication of install prefix
