+++ src/lib/mod_base_layer.f90
@@ -83,99 +83,99 @@ module base_layer
   end type base_layer_type
 
   interface
-    !!--------------------------------------------------------------------------
-    !! print layer to file (do nothing for a base layer)
-    !!--------------------------------------------------------------------------
-    !! this = (T, in) base_layer_type
-    !! file = (I, in) file name
-    module subroutine print_base(this, file)
-      class(base_layer_type), intent(in) :: this
-      character(*), intent(in) :: file
-    end subroutine print_base
+     !!--------------------------------------------------------------------------
+     !! print layer to file (do nothing for a base layer)
+     !!--------------------------------------------------------------------------
+     !! this = (T, in) base_layer_type
+     !! file = (I, in) file name
+     module subroutine print_base(this, file)
+       class(base_layer_type), intent(in) :: this
+       character(*), intent(in) :: file
+     end subroutine print_base
   end interface
 
   interface
-    !!--------------------------------------------------------------------------
-    !! setup input layer shape
-    !!--------------------------------------------------------------------------
-    !! this        = (T, inout) base_layer_type
-    !! input_shape = (I, in) input shape
-    module subroutine set_shape_base(this, input_shape)
-      class(base_layer_type), intent(inout) :: this
-      integer, dimension(:), intent(in) :: input_shape
-    end subroutine set_shape_base
+     !!--------------------------------------------------------------------------
+     !! setup input layer shape
+     !!--------------------------------------------------------------------------
+     !! this        = (T, inout) base_layer_type
+     !! input_shape = (I, in) input shape
+     module subroutine set_shape_base(this, input_shape)
+       class(base_layer_type), intent(inout) :: this
+       integer, dimension(:), intent(in) :: input_shape
+     end subroutine set_shape_base
   end interface
 
 
   abstract interface
-    !!--------------------------------------------------------------------------
-    !! initialise layer
-    !!--------------------------------------------------------------------------
-    !! this        = (T, inout) base_layer_type
-    !! input_shape = (I, in) input shape
-    !! batch_size  = (I, in) batch size
-    !! verbose     = (I, in) verbosity level
-    module subroutine initialise(this, input_shape, batch_size, verbose)
-      class(base_layer_type), intent(inout) :: this
-      integer, dimension(:), intent(in) :: input_shape
-      integer, optional, intent(in) :: batch_size
-      integer, optional, intent(in) :: verbose
-    end subroutine initialise
+     !!--------------------------------------------------------------------------
+     !! initialise layer
+     !!--------------------------------------------------------------------------
+     !! this        = (T, inout) base_layer_type
+     !! input_shape = (I, in) input shape
+     !! batch_size  = (I, in) batch size
+     !! verbose     = (I, in) verbosity level
+     module subroutine initialise(this, input_shape, batch_size, verbose)
+       class(base_layer_type), intent(inout) :: this
+       integer, dimension(:), intent(in) :: input_shape
+       integer, optional, intent(in) :: batch_size
+       integer, optional, intent(in) :: verbose
+     end subroutine initialise
 
-    !!--------------------------------------------------------------------------
-    !! set batch size
-    !!--------------------------------------------------------------------------
-    !! this       = (T, inout) base_layer_type
-    !! batch_size = (I, in) batch size
-    !! verbose    = (I, in) verbosity level
-    module subroutine set_batch_size(this, batch_size, verbose)
-      class(base_layer_type), intent(inout) :: this
-      integer, intent(in) :: batch_size
-      integer, optional, intent(in) :: verbose
-    end subroutine set_batch_size
+     !!--------------------------------------------------------------------------
+     !! set batch size
+     !!--------------------------------------------------------------------------
[…]
