+++ src/lib/mod_network_sub.f90
@@ -1,0 +1,1526 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ARTEMIS group (Hepplestone research group)
+!!! Think Hepplestone, think HRG
+!!!#############################################################################
+submodule(network) network_submodule
+#ifdef _OPENMP
+  use omp_lib
+#endif
+  use misc_ml, only: shuffle
+
+  use accuracy, only: categorical_score, mae_score, mse_score, r2_score
+  use base_layer, only: &
+       input_layer_type, flatten_layer_type, &
+       drop_layer_type, &
+       learnable_layer_type, &
+       batch_layer_type, &
+       conv_layer_type, &
+       pool_layer_type
+#if defined(GFORTRAN)
+  use container_layer, only: container_reduction
+#endif
+
+  !! input layer types
+  use input1d_layer,   only: input1d_layer_type
+  use input3d_layer,   only: input3d_layer_type
+  use input4d_layer,   only: input4d_layer_type
+
+  !! batch normalisation layer types
+  use batchnorm1d_layer, only: batchnorm1d_layer_type, read_batchnorm1d_layer
+  use batchnorm2d_layer, only: batchnorm2d_layer_type, read_batchnorm2d_layer
+  use batchnorm3d_layer, only: batchnorm3d_layer_type, read_batchnorm3d_layer
+
+  !! convolution layer types
+  use conv2d_layer,    only: conv2d_layer_type, read_conv2d_layer
+  use conv3d_layer,    only: conv3d_layer_type, read_conv3d_layer
+
+  !! dropout layer types
+  use dropout_layer, only: dropout_layer_type, read_dropout_layer
+  use dropblock2d_layer, only: dropblock2d_layer_type, read_dropblock2d_layer
+  use dropblock3d_layer, only: dropblock3d_layer_type, read_dropblock3d_layer
+
+  !! pooling layer types
+  use avgpool2d_layer, only: avgpool2d_layer_type, read_avgpool2d_layer
+  use avgpool3d_layer, only: avgpool3d_layer_type, read_avgpool3d_layer
+  use maxpool2d_layer, only: maxpool2d_layer_type, read_maxpool2d_layer
+  use maxpool3d_layer, only: maxpool3d_layer_type, read_maxpool3d_layer
+
+  !! flatten layer types
+  use flatten1d_layer, only: flatten1d_layer_type
+  use flatten2d_layer, only: flatten2d_layer_type
+  use flatten3d_layer, only: flatten3d_layer_type
+  use flatten4d_layer, only: flatten4d_layer_type
+
+  !! fully connected (dense) layer types
+  use full_layer,      only: full_layer_type, read_full_layer
+
+  implicit none
+
+! #ifdef _OPENMP
+!   !$omp declare reduction(network_reduction:network_type:omp_out%network_reduction(omp_in)) &
+!   !$omp& initializer(omp_priv = omp_orig)
+! #endif
+
+contains
+
+!!!#############################################################################
+!!! network addition
+!!!#############################################################################
+  module subroutine network_reduction(this, source)
+    implicit none
+    class(network_type), intent(inout) :: this
+    type(network_type), intent(in) :: source
+
+    integer :: i
+    
+    this%metrics(1)%val = this%metrics(1)%val + source%metrics(1)%val
+    this%metrics(2)%val = this%metrics(2)%val + source%metrics(2)%val
+    do i=1,size(this%model)
+       select type(layer_this => this%model(i)%layer)
+       class is(learnable_layer_type)
+          select type(layer_source => source%model(i)%layer)
+          class is(learnable_layer_type)
+             call layer_this%merge(layer_source)
+          end select
+       end select
+    end do
+
+  end subroutine network_reduction
+!!!#############################################################################
+
+
[…]
