+++ src/lib/mod_network.f90
@@ -3,69 +3,24 @@
 !!! Code part of the ARTEMIS group (Hepplestone research group)
 !!! Think Hepplestone, think HRG
 !!!#############################################################################
+!!! module contains the network class, which is used to define a neural network
+!!! module contains the following derived types:
+!!! - network_type
 module network
-#ifdef _OPENMP
-  use omp_lib
-#endif
   use constants, only: real12
-  use misc_ml, only: shuffle
-
-  use accuracy, only: categorical_score, mae_score, mse_score, r2_score
-
   use metrics, only: metric_dict_type
   use optimiser, only: base_optimiser_type
   use loss, only: &
        comp_loss_func => compute_loss_function, &
        comp_loss_deriv => compute_loss_derivative
-
-  use base_layer,      only: base_layer_type, &
-       input_layer_type, flatten_layer_type, &
-       drop_layer_type, &
-       learnable_layer_type, &
-       batch_layer_type, &
-       conv_layer_type, &
-       pool_layer_type
-#if defined(GFORTRAN)
-  use container_layer, only: container_layer_type, container_reduction
-#else
+  use base_layer, only: base_layer_type
   use container_layer, only: container_layer_type
-#endif
-
-  !! input layer types
-  use input1d_layer,   only: input1d_layer_type
-  use input3d_layer,   only: input3d_layer_type
-  use input4d_layer,   only: input4d_layer_type
-
-  !! batch normalisation layer types
-  use batchnorm1d_layer, only: batchnorm1d_layer_type, read_batchnorm1d_layer
-  use batchnorm2d_layer, only: batchnorm2d_layer_type, read_batchnorm2d_layer
-  use batchnorm3d_layer, only: batchnorm3d_layer_type, read_batchnorm3d_layer
-
-  !! convolution layer types
-  use conv2d_layer,    only: conv2d_layer_type, read_conv2d_layer
-  use conv3d_layer,    only: conv3d_layer_type, read_conv3d_layer
-
-  !! dropout layer types
-  use dropout_layer, only: dropout_layer_type, read_dropout_layer
-  use dropblock2d_layer, only: dropblock2d_layer_type, read_dropblock2d_layer
-  use dropblock3d_layer, only: dropblock3d_layer_type, read_dropblock3d_layer
-
-  !! pooling layer types
-  use avgpool2d_layer, only: avgpool2d_layer_type, read_avgpool2d_layer
-  use avgpool3d_layer, only: avgpool3d_layer_type, read_avgpool3d_layer
-  use maxpool2d_layer, only: maxpool2d_layer_type, read_maxpool2d_layer
-  use maxpool3d_layer, only: maxpool3d_layer_type, read_maxpool3d_layer
-
-  !! flatten layer types
-  use flatten1d_layer, only: flatten1d_layer_type
-  use flatten2d_layer, only: flatten2d_layer_type
-  use flatten3d_layer, only: flatten3d_layer_type
-  use flatten4d_layer, only: flatten4d_layer_type
-
-  !! fully connected (dense) layer types
-  use full_layer,      only: full_layer_type, read_full_layer
-
   implicit none
+
+  private
+
+  public :: network_type
+
 
   type :: network_type
      real(real12) :: accuracy, loss
@@ -106,1487 +61,300 @@ module network
   end type network_type
 
   interface network_type
+     !!-------------------------------------------------------------------------
+     !! setup the network (network initialisation)
+     !!-------------------------------------------------------------------------
+     !! layers      = (T, in) layer container
+     !! optimiser   = (T, in, opt) optimiser
+     !! loss_method = (S, in, opt) loss method
+     !! metrics     = (*, in, opt) metrics, either string or metric_dict_type
+     !! batch_size  = (I, in, opt) batch size
[…]
