+++ CMakeLists.txt
@@ -46,8 +46,6 @@ if (CMAKE_BUILD_TYPE MATCHES "Debug*" OR CMAKE_BUILD_TYPE MATCHES "Dev*")
         EXECUTABLE ctest
         EXCLUDE "${PROJECT_SOURCE_DIR}/test/*")
   endif()
-  # add coverage compiler flags
-  append_coverage_compiler_flags()
 endif()
 
 # enable testing
@@ -219,5 +217,10 @@ target_compile_options(${PROJECT_NAME} PUBLIC "$<$<CONFIG:Parallel_Dev>:${DEVFLA
 
 # include the build test directory
 if(BUILD_TESTS)
-    add_subdirectory(test)
+   add_subdirectory(test)
+endif()
+
+# add coverage compiler flags
+if (CMAKE_BUILD_TYPE MATCHES "Debug*" OR CMAKE_BUILD_TYPE MATCHES "Dev*")
+  append_coverage_compiler_flags()
 endif()
