+++ src/lib/mod_activation_sigmoid.f90
@@ -1,7 +1,8 @@
 !!!#############################################################################
-!!! Code written by Ned Thaddeus Taylor 
-!!! Code part of the ARTEMIS group (Hepplestone research group)
-!!! Think Hepplestone, think HRG
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains implementation of the sigmoid activation function
 !!!#############################################################################
 module activation_sigmoid
   use constants, only: real12
