+++ src/lib/mod_full_layer.f90
@@ -1,7 +1,8 @@
 !!!#############################################################################
 !!! Code written by Ned Thaddeus Taylor
-!!! Code part of the ARTEMIS group (Hepplestone research group)
-!!! Think Hepplestone, think HRG
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains implementation of a fully connected (dense) layer
 !!!#############################################################################
 module full_layer
   use constants, only: real12
