+++ src/athena.f90
@@ -1,7 +1,10 @@
 !!!#############################################################################
 !!! Code written by Ned Thaddeus Taylor
-!!! Code part of the ARTEMIS group (Hepplestone research group)
-!!! Think Hepplestone, think HRG
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains all publicly available types and procedures within the ...
+!!! ... ATHENA library
+!!! No other types or procedures should be needed to be accessed by the user
 !!!#############################################################################
 module athena
   use misc_ml, only: shuffle, split, pad_data
