### 2024-03-13 #6364 @nedtaylor
Hi @HaoZeke. I would definitely recommend authors of neural-fortran. I'm happy to wait until the 25th of March for Milan Curcic.

If other reviewers are needed, from the JOSS reviewers list, it appears that José Luiz Rybarczyk Filho, Rocco Meli, and Pankaj K Mishra look like good fits.

Additionally, am I allowed to update the version and branch?
