### 2024-03-13 03fa994
Remove unnecessary test
