### 2024-03-13 ff51c37
Add gcovr and comment out apt-get
