+++ .github/workflows/coverage.yml
@@ -32,12 +32,11 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest]
         compiler: [gcc]
         version: [13]
     steps:
       - name: checkout repo
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
       
       - name: actions-setup-cmake
         uses: jwlawson/actions-setup-cmake@v2.0.1
