+++ .github/workflows/cmake.yml
@@ -36,7 +36,7 @@ jobs:
         version: [13]
     steps:
       - name: checkout repo
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
       
       - name: actions-setup-cmake
         uses: jwlawson/actions-setup-cmake@v2.0.1
@@ -82,7 +82,7 @@ jobs:
         version: [13]
     steps:
       - name: checkout repo
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
       
       - name: actions-setup-cmake
         uses: jwlawson/actions-setup-cmake@v2.0.1
