### 2024-03-13 #6364 @HaoZeke
hi @jrybarczyk @RMeli @pankajkmishra 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

If not, could you recommend any potential reviewers?
