### 2024-03-13 #6364 @RMeli
Hi @HaoZeke, I just started editing for JOSS and I'd rather edit than review at this time. However, if you struggle to find other reviewers quickly I'd be happy to review this.

Maybe @awvwgk might be interested in reviewing this submission? (Or knows someone in the Fortran community that might be interested?)
