### 2024-03-15 #6492 @nedtaylor
Hi @HaoZeke, a few questions. Firstly, the paper version is now 1.3.0 (updated around 2-3 days ago, so before the paper went into review). Can this be updated here?

Second, I notice that there are still some issues with DOIs for references. I assume I should fix them:
https://dl.acm.org/doi/10.5555/3327546.3327732
https://dl.acm.org/doi/10.5555/3045118.3045167
https://dl.acm.org/doi/10.5555/2627435.2670313

Should I add them to the paper.bib as urls (it appears they don’t work exactly as DOIs from the journal)? Or should I wait for review to be over now?
