### 2024-03-15 #6492 @HaoZeke
> Hi @HaoZeke, a few questions. Firstly, the paper version is now 1.3.0 (updated around 2-3 days ago, so before the paper went into review). Can this be updated here?
>
> Second, I notice that there are still some issues with DOIs for references. I assume I should fix them: https://dl.acm.org/doi/10.5555/3327546.3327732 https://dl.acm.org/doi/10.5555/3045118.3045167 https://dl.acm.org/doi/10.5555/2627435.2670313
>
> Should I add them to the paper.bib as urls (it appears they don’t work exactly as DOIs from the journal)? Or should I wait for review to be over now?

Hi @nedtaylor, great questions. I wouldn't worry about the version for now, at the end of the review we'll set the version to the latest (which should also include changes in response to the reviewers, if any).

Modifying the paper.bib should be alright, but can also wait until the first round of reviews are in (you can mention it in the response / changes made, if any).
