+++ example/simple/src/main.f90
@@ -16,9 +16,8 @@ program simple
   
 
   !! set random seed
-  seed_size = 8
   call random_seed(size=seed_size)
-  seed = [1,1,1,1,1,1,1,1]
+  allocate(seed(seed_size), source = 1)
   call random_seed(put=seed)
 
   write(*,*) "Simple function approximation using a fully-connected neural network"
