+++ example/mnist/src/main.f90
@@ -71,7 +71,7 @@ program mnist_example
 !!!-----------------------------------------------------------------------------
 !!! initialise random seed
 !!!-----------------------------------------------------------------------------
-  call random_setup(seed, num_seed=1, restart=.false.)
+  call random_setup(seed, restart=.false.)
 
 
 !!!-----------------------------------------------------------------------------
