+++ test/test_dropblock2d_layer.f90
@@ -15,7 +15,7 @@ program test_dropblock2d_layer
   integer :: i, j, output_width
   real, parameter :: max_value = 3.0
 
-  integer :: seed_size = 1
+  integer :: seed_size
   integer, allocatable, dimension(:) :: seed
 
 
