### 2024-04-10 c4ea8a8
Add deepset layer type
