+++ src/lib/mod_deepset_layer.f90
@@ -1,0 +1,505 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains implementation of a deep setp layer
+!!!#############################################################################
+module deepset_layer
+  use constants, only: real12
+  use base_layer, only: learnable_layer_type
+  use custom_types, only: activation_type, initialiser_type
+  implicit none
+  
+
+!!!-----------------------------------------------------------------------------
+!!! fully connected network layer type
+!!!-----------------------------------------------------------------------------
+  type, extends(learnable_layer_type) :: deepset_layer_type
+     integer :: num_inputs, num_addit_inputs = 0
+     integer :: num_outputs
+     real(real12) :: lambda, gamma, bias
+     real(real12), allocatable, dimension(:) :: dg, dl, db ! weight gradient
+     real(real12), allocatable, dimension(:,:) :: output !output and activation
+     real(real12), allocatable, dimension(:,:) :: di ! input gradient (i.e. delta)
+   contains
+     procedure, pass(this) :: get_num_params => get_num_params_deepset
+     procedure, pass(this) :: get_params => get_params_deepset
+     procedure, pass(this) :: set_params => set_params_deepset
+     procedure, pass(this) :: get_gradients => get_gradients_deepset
+     procedure, pass(this) :: set_gradients => set_gradients_deepset
+     procedure, pass(this) :: get_output => get_output_deepset
+
+     procedure, pass(this) :: print => print_deepset
+     procedure, pass(this) :: init => init_deepset
+     procedure, pass(this) :: set_batch_size => set_batch_size_deepset
+     
+     procedure, pass(this) :: forward  => forward_rank
+     procedure, pass(this) :: backward => backward_rank
+     procedure, private, pass(this) :: forward_2d
+     procedure, private, pass(this) :: backward_2d
+
+     procedure, pass(this) :: reduce => layer_reduction
+     procedure, pass(this) :: merge => layer_merge
+     procedure :: add_t_t => layer_add  !t = type, r = real, i = int
+     generic :: operator(+) => add_t_t !, public
+  end type deepset_layer_type
+
+
+!!!-----------------------------------------------------------------------------
+!!! interface for layer set up
+!!!-----------------------------------------------------------------------------
+  interface deepset_layer_type
+     module function layer_setup( &
+          input_shape, batch_size, &
+          num_inputs, &
+          lambda, gamma) result(layer)
+       integer, dimension(:), optional, intent(in) :: input_shape
+       integer, optional, intent(in) :: num_inputs
+       integer, optional, intent(in) :: batch_size
+       real(real12), optional, intent(in) :: lambda, gamma
+       type(deepset_layer_type) :: layer
+     end function layer_setup
+  end interface deepset_layer_type
+
+
+  private
+  public :: deepset_layer_type
+  public :: read_deepset_layer
+
+
+contains
+
+!!!#############################################################################
+!!! layer reduction
+!!!#############################################################################
+  subroutine layer_reduction(this, rhs)
+    implicit none
+    class(deepset_layer_type), intent(inout) :: this
+    class(learnable_layer_type), intent(in) :: rhs
+
+    select type(rhs)
+    type is(deepset_layer_type)
+       this%dl = this%dl + rhs%dl
+       this%dg = this%dg + rhs%dg
+       this%db = this%db + rhs%db
+    end select
+
+  end subroutine  layer_reduction
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! layer addition
+!!!#############################################################################
+  function layer_add(a, b) result(output)
+    implicit none
[…]
