### 2024-04-12 8068662
Remove batch print
