+++ src/lib/mod_misc_ml.f90
@@ -28,6 +28,7 @@ module misc_ml
   procedure shuffle_1Dilist, &
        shuffle_2Drdata, shuffle_3Didata, shuffle_3Drdata, &
        shuffle_4Drdata, shuffle_5Drdata, &
+       shuffle_2Drdata_1Drlist, &
        shuffle_3Didata_1Dilist, shuffle_3Didata_1Drlist, &
        shuffle_4Drdata_1Dilist, shuffle_5Drdata_1Dilist, shuffle_5Drdata_1Drlist
   end interface shuffle
@@ -52,6 +53,7 @@ module misc_ml
 !!! split_list  = (I, out, opt) index array of split
   interface split
   procedure &
+       split_2Drdata_1Drlist, &
        split_3Didata_1Dilist, split_3Didata_1Drlist, &
        split_5Drdata, &
        split_5Drdata_1Drlist
@@ -488,6 +490,86 @@ do i=1,n_data
 end do
 
 end subroutine shuffle_5Drdata
+!!!-----------------------------------------------------
+!!!-----------------------------------------------------
+subroutine shuffle_2Drdata_1Drlist(data,label,dim,seed,shuffle_list)
+implicit none
+integer :: istart,seed_size
+integer :: i,j,n_data
+real(real12) :: rtmp1
+real(real12) :: r
+integer, dimension(2) :: idx_s,idx_e,jdx_s,jdx_e
+integer, dimension(2,2) :: t_size
+integer, allocatable, dimension(:) :: iseed
+real(real12), allocatable, dimension(:,:) :: tlist
+
+integer, intent(in) :: dim
+real(real12), dimension(:,:), intent(inout) :: data
+real(real12), dimension(:), intent(inout) :: label
+
+integer, optional, intent(in) :: seed
+integer, optional, dimension(size(data,dim)), intent(out) :: shuffle_list
+
+
+!! set or get random seed
+call random_seed(size=seed_size)
+allocate(iseed(seed_size))
+if(present(seed))then
+   iseed = seed
+   call random_seed(put=iseed)
+else
+   call random_seed(get=iseed)
+end if
+
+!! get the size of the data
+n_data = size(data,dim=dim)
+do i=1,2
+  t_size(i,1) = 1
+  jdx_s(i) = 1
+  jdx_e(i) = size(data,dim=i)
+  idx_s(i) = 1
+  idx_e(i) = size(data,dim=i)
+  if(i.eq.dim) then
+     t_size(i,2) = 1
+  else
+     t_size(i,2) = size(data,dim=i)
+  end if
+end do
+
+allocate(tlist(t_size(1,2),t_size(2,2)))
+
+istart=1
+do i=1,n_data
+  call random_number(r)
+  j = istart + floor((n_data+1-istart)*r)
+  if(present(shuffle_list)) shuffle_list(i) = j
+  idx_s(dim) = i
+  idx_e(dim) = i
+  jdx_s(dim) = j
+  jdx_e(dim) = j
+  tlist(&
+       t_size(1,1):t_size(1,2),&
+       t_size(2,1):t_size(2,2)) = data(&
+       idx_s(1):idx_e(1),&
+       idx_s(2):idx_e(2))
+  data(&
+       idx_s(1):idx_e(1),&
+       idx_s(2):idx_e(2)) = data(&
+       jdx_s(1):jdx_e(1),&
+       jdx_s(2):jdx_e(2))
+  data(&
+       jdx_s(1):jdx_e(1),&
+       jdx_s(2):jdx_e(2)) = tlist(&
+       t_size(1,1):t_size(1,2),&
+       t_size(2,1):t_size(2,2))
+
+  rtmp1 = label(i)
+  label(i) = label(j)
+  label(j) = rtmp1
+
+end do
+
[…]
