### 2024-04-12 0e0bef6
Make accuracy conform with loss format
