+++ src/lib/mod_accuracy.f90
@@ -16,10 +16,23 @@ module accuracy
 
   private
 
+  public :: compute_accuracy_function
   public :: categorical_score
-  public :: mae_score, mse_score
+  public :: mae_score, mse_score, rmse_score
   public :: r2_score
 
+
+  abstract interface
+     !! compute the accuracy function
+     !! predicted = (R, in) predicted values
+     !! expected  = (R, in) expected values
+     !! output    = (R, in) accuracy function
+     pure function compute_accuracy_function(predicted, expected) result(output)
+       import real12
+       real(real12), dimension(:,:), intent(in) :: predicted, expected
+       real(real12), dimension(size(predicted,2)) :: output
+     end function compute_accuracy_function
+  end interface
 
 contains
 
@@ -27,20 +40,19 @@ contains
 !!! compute accuracy
 !!! this only works (and is only valid for?) categorisation problems
 !!!#############################################################################
-  function categorical_score(output, expected) result(accuracy)
+  pure function categorical_score(predicted, expected) result(output)
     implicit none
-    real(real12), dimension(:,:), intent(in) :: output
-    integer, dimension(:,:) :: expected
-    real(real12), dimension(size(expected,2)) :: accuracy
+    real(real12), dimension(:,:), intent(in) :: predicted, expected
+    real(real12), dimension(size(expected,2)) :: output
 
     integer :: s
 
     !! Compute the accuracy
     do concurrent(s=1:size(expected,2))
-       if (maxloc(expected(:,s),dim=1).eq.maxloc(output(:,s),dim=1)) then
-          accuracy(s) = 1._real12
+       if (maxloc(expected(:,s),dim=1).eq.maxloc(predicted(:,s),dim=1)) then
+          output(s) = 1._real12
        else
-          accuracy(s) = 0._real12
+          output(s) = 0._real12
        end if
     end do
 
@@ -48,35 +60,49 @@ contains
 !!!-----------------------------------------------------------------------------
 !!!-----------------------------------------------------------------------------
 !!! works for continuous datasets
-  function mae_score(output, expected) result(accuracy)
+  pure function mae_score(predicted, expected) result(output)
     implicit none
-    real(real12), dimension(:,:), intent(in) :: output, expected
-    real(real12), dimension(size(expected,2)) :: accuracy
+    real(real12), dimension(:,:), intent(in) :: predicted, expected
+    real(real12), dimension(size(expected,2)) :: output
 
     !! Compute the accuracy
-    accuracy = 1._real12 - sum(abs(expected - output),dim=1)/size(expected(:,1))
+    output = 1._real12 - sum(abs(expected - predicted),dim=1)/size(expected,1)
 
   end function mae_score
 !!!-----------------------------------------------------------------------------
 !!!-----------------------------------------------------------------------------
 !!! works for continuous datasets
-  function mse_score(output, expected) result(accuracy)
+  pure function mse_score(predicted, expected) result(output)
     implicit none
-    real(real12), dimension(:,:), intent(in) :: output, expected
-    real(real12), dimension(size(expected,2)) :: accuracy
+    real(real12), dimension(:,:), intent(in) :: predicted, expected
+    real(real12), dimension(size(expected,2)) :: output
 
     !! Compute the accuracy
-    accuracy = 1._real12 - sum((expected - output)**2._real12,dim=1)/size(expected(:,1))
+    output = 1._real12 - &
+         sum((expected - predicted)**2._real12,dim=1)/size(expected,1)
 
   end function mse_score
[…]
