### 2024-04-12 71596fc
Add Hubber loss and conv1d layer
