+++ src/lib/mod_container_layer_sub.f90
@@ -8,11 +8,13 @@
 submodule(container_layer) container_layer_submodule
   use base_layer, only: learnable_layer_type, flatten_layer_type
   use input1d_layer, only: input1d_layer_type
+  use input2d_layer, only: input2d_layer_type
   use input3d_layer, only: input3d_layer_type
   use input4d_layer, only: input4d_layer_type
   use batchnorm1d_layer, only: batchnorm1d_layer_type
   use batchnorm2d_layer, only: batchnorm2d_layer_type
   use batchnorm3d_layer, only: batchnorm3d_layer_type
+  use conv1d_layer, only: conv1d_layer_type
   use conv2d_layer, only: conv2d_layer_type
   use conv3d_layer, only: conv3d_layer_type
   use deepset_layer, only: deepset_layer_type
@@ -35,6 +37,8 @@ contains
     select type(previous => input%layer)
     type is(input1d_layer_type)
        call this%layer%forward(previous%output)
+    type is(input2d_layer_type)
+       call this%layer%forward(previous%output)
     type is(input3d_layer_type)
        call this%layer%forward(previous%output)
     type is(input4d_layer_type)
@@ -47,6 +51,8 @@ contains
     type is(batchnorm3d_layer_type)
        call this%layer%forward(previous%output)
 
+    type is(conv1d_layer_type)
+       call this%layer%forward(previous%output)
     type is(conv2d_layer_type)
        call this%layer%forward(previous%output)
     type is(conv3d_layer_type)
@@ -90,6 +96,8 @@ contains
     select type(previous => input%layer)
     type is(input1d_layer_type)
        call this%layer%backward(previous%output, gradient)
+    type is(input2d_layer_type)
+       call this%layer%backward(previous%output, gradient)
     type is(input3d_layer_type)
        call this%layer%backward(previous%output, gradient)
     type is(input4d_layer_type)
@@ -102,6 +110,8 @@ contains
     type is(batchnorm3d_layer_type)
        call this%layer%backward(previous%output, gradient)
 
+    type is(conv1d_layer_type)
+       call this%layer%backward(previous%output, gradient)
     type is(conv2d_layer_type)
        call this%layer%backward(previous%output, gradient)
     type is(conv3d_layer_type)
