+++ src/athena.f90
@@ -49,6 +49,7 @@ module athena
 
   !! input layer types
   use input1d_layer,   only: input1d_layer_type
+  use input2d_layer,   only: input2d_layer_type
   use input3d_layer,   only: input3d_layer_type
   use input4d_layer,   only: input4d_layer_type
 
@@ -58,6 +59,7 @@ module athena
   use batchnorm3d_layer, only: batchnorm3d_layer_type, read_batchnorm3d_layer
 
   !! convolution layer types
+  use conv1d_layer,    only: conv1d_layer_type, read_conv1d_layer
   use conv2d_layer,    only: conv2d_layer_type, read_conv2d_layer
   use conv3d_layer,    only: conv3d_layer_type, read_conv3d_layer
 
