### 2024-04-15 nedtaylor/athena#20 @milancurcic
gfortran-11.4 Internal Compiler Error building test_network.f90
```
...
test_dropblock3d_layer.f90             done.
test_network.f90                       failed.
[ 63%] Compiling...
f951: internal compiler error: Segmentation fault
0x74e33f24251f ???
        ./signal/../sysdeps/unix/sysv/linux/x86_64/libc_sigaction.c:0
0x74e33f229d8f __libc_start_call_main
        ../sysdeps/nptl/libc_start_call_main.h:58
0x74e33f229e3f __libc_start_main_impl
        ../csu/libc-start.c:392
Please submit a full bug report,
with preprocessed source if appropriate.
Please include the complete backtrace with any bug report.
See <file:///usr/share/doc/gcc-11/README.Bugs> for instructions.
<ERROR> Compilation failed for object " test_test_network.f90.o "
<ERROR> stopping due to failed compilation
STOP 1
```

I can see that the README states that the lowest tested gfortran is v13, and that the library is possibly not expected to build with lower version. Reporting here nevertheless to document the issue.
