### 2024-04-15 nedtaylor/athena#20 @milancurcic
The issue occurs with `--profile debug` as well.
