### 2024-04-15 nedtaylor/athena#20 @milancurcic
Correction, v12.3 doesn't encounter this issue.
