### 2024-04-15 nedtaylor/athena#21 @milancurcic
Error compiling mod_base_layer.f90 with ifort and ifx
With ifort-2021.10:

```
mod_base_layer.f90                     failed.
[ 18%] Compiling...
././src/lib/mod_base_layer.f90(117): error #6118: The interface block containing the separate interface body must not be abstract.
     module subroutine initialise(this, input_shape, batch_size, verbose)
-----^
/tmp/ifort9QfSWv.i90(424): catastrophic error: Too many errors, exiting
compilation aborted for ././src/lib/mod_base_layer.f90 (code 1)
<ERROR> Compilation failed for object " src_lib_mod_base_layer.f90.o "
<ERROR> stopping due to failed compilation
STOP 1
```

With ifx-2023.2:

```
mod_base_layer.f90                     failed.
[ 18%] Compiling...
././src/lib/mod_base_layer.f90(117): error #6118: The interface block containing the separate interface body must not be abstract.
     module subroutine initialise(this, input_shape, batch_size, verbose)
-----^
/tmp/ifx15633005958Xufq1/ifxtP3YAF.i90(424): catastrophic error: Too many errors, exiting
compilation aborted for ././src/lib/mod_base_layer.f90 (code 1)
<ERROR> Compilation failed for object " src_lib_mod_base_layer.f90.o "
<ERROR> stopping due to failed compilation
STOP 1
```

Part of https://github.com/openjournals/joss-reviews/issues/6492
