### 2024-04-16 29d05ae
Fix abstract interface
