### 2024-04-16 ac47dc2
Fix optimiser module variable name conflict
