+++ test/test_optimiser.f90
@@ -4,7 +4,7 @@ program test_optimiser
   use learning_rate_decay
   implicit none
 
-  class(base_optimiser_type), allocatable :: optimiser
+  class(base_optimiser_type), allocatable :: optimiser_var
 
 
   integer, parameter :: num_params = 10
@@ -15,36 +15,36 @@ program test_optimiser
 
 
 !!!-----------------------------------------------------------------------------
-!!! test empty base optimiser
+!!! test empty base optimiser_var
 !!!-----------------------------------------------------------------------------
-  allocate(optimiser, source=base_optimiser_type())
+  allocate(optimiser_var, source=base_optimiser_type())
 
 
 !!!-----------------------------------------------------------------------------
 !!! test base optimiser
 !!!-----------------------------------------------------------------------------
-  deallocate(optimiser)
-  allocate(optimiser, source=base_optimiser_type( &
+  deallocate(optimiser_var)
+  allocate(optimiser_var, source=base_optimiser_type( &
        learning_rate = 0.1, &
        num_params = num_params, &
        lr_decay = base_lr_decay_type(), &
        regulariser=l1_regulariser_type() &
        ))
-  select type(optimiser)
+  select type(optimiser_var)
   type is (base_optimiser_type)
      write(*,*) "base_optimiser_type"
   class default
      write(0,*) "Failed to allocate base optimiser"
      success = .false.
   end select
-  call optimiser%minimise(param, gradient)
+  call optimiser_var%minimise(param, gradient)
 
 
 !!!-----------------------------------------------------------------------------
 !!! test sgd optimiser
 !!!-----------------------------------------------------------------------------
-  deallocate(optimiser)
-  allocate(optimiser, source=sgd_optimiser_type( &
+  deallocate(optimiser_var)
+  allocate(optimiser_var, source=sgd_optimiser_type( &
        learning_rate = 0.1, &
        num_params = num_params, &
        momentum = 0.9, &
@@ -52,21 +52,21 @@ program test_optimiser
        lr_decay = base_lr_decay_type(), &
        regulariser=l1_regulariser_type() &
        ))
-  select type(optimiser)
+  select type(optimiser_var)
   type is (sgd_optimiser_type)
      write(*,*) "sgd_optimiser_type"
   class default
      write(0,*) "Failed to allocate sgd optimiser"
      success = .false.
   end select
-  call optimiser%minimise(param, gradient)
+  call optimiser_var%minimise(param, gradient)
 
 
 !!!-----------------------------------------------------------------------------
 !!! test rmsprop optimiser
 !!!-----------------------------------------------------------------------------
-  deallocate(optimiser)
-  allocate(optimiser, source=rmsprop_optimiser_type( &
+  deallocate(optimiser_var)
+  allocate(optimiser_var, source=rmsprop_optimiser_type( &
        learning_rate = 0.1, &
        num_params = num_params, &
        beta = 0.9, &
@@ -74,42 +74,42 @@ program test_optimiser
        lr_decay = base_lr_decay_type(), &
        regulariser=l1_regulariser_type() &
        ))
-  select type(optimiser)
+  select type(optimiser_var)
   type is (rmsprop_optimiser_type)
      write(*,*) "rmsprop_optimiser_type"
   class default
      write(0,*) "Failed to allocate rmsprop optimiser"
      success = .false.
   end select
-  call optimiser%minimise(param, gradient)
+  call optimiser_var%minimise(param, gradient)
 
 
 !!!-----------------------------------------------------------------------------
 !!! test adagrad optimiser
 !!!-----------------------------------------------------------------------------
-  deallocate(optimiser)
[…]
