### 2024-04-16 5b21440
Fix activation module variable name conflict
