+++ test/test_activations.f90
@@ -14,7 +14,7 @@ program test_activations
    implicit none
  
    class(base_layer_type), allocatable :: full_layer, conv2d_layer, conv3d_layer
-   class(activation_type), allocatable :: activation
+   class(activation_type), allocatable :: activation_var
    logical :: success = .true.
  
    integer :: i
@@ -83,12 +83,12 @@ program test_activations
 !!!-----------------------------------------------------------------------------
 !!! check gaussian setup
 !!!-----------------------------------------------------------------------------
-   activation = gaussian_setup(threshold = 2.E0, sigma = 2.E0)
-   if(.not. activation%name .eq. 'gaussian')then
+   activation_var = gaussian_setup(threshold = 2.E0, sigma = 2.E0)
+   if(.not. activation_var%name .eq. 'gaussian')then
       success = .false.
       write(0,*) 'activation has wrong name for gaussian'
    else
-      if (activation%threshold .ne. 2.E0) then
+      if (activation_var%threshold .ne. 2.E0) then
          success = .false.
          write(0,*) 'activation has wrong threshold for gaussian'
       end if
@@ -98,8 +98,8 @@ program test_activations
 !!!-----------------------------------------------------------------------------
 !!! check piecewise setup
 !!!-----------------------------------------------------------------------------
-   activation = piecewise_setup(intercept = 2.E0)
-   if(.not. activation%name .eq. 'piecewise')then
+   activation_var = piecewise_setup(intercept = 2.E0)
+   if(.not. activation_var%name .eq. 'piecewise')then
       success = .false.
       write(0,*) 'activation has wrong name for piecewise'
    end if
@@ -108,12 +108,12 @@ program test_activations
 !!!-----------------------------------------------------------------------------
 !!! check sigmoid setup
 !!!-----------------------------------------------------------------------------
-   activation = sigmoid_setup(threshold = 2.E0)
-   if(.not. activation%name .eq. 'sigmoid')then
+   activation_var = sigmoid_setup(threshold = 2.E0)
+   if(.not. activation_var%name .eq. 'sigmoid')then
       success = .false.
       write(0,*) 'activation has wrong name for sigmoid'
    else
-      if (activation%threshold .ne. 2.E0) then
+      if (activation_var%threshold .ne. 2.E0) then
          success = .false.
          write(0,*) 'activation has wrong threshold for sigmoid'
       end if
@@ -123,12 +123,12 @@ program test_activations
 !!!-----------------------------------------------------------------------------
 !!! check softmax setup
 !!!-----------------------------------------------------------------------------
-   activation = softmax_setup(threshold = 2.E0)
-   if(.not. activation%name .eq. 'softmax')then
+   activation_var = softmax_setup(threshold = 2.E0)
+   if(.not. activation_var%name .eq. 'softmax')then
       success = .false.
       write(0,*) 'activation has wrong name for softmax'
    else
-      if (activation%threshold .ne. 2.E0) then
+      if (activation_var%threshold .ne. 2.E0) then
          success = .false.
          write(0,*) 'activation has wrong threshold for softmax'
       end if
@@ -138,12 +138,12 @@ program test_activations
 !!!-----------------------------------------------------------------------------
 !!! check tanh setup
 !!!-----------------------------------------------------------------------------
-   activation = tanh_setup(threshold = 2.E0)
-   if(.not. activation%name .eq. 'tanh')then
+   activation_var = tanh_setup(threshold = 2.E0)
+   if(.not. activation_var%name .eq. 'tanh')then
       success = .false.
       write(0,*) 'activation has wrong name for tanh'
    else
-      if (activation%threshold .ne. 2.E0) then
[…]
