### 2024-04-16 43362b6
Remove coverage compiler flag option for intel
