### 2024-04-16 nedtaylor/athena#21 @nedtaylor
This is now fixed in commit 29d05ae2d26afc3f54883a90dbd322de1ff27512

Other issues have arisen which have been resolved in branch 21-ifx-compilation-error
