### 2024-04-19 nedtaylor/athena#21 @milancurcic
Thank @nedtaylor, can you keep open until this is merged into `main`, for the sake of bookkeeping for https://github.com/openjournals/joss-reviews/issues/6492.
