### 2024-04-19 nedtaylor/athena#23 @milancurcic
CONTRIBUTING.md was largely copied from neural-fortran
CONTRIBUTING.md, as of 0d3f02bef9439f8b3fd0355463fd8b0a9fe8dac9, seems copied from https://github.com/modern-fortran/neural-fortran/blob/main/CONTRIBUTING.md and only edited to change the occasional wording and to account for different features of the library. For example, the layout, sections, and sentence structures and semantics are almost exactly the same.

Personally, as creator of neural-fortran and its CONTRIBUTING.md doc, I don't mind as long as you reference it at the top (e.g. "This document was adapted from https://github.com/modern-fortran/neural-fortran/blob/main/CONTRIBUTING.md", or similar). But given that the software and its docs are part of the submission itself, which is expected to be original, this may be problematic.

This CONTRIBUTING.md is not the kind of generic CONTRIBUTING guide for community projects (in which case a copied CONTRIBUTING.md may not be a problem), but bespoke and very specific to the library.

@haozeke what's your take or is there a JOSS guideline for this? If it's problematic, a simple solution could be to simply rewrite CONTRIBUTING.md in original language.

To be clear, I don't think this was intentional but merely an honest omission by @nedtaylor.

I see similar issues with parts of the code, but I'll open separate issues for those.

Part of https://github.com/openjournals/joss-reviews/issues/6492
