### 2024-04-19 nedtaylor/athena#23 @nedtaylor
You are completely correct, @milancurcic. I used the neural-fortran as a template. I definitely should have referenced your original document in the CONTRIBUTING.md file (as I had not written one before, I thought it was a good one to use as an example).

@HaoZeke, as Milan has mentioned, if the CONTRIBUTING.md needs to be clearly original, then I am happy to rewrite it and any other offending parts of the code.

@milancurcic I am sorry about that. I will include references in CONTRIBUTING.md (and other parts of the code) where clear inspiration has been taken from neural-fortran. This is why I felt it good to have you as a one of the referees to ascertain whether ATHENA is different enough to be worth being its own code or whether it should be integrated into neural-fortran.
