### 2024-04-19 nedtaylor/athena#23 @HaoZeke
I would prefer to have clearer attribution than obfuscation just for JOSS. Since these are both MIT licensed codes, I think providing attribution is enough. Note that this doesn't extend to the article itself or the principal claims of the software, which should stand-alone as a scientific contribution (and will be determined separately by the reviewers).
