### 2024-04-19 nedtaylor/athena#23 @milancurcic
I agree with the approach suggested by @HaoZeke.
