### 2024-04-19 nedtaylor/athena#23 @nedtaylor
Thank you for the clarification here. I have now included a reference in commit 35ff066 in the `development` branch.

I will start combing through the rest of the code to provide more explicit references to neural-fortran (but as you said, that's for a the opening of a different issue).
