### 2024-04-19 7ff9512
Include attribution statements
