### 2024-04-20 33adc77
Add graph type
