+++ src/lib/mod_cfconv1d_layer.f90
@@ -1,0 +1,949 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains implementation of a 1D continuous filter convolutional layer
+!!!#############################################################################
+module cfconv1d_layer
+  use constants, only: real12
+  use base_layer, only: learnable_layer_type, conv_layer_type
+  use custom_types, only: initialiser_type
+  implicit none
+  
+  
+  type, extends(conv_layer_type) :: cfconv1d_layer_type
+     real(real12), allocatable, dimension(:,:,:) :: weight !!!REMOVE
+     real(real12), allocatable, dimension(:,:,:) :: location
+     real(real12), allocatable, dimension(:,:,:,:) :: dw ! weight gradient
+     real(real12), allocatable, dimension(:,:,:) :: output, z
+     real(real12), allocatable, dimension(:,:,:) :: di ! input gradient
+   contains
+     procedure, pass(this) :: get_params => get_params_cfconv1d
+     procedure, pass(this) :: set_params => set_params_cfconv1d
+     procedure, pass(this) :: get_gradients => get_gradients_cfconv1d
+     procedure, pass(this) :: set_gradients => set_gradients_cfconv1d
+     procedure, pass(this) :: get_output => get_output_cfconv1d
+
+     procedure, pass(this) :: init => init_cfconv1d
+     procedure, pass(this) :: set_batch_size => set_batch_size_cfconv1d
+     procedure, pass(this) :: print => print_cfconv1d
+
+     procedure, pass(this) :: forward  => forward_rank
+     procedure, pass(this) :: backward => backward_rank
+     procedure, private, pass(this) :: forward_3d
+     procedure, private, pass(this) :: backward_3d
+
+     procedure, pass(this) :: filter => null()
+     
+     procedure, pass(this) :: reduce => layer_reduction
+     procedure, pass(this) :: merge => layer_merge
+     procedure :: add_t_t => layer_add  !t = type, r = real, i = int
+     generic :: operator(+) => add_t_t !, public
+  end type cfconv1d_layer_type
+
+  
+!!!-----------------------------------------------------------------------------
+!!! interface for layer set up
+!!!-----------------------------------------------------------------------------
+  interface cfconv1d_layer_type
+     module function layer_setup( &
+          input_shape, batch_size, &
+          num_filters, kernel_size, stride, padding, &
+          activation_function, activation_scale, &
+          kernel_initialiser, bias_initialiser, &
+          calc_input_gradients) result(layer)
+       integer, dimension(:), optional, intent(in) :: input_shape
+       integer, optional, intent(in) :: batch_size
+       integer, optional, intent(in) :: num_filters
+       integer, dimension(..), optional, intent(in) :: kernel_size
+       integer, dimension(..), optional, intent(in) :: stride
+       real(real12), optional, intent(in) :: activation_scale
+       character(*), optional, intent(in) :: activation_function, &
+            kernel_initialiser, bias_initialiser, padding
+       logical, optional, intent(in) :: calc_input_gradients
+       type(cfconv1d_layer_type) :: layer
+     end function layer_setup
+  end interface cfconv1d_layer_type
+
+
+  private
+  public :: cfconv1d_layer_type
+  public :: read_cfconv1d_layer
+
+
+contains
+
+!!!#############################################################################
+!!! layer reduction
+!!!#############################################################################
+  subroutine layer_reduction(this, rhs)
+    implicit none
+    class(cfconv1d_layer_type), intent(inout) :: this
+    class(learnable_layer_type), intent(in) :: rhs
+
+    select type(rhs)
+    class is(cfconv1d_layer_type)
[…]
