+++ src/lib/mod_message_passing_layer.f90
@@ -15,14 +15,24 @@ module mpnn_module
   public :: mpnn_type
 
   type :: mpnn_type
-     real(real12), dimension(:), allocatable :: message
+     integer :: num_vertices
+     integer :: num_time_steps
+     !! hidden features has dimensions (vertex, time step)
+     type(feature_type), dimension(:,:), allocatable :: hidden
+!     type(feature_type), dimension(:,:), allocatable :: hidden_vertex
+!     type(feature_type), dimension(:,:), allocatable :: hidden_edge
+     !! message has dimensions (vertex, time step)
+     type(feature_type), dimension(:,:), allocatable :: message
    contains
-     procedure(message_function), pass(this), pointer :: message_function => null()
-     procedure, pass(this) :: message_function
-     procedure, pass(this) :: update_function
-     procedure, pass(this) :: readout_function
+     procedure(message_update), pass(this), pointer :: message_update => null()
+     procedure(state_update), pass(this), pointer :: state_update => null()
+     procedure, pass(this) :: readout
      procedure, pass(this) :: forward
   end type mpnn_type
+
+  type :: feature_type
+     real(real12), dimension(:), allocatable :: feature
+  end type feature_type
 
   
 !!!-----------------------------------------------------------------------------
@@ -50,15 +60,17 @@ module mpnn_module
 
 
   abstract interface
-     !! compute the loss function
-     !! predicted = (R, in) predicted values
-     !! expected  = (R, in) expected values
-     !! output    = (R, in) loss function
-     pure subroutine message_function(this, graph)
+     pure subroutine message_update(this, graph)
        implicit none
        class(mpnn_type), intent(in) :: this
        type(graph_type), intent(in) :: graph
-     end subroutine message_function
+     end subroutine message_update
+
+     pure subroutine state_update(this, graph)
+       implicit none
+       class(mpnn_type), intent(in) :: this
+       type(graph_type), intent(in) :: graph
+     end subroutine state_update
   end interface
   
 
@@ -68,20 +80,64 @@ contains
 !!!#############################################################################
 !!! message function
 !!!#############################################################################
-  pure subroutine convolutional_message_function(this, graph)
+  pure subroutine convolutional_message_update(this, graph, time_step)
+    implicit none
+    class(mpnn_type), intent(inout) :: this
+    type(graph_type), intent(in) :: graph
+    integer, intent(in) :: time_step
+
+    integer :: i, j
+    integer :: num_features
+
+    !! assume all hidden vertices for one time_step have the same number of features
+    num_features = size(this%hidden(1,time_step)%feature)
+    do i = 1, graph%num_vertices
+       allocate(this%message(i,time_step+1)%feature(num_features), source=0._real12)
+       do j = 1, graph%num_vertices
+          this%message(i,time_step+1)%feature = &
+               this%message(i,time_step+1)%feature + &
+               [ this%hidden(j,time_step), graph%edge(i,j)%feature ]
+               ![ graph%vertex(j)%feature, graph%edge(i,j)%feature ]
+               ! at time step 1, set hidden to graph vertex features
+       end do
+    end do
+
+  end subroutine convolutional_message_update
+
+  pure subroutine convolutional_state_update(this, graph, time_step)
+    implicit none
+    class(mpnn_type), intent(inout) :: this
+    type(graph_type), intent(in) :: graph
+    integer, intent(in) :: time_step
+
+    integer :: i
+
+    do i = 1, graph%num_vertices
+       this%hidden(:,i,time_step+1) = sigmoid( &
+            matmul( this%update_matrix(graph%get_degree(i),time_step), &
+                    this%message(i,time_step+1) ) )
+    end do
+
[…]
