+++ src/lib/mod_message_passing_layer.f90
@@ -6,7 +6,7 @@
 !!!#############################################################################
 module mpnn_module
   use constants, only: real12
-  use custom_types, only: initialiser_type, graph_type
+  use custom_types, only: graph_type
   implicit none
   
 
@@ -14,25 +14,41 @@ module mpnn_module
 
   public :: mpnn_type
 
-  type :: mpnn_type
-     integer :: num_vertices
-     integer :: num_time_steps
-     !! hidden features has dimensions (vertex, time step)
-     type(feature_type), dimension(:,:), allocatable :: hidden
-!     type(feature_type), dimension(:,:), allocatable :: hidden_vertex
-!     type(feature_type), dimension(:,:), allocatable :: hidden_edge
-     !! message has dimensions (vertex, time step)
-     type(feature_type), dimension(:,:), allocatable :: message
-   contains
-     procedure(message_update), pass(this), pointer :: message_update => null()
-     procedure(state_update), pass(this), pointer :: state_update => null()
-     procedure, pass(this) :: readout
-     procedure, pass(this) :: forward
-  end type mpnn_type
-
   type :: feature_type
      real(real12), dimension(:), allocatable :: feature
   end type feature_type
+
+
+  type :: mpnn_type
+     integer :: num_features
+     integer :: num_vertices
+     integer :: num_time_steps
+     integer :: batch_size
+     class(state_method_type), allocatable :: state
+     class(message_method_type), allocatable :: message
+     class(readout_method_type), allocatable :: readout
+     real(real12), dimension(:,:), allocatable :: output
+     real(real12), dimension(:,:,:,:), allocatable :: di
+     !! hidden features has dimensions (vertex, time step, batch_size)
+     !type(feature_type), dimension(:,:,:), allocatable :: hidden
+!     type(feature_type), dimension(:,:), allocatable :: hidden_vertex
+!     type(feature_type), dimension(:,:), allocatable :: hidden_edge
+     !! message has dimensions (vertex, time step, batch_size)
+     !type(feature_type), dimension(:,:,:), allocatable :: message
+     !type(feature_type), dimension(:,:,:), allocatable :: weight
+     !! output has dimensions (num_outputs, batch_size)
+     !! v and dw have dimensions (num_features_t, num_features_t, time_step, batch_size)
+     !real(real12), dimension(:,:,:,:), allocatable :: v
+     !real(real12), dimension(:,:,:,:), allocatable :: dw
+     !! di has dimensions (feature, vertex, time_step, batch_size)
+     !procedure(message_update), pass(this), pointer :: message_update => null()
+     !procedure(state_update), pass(this), pointer :: state_update => null()
+     !procedure(readout), pass(this), pointer :: readout => null()
+   contains
+     procedure, pass(this) :: forward
+     procedure, pass(this) :: backward
+  end type mpnn_type
+
 
   
 !!!-----------------------------------------------------------------------------
@@ -59,102 +75,187 @@ module mpnn_module
   ! end interface mpnn_type
 
 
+  type, abstract :: state_method_type
+     !! feature has dimensions (feature, vertex, time_step, batch_size)
+     real(real12), dimension(:,:,:,:), allocatable :: feature     
+   contains
+     procedure(state_update), deferred, pass(this) :: update
+     procedure(get_state_differential), deferred, pass(this) :: get_differential
+  end type state_method_type
+
+  type, abstract :: message_method_type
+     !! feature has dimensions (feature, vertex, time_step, batch_size)
+     real(real12), dimension(:,:,:,:), allocatable :: feature     
+   contains
+     procedure(message_update), deferred, pass(this) :: update
+     procedure(get_message_differential), deferred, pass(this) :: get_differential
+  end type message_method_type
+
[…]
