+++ src/lib/mod_message_passing_layer.f90
@@ -24,8 +24,9 @@ module mpnn_module
      integer :: num_vertices
      integer :: num_time_steps
      integer :: batch_size
-     class(state_method_type), allocatable :: state
-     class(message_method_type), allocatable :: message
+     !! state and message dimension is (time_step)
+     class(state_method_type), dimension(:), allocatable :: state
+     class(message_method_type), dimension(:), allocatable :: message
      class(readout_method_type), allocatable :: readout
      real(real12), dimension(:,:), allocatable :: output
      real(real12), dimension(:,:,:,:), allocatable :: di
@@ -50,34 +51,11 @@ module mpnn_module
   end type mpnn_type
 
 
-  
-!!!-----------------------------------------------------------------------------
-!!! interface for layer set up
-!!!-----------------------------------------------------------------------------
-  ! interface mpnn_type
-  !    module function layer_setup( &
-  !         input_shape, batch_size, &
-  !         num_filters, kernel_size, stride, padding, &
-  !         activation_function, activation_scale, &
-  !         kernel_initialiser, bias_initialiser, &
-  !         calc_input_gradients) result(layer)
-  !      integer, dimension(:), optional, intent(in) :: input_shape
-  !      integer, optional, intent(in) :: batch_size
-  !      integer, optional, intent(in) :: num_filters
-  !      integer, dimension(..), optional, intent(in) :: kernel_size
-  !      integer, dimension(..), optional, intent(in) :: stride
-  !      real(real12), optional, intent(in) :: activation_scale
-  !      character(*), optional, intent(in) :: activation_function, &
-  !           kernel_initialiser, bias_initialiser, padding
-  !      logical, optional, intent(in) :: calc_input_gradients
-  !      type(conv1d_layer_type) :: layer
-  !    end function layer_setup
-  ! end interface mpnn_type
 
 
   type, abstract :: state_method_type
      !! feature has dimensions (feature, vertex, time_step, batch_size)
-     real(real12), dimension(:,:,:,:), allocatable :: feature     
+     real(real12), dimension(:,:,:), allocatable :: feature     
    contains
      procedure(state_update), deferred, pass(this) :: update
      procedure(get_state_differential), deferred, pass(this) :: get_differential
@@ -85,7 +63,7 @@ module mpnn_module
 
   type, abstract :: message_method_type
      !! feature has dimensions (feature, vertex, time_step, batch_size)
-     real(real12), dimension(:,:,:,:), allocatable :: feature     
+     real(real12), dimension(:,:,:), allocatable :: feature     
    contains
      procedure(message_update), deferred, pass(this) :: update
      procedure(get_message_differential), deferred, pass(this) :: get_differential
@@ -99,13 +77,12 @@ module mpnn_module
 
 
   abstract interface
-     subroutine message_update(this, hidden, graph, time_step)
+     subroutine message_update(this, hidden, graph)
        import :: message_method_type, real12, graph_type
        class(message_method_type), intent(inout) :: this
        !! hidden features has dimensions (feature, vertex, batch_size)
        real(real12), dimension(:,:,:), intent(in) :: hidden
        type(graph_type), dimension(:), intent(in) :: graph
-       integer, intent(in) :: time_step
      end subroutine message_update
 
      pure function get_message_differential(this, hidden, graph)
@@ -117,12 +94,11 @@ module mpnn_module
      end function get_message_differential
 
 
-     subroutine state_update(this, message, time_step)
+     subroutine state_update(this, message)
        import :: state_method_type, real12
        class(state_method_type), intent(inout) :: this
        !! message has dimensions (feature, vertex, batch_size)
[…]
