+++ src/lib/mod_message_passing_layer.f90
@@ -2,7 +2,7 @@
 !!! Code written by Ned Thaddeus Taylor
 !!! Code part of the ATHENA library - a feedforward neural network library
 !!!#############################################################################
-!!! module contains implementation of a 1D convolutional layer
+!!! module contains implementation of a message passing neural network
 !!!#############################################################################
 module mpnn_module
   use constants, only: real12
@@ -12,11 +12,11 @@ module mpnn_module
 
   private
 
-  public :: mpnn_type
-
-  type :: feature_type
-     real(real12), dimension(:), allocatable :: feature
-  end type feature_type
+  public :: mpnn_type, feature_type
+  public :: state_method_type, message_method_type, readout_method_type
+  public :: state_update, get_state_differential
+  public :: message_update, get_message_differential
+  public :: get_readout_output, get_readout_differential
 
 
   type :: mpnn_type
@@ -30,46 +30,42 @@ module mpnn_module
      class(readout_method_type), allocatable :: readout
      real(real12), dimension(:,:), allocatable :: output
      real(real12), dimension(:,:,:,:), allocatable :: di
-     !! hidden features has dimensions (vertex, time step, batch_size)
-     !type(feature_type), dimension(:,:,:), allocatable :: hidden
-!     type(feature_type), dimension(:,:), allocatable :: hidden_vertex
-!     type(feature_type), dimension(:,:), allocatable :: hidden_edge
-     !! message has dimensions (vertex, time step, batch_size)
-     !type(feature_type), dimension(:,:,:), allocatable :: message
-     !type(feature_type), dimension(:,:,:), allocatable :: weight
-     !! output has dimensions (num_outputs, batch_size)
-     !! v and dw have dimensions (num_features_t, num_features_t, time_step, batch_size)
-     !real(real12), dimension(:,:,:,:), allocatable :: v
-     !real(real12), dimension(:,:,:,:), allocatable :: dw
-     !! di has dimensions (feature, vertex, time_step, batch_size)
-     !procedure(message_update), pass(this), pointer :: message_update => null()
-     !procedure(state_update), pass(this), pointer :: state_update => null()
-     !procedure(readout), pass(this), pointer :: readout => null()
    contains
      procedure, pass(this) :: forward
      procedure, pass(this) :: backward
   end type mpnn_type
 
 
+  type :: feature_type
+     real(real12), dimension(:,:), allocatable :: val
+  end type feature_type
 
 
   type, abstract :: state_method_type
-     !! feature has dimensions (feature, vertex, time_step, batch_size)
-     real(real12), dimension(:,:,:), allocatable :: feature     
+     integer :: batch_size
+     integer :: num_features
+     !! feature has dimensions (feature, vertex)
+     type(feature_type), dimension(:), allocatable :: feature
+     type(feature_type), dimension(:), allocatable :: di
    contains
      procedure(state_update), deferred, pass(this) :: update
      procedure(get_state_differential), deferred, pass(this) :: get_differential
   end type state_method_type
 
   type, abstract :: message_method_type
-     !! feature has dimensions (feature, vertex, time_step, batch_size)
-     real(real12), dimension(:,:,:), allocatable :: feature     
+     integer :: batch_size
+     integer :: num_features
+     !! feature has dimensions (feature, vertex)
+     type(feature_type), dimension(:), allocatable :: feature
+     type(feature_type), dimension(:), allocatable :: di
    contains
      procedure(message_update), deferred, pass(this) :: update
      procedure(get_message_differential), deferred, pass(this) :: get_differential
   end type message_method_type
 
-  type, abstract :: readout_method_type    
[…]
