### 2024-04-23 d1a9e24
Ad operator handling
