+++ src/lib/mod_message_passing_layer.f90
@@ -38,12 +38,18 @@ module mpnn_module
 
   type :: feature_type
      real(real12), dimension(:,:), allocatable :: val
+   contains
+     ! t = type, r = real, i = int
+     procedure :: add_t_t => feature_add
+     procedure :: multiply_t_t => feature_multiply
+     generic :: operator(+) => add_t_t
+     generic :: operator(*) => multiply_t_t
   end type feature_type
 
 
   type, abstract :: state_method_type
-     integer :: batch_size
      integer :: num_features
+     integer :: batch_size
      !! feature has dimensions (feature, vertex)
      type(feature_type), dimension(:), allocatable :: feature
      type(feature_type), dimension(:), allocatable :: di
@@ -53,8 +59,8 @@ module mpnn_module
   end type state_method_type
 
   type, abstract :: message_method_type
-     integer :: batch_size
      integer :: num_features
+     integer :: batch_size
      !! feature has dimensions (feature, vertex)
      type(feature_type), dimension(:), allocatable :: feature
      type(feature_type), dimension(:), allocatable :: di
@@ -74,38 +80,38 @@ module mpnn_module
 
   abstract interface
      subroutine state_update(this, message, graph)
-       import :: state_method_type, real12, graph_type
+       import :: state_method_type, feature_type, graph_type
        class(state_method_type), intent(inout) :: this
        !! message has dimensions (feature, vertex, batch_size)
-       real(real12), dimension(:,:), intent(in) :: message
+       type(feature_type), dimension(this%batch_size), intent(in) :: message
        type(graph_type), dimension(:), intent(in) :: graph
      end subroutine state_update
 
      pure function get_state_differential(this, message, graph) result(output)
-       import :: state_method_type, real12, graph_type
+       import :: state_method_type, feature_type, graph_type
        class(state_method_type), intent(in) :: this
        !! message has dimensions (feature, vertex, batch_size)
-       real(real12), dimension(:,:), intent(in) :: message
+       type(feature_type), dimension(this%batch_size), intent(in) :: message
        type(graph_type), dimension(:), intent(in) :: graph
-       real(real12), dimension(:,:), allocatable :: output
+       type(feature_type), dimension(this%batch_size) :: output
      end function get_state_differential
 
 
      subroutine message_update(this, hidden, graph)
-       import :: message_method_type, real12, graph_type
+       import :: message_method_type, feature_type, graph_type
        class(message_method_type), intent(inout) :: this
        !! hidden features has dimensions (feature, vertex, batch_size)
-       real(real12), dimension(:,:), intent(in) :: hidden
+       type(feature_type), dimension(this%batch_size), intent(in) :: hidden
        type(graph_type), dimension(:), intent(in) :: graph
      end subroutine message_update
 
      pure function get_message_differential(this, hidden, graph) result(output)
-       import :: message_method_type, real12, graph_type
+       import :: message_method_type, feature_type, graph_type
        class(message_method_type), intent(in) :: this
        !! hidden features has dimensions (feature, vertex, batch_size)
-       real(real12), dimension(:,:), intent(in) :: hidden
+       type(feature_type), dimension(this%batch_size), intent(in) :: hidden
        type(graph_type), dimension(:), intent(in) :: graph
-       real(real12), dimension(:,:), allocatable :: output
+       type(feature_type), dimension(this%batch_size) :: output
      end function get_message_differential
 
 
@@ -116,11 +122,12 @@ module mpnn_module
        real(real12), dimension(:,:), allocatable :: output
[…]
