### 2024-04-23 2071572
Add calculate_partials procedure
