+++ src/lib/mod_message_passing_layer.f90
@@ -25,8 +25,8 @@ module mpnn_module
      integer :: num_time_steps
      integer :: batch_size
      !! state and message dimension is (time_step)
-     class(state_method_type), dimension(:), allocatable :: state
      class(message_method_type), dimension(:), allocatable :: message
+     class(state_method_type), dimension(:), allocatable :: state
      class(readout_method_type), allocatable :: readout
      real(real12), dimension(:,:), allocatable :: output
      real(real12), dimension(:,:,:,:), allocatable :: di
@@ -47,6 +47,18 @@ module mpnn_module
   end type feature_type
 
 
+   type, abstract :: message_method_type
+     integer :: num_features
+     integer :: batch_size
+     !! feature has dimensions (feature, vertex)
+     type(feature_type), dimension(:), allocatable :: feature
+     type(feature_type), dimension(:), allocatable :: di
+   contains
+     procedure(message_update), deferred, pass(this) :: update
+     procedure(get_message_differential), deferred, pass(this) :: get_differential
+     procedure(calculate_message_partials), deferred, pass(this) :: calculate_partials
+  end type message_method_type
+
   type, abstract :: state_method_type
      integer :: num_features
      integer :: batch_size
@@ -56,53 +68,27 @@ module mpnn_module
    contains
      procedure(state_update), deferred, pass(this) :: update
      procedure(get_state_differential), deferred, pass(this) :: get_differential
+     procedure(calculate_state_partials), deferred, pass(this) :: calculate_partials
   end type state_method_type
-
-  type, abstract :: message_method_type
-     integer :: num_features
-     integer :: batch_size
-     !! feature has dimensions (feature, vertex)
-     type(feature_type), dimension(:), allocatable :: feature
-     type(feature_type), dimension(:), allocatable :: di
-   contains
-     procedure(message_update), deferred, pass(this) :: update
-     procedure(get_message_differential), deferred, pass(this) :: get_differential
-  end type message_method_type
 
   type, abstract :: readout_method_type
      integer :: batch_size
+     integer :: num_time_steps
      integer :: num_outputs
    contains
      procedure(get_readout_output), deferred, pass(this) :: get_output
      procedure(get_readout_differential), deferred, pass(this) :: get_differential
+     procedure(calculate_readout_partials), deferred, pass(this) :: calculate_partials
   end type readout_method_type
 
 
   abstract interface
-     subroutine state_update(this, message, graph)
-       import :: state_method_type, feature_type, graph_type
-       class(state_method_type), intent(inout) :: this
-       !! message has dimensions (feature, vertex, batch_size)
-       type(feature_type), dimension(this%batch_size), intent(in) :: message
-       type(graph_type), dimension(:), intent(in) :: graph
-     end subroutine state_update
-
-     pure function get_state_differential(this, message, graph) result(output)
-       import :: state_method_type, feature_type, graph_type
-       class(state_method_type), intent(in) :: this
-       !! message has dimensions (feature, vertex, batch_size)
-       type(feature_type), dimension(this%batch_size), intent(in) :: message
-       type(graph_type), dimension(:), intent(in) :: graph
-       type(feature_type), dimension(this%batch_size) :: output
-     end function get_state_differential
-
-
      subroutine message_update(this, hidden, graph)
        import :: message_method_type, feature_type, graph_type
        class(message_method_type), intent(inout) :: this
        !! hidden features has dimensions (feature, vertex, batch_size)
        type(feature_type), dimension(this%batch_size), intent(in) :: hidden
-       type(graph_type), dimension(:), intent(in) :: graph
+       type(graph_type), dimension(this%batch_size), intent(in) :: graph
      end subroutine message_update
 
[…]
