### 2024-04-23 28fe61b
Add variables to derived type
