### 2024-04-23 85b2f21
Add generalised handling of partial derivatives
