+++ src/lib/mod_conv_mp_layer.f90
@@ -1,0 +1,253 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+module conv_mp_methods
+  use constants, only: real12
+  use custom_types, only: graph_type, activation_type
+  use activation_softmax, only: softmax_setup
+  use mpnn_module, only: &
+       mpnn_type, state_method_type, message_method_type, readout_method_type, &
+       feature_type
+  implicit none
+  
+
+  private
+
+
+  type, extends(message_method_type) :: convolutional_message_method_type
+   contains
+     procedure :: update => convolutional_message_update
+     procedure :: get_differential => convolutional_get_message_differential
+     procedure :: calculate_partials => convolutional_calculate_message_partials
+  end type convolutional_message_method_type
+
+  type, extends(state_method_type) :: convolutional_state_method_type
+     !! weight has dimensions (feature_out, feature_in, vertex_degree, batch_size)
+     real(real12), dimension(:,:,:), allocatable :: weight
+     real(real12), dimension(:,:,:,:), allocatable :: dw
+     class(activation_type), allocatable :: transfer
+   contains
+     procedure :: update => convolutional_state_update
+     procedure :: get_differential => convolutional_get_state_differential
+     procedure :: calculate_partials => convolutional_calculate_state_partials
+  end type convolutional_state_method_type
+
+  type, extends(readout_method_type) :: convolutional_readout_method_type
+    real(real12), dimension(:,:,:), allocatable :: weight
+    real(real12), dimension(:,:,:,:), allocatable :: dw
+    class(activation_type), allocatable :: transfer
+   contains
+     procedure :: get_output => convolutional_get_readout_output
+     procedure :: get_differential => convolutional_get_readout_differential
+     procedure :: calculate_partials => convolutional_calculate_readout_partials
+  end type convolutional_readout_method_type
+
+
+contains
+
+  subroutine convolutional_message_udpate(this, hidden, graph)
+    implicit none
+    class(convolutional_message_method_type), intent(inout) :: this
+    type(feature_type), dimension(this%batch_size), intent(in) :: hidden
+    type(graph_type), dimension(this%batch_size), intent(in) :: graph
+
+    integer :: s, v, w
+
+    do s = 1, this%batch_size
+       do v = 1, graph(s)%num_vertices
+         this%feature(s)%val(:,v) = 0._real12
+         do w = 1, graph%num_vertices
+             if(graph%adjacency(v,w) == 1) then
+               this%feature(s)%val(:,v) = &
+                     this%feature(:,v) + &
+                     [ hidden(s)%val(:,w), graph(s)%edge(v,w)%feature(:) ]
+             end if
+         end do
+       end do
+    end do
+
+  end subroutine convolutional_message_udpate
+
+  function convolutional_get_message_differential(this, hidden, graph), &
+       result(output)
+    implicit none
+    class(convolutional_message_method_type), intent(in) :: this
+    type(feature_type), dimension(this%batch_size), intent(in) :: hidden
+    type(graph_type), dimension(this%batch_size), intent(in) :: graph
+    
+    type(feature_type), dimension(this%batch_size) :: output
+
+    integer :: s, v
+
+    do s = 1, this%batch_size
+       do v = 1, graph(s)%num_vertices
+          output(s)%val(:,v) = 1._real12
+       end do
+    end do
+
+  end function convolutional_get_message_differential
+
+  subroutine convolutional_calculate_message_partials(this, output_state, graph, input)
+    implicit none
[…]
