### 2024-04-23 6aa8fcb
Add message passing module to compilation
