### 2024-04-24 f02fa60
Change degree handling
