### 2024-04-24 1bcfcc4
Add outer product procedure
