### 2024-04-24 915578f
Fix readout interfaces
