+++ src/lib/mod_conv_mp_layer.f90
@@ -4,6 +4,7 @@
 !!!#############################################################################
 module conv_mp_methods
   use constants, only: real12
+  use misc, only: outer_product
   use custom_types, only: graph_type, activation_type
   use activation_softmax, only: softmax_setup
   use mpnn_module, only: &
@@ -26,6 +27,7 @@ module conv_mp_methods
      !! weight has dimensions (feature_out, feature_in, vertex_degree, batch_size)
      real(real12), dimension(:,:,:), allocatable :: weight
      real(real12), dimension(:,:,:,:), allocatable :: dw
+     real(real12), dimension(:,:), allocatable :: z
      class(activation_type), allocatable :: transfer
    contains
      procedure :: update => convolutional_state_update
@@ -34,8 +36,10 @@ module conv_mp_methods
   end type convolutional_state_method_type
 
   type, extends(readout_method_type) :: convolutional_readout_method_type
+    integer :: num_time_steps
     real(real12), dimension(:,:,:), allocatable :: weight
     real(real12), dimension(:,:,:,:), allocatable :: dw
+    real(real12), dimension(:,:), allocatable :: z
     class(activation_type), allocatable :: transfer
    contains
      procedure :: get_output => convolutional_get_readout_output
@@ -46,10 +50,10 @@ module conv_mp_methods
 
 contains
 
-  subroutine convolutional_message_udpate(this, hidden, graph)
+  subroutine convolutional_message_update(this, input, graph)
     implicit none
     class(convolutional_message_method_type), intent(inout) :: this
-    type(feature_type), dimension(this%batch_size), intent(in) :: hidden
+    type(feature_type), dimension(this%batch_size), intent(in) :: input
     type(graph_type), dimension(this%batch_size), intent(in) :: graph
 
     integer :: s, v, w
@@ -57,23 +61,23 @@ contains
     do s = 1, this%batch_size
        do v = 1, graph(s)%num_vertices
          this%feature(s)%val(:,v) = 0._real12
-         do w = 1, graph%num_vertices
-             if(graph%adjacency(v,w) == 1) then
+         do w = 1, graph(s)%num_vertices
+             if(graph(s)%adjacency(v,w) .eq. 1) then
                this%feature(s)%val(:,v) = &
-                     this%feature(:,v) + &
-                     [ hidden(s)%val(:,w), graph(s)%edge(v,w)%feature(:) ]
+                     this%feature(s)%val(:,v) + &
+                     [ input(s)%val(:,w), graph(s)%edge(v,w)%feature(:) ]
              end if
          end do
        end do
     end do
 
-  end subroutine convolutional_message_udpate
+  end subroutine convolutional_message_update
 
-  function convolutional_get_message_differential(this, hidden, graph), &
+  pure function convolutional_get_message_differential(this, input, graph) &
        result(output)
     implicit none
     class(convolutional_message_method_type), intent(in) :: this
-    type(feature_type), dimension(this%batch_size), intent(in) :: hidden
+    type(feature_type), dimension(this%batch_size), intent(in) :: input
     type(graph_type), dimension(this%batch_size), intent(in) :: graph
     
     type(feature_type), dimension(this%batch_size) :: output
@@ -88,13 +92,15 @@ contains
 
   end function convolutional_get_message_differential
 
-  subroutine convolutional_calculate_message_partials(this, output_state, graph, input)
+  subroutine convolutional_calculate_message_partials(this, input, gradient, graph)
     implicit none
-    class(message_method_type), intent(inout) :: this
-    class(state_method_type), intent(in) :: output_state
-    type(graph_type), dimension(this%batch_size), intent(in) :: graph
+    class(convolutional_message_method_type), intent(inout) :: this
[…]
