### 2024-04-24 8e8b857
Implement readout partials
