+++ src/lib/mod_message_passing_layer.f90
@@ -1,351 +1,0 @@
-!!!#############################################################################
-!!! Code written by Ned Thaddeus Taylor
-!!! Code part of the ATHENA library - a feedforward neural network library
-!!!#############################################################################
-!!! module contains implementation of a message passing neural network
-!!!#############################################################################
-module mpnn_module
-  use constants, only: real12
-  use custom_types, only: graph_type
-  implicit none
-  
-
-  private
-
-  public :: mpnn_type, feature_type
-  public :: state_method_type, message_method_type, readout_method_type
-  public :: state_update, get_state_differential
-  public :: message_update, get_message_differential
-  public :: get_readout_output, get_readout_differential
-
-
-  type :: mpnn_type
-     integer :: num_features
-     integer :: num_vertices
-     integer :: num_time_steps
-     integer :: batch_size
-     !! state and message dimension is (time_step)
-     class(message_method_type), dimension(:), allocatable :: message
-     class(state_method_type), dimension(:), allocatable :: state
-     class(readout_method_type), allocatable :: readout
-     real(real12), dimension(:,:), allocatable :: output
-     real(real12), dimension(:,:,:), allocatable :: di
-   contains
-     procedure, pass(this) :: forward
-     procedure, pass(this) :: backward
-  end type mpnn_type
-
-
-  type :: feature_type
-     real(real12), dimension(:,:), allocatable :: val
-   contains
-     ! t = type, r = real, i = int
-     procedure :: add_t_t => feature_add
-     procedure :: multiply_t_t => feature_multiply
-     generic :: operator(+) => add_t_t
-     generic :: operator(*) => multiply_t_t
-  end type feature_type
-
-
-   type, abstract :: message_method_type
-     integer :: num_features
-     integer :: batch_size
-     !! feature has dimensions (feature, vertex)
-     type(feature_type), dimension(:), allocatable :: feature
-     type(feature_type), dimension(:), allocatable :: di
-   contains
-     procedure(message_update), deferred, pass(this) :: update
-     procedure(get_message_differential), deferred, pass(this) :: get_differential
-     procedure(calculate_message_partials), deferred, pass(this) :: calculate_partials
-  end type message_method_type
-
-  type, abstract :: state_method_type
-     integer :: num_features
-     integer :: batch_size
-     !! feature has dimensions (feature, vertex)
-     type(feature_type), dimension(:), allocatable :: feature
-     type(feature_type), dimension(:), allocatable :: di
-   contains
-     procedure(state_update), deferred, pass(this) :: update
-     procedure(get_state_differential), deferred, pass(this) :: get_differential
-     procedure(calculate_state_partials), deferred, pass(this) :: calculate_partials
-  end type state_method_type
-
-  type, abstract :: readout_method_type
-     integer :: batch_size
-     integer :: num_outputs
-     type(feature_type), dimension(:), allocatable :: di
-   contains
-     procedure(get_readout_output), deferred, pass(this) :: get_output
-     procedure(get_readout_differential), deferred, pass(this) :: get_differential
-     procedure(calculate_readout_partials), deferred, pass(this) :: calculate_partials
-  end type readout_method_type
-
-
-  abstract interface
-     subroutine message_update(this, input, graph)
-       import :: message_method_type, feature_type, graph_type
-       class(message_method_type), intent(inout) :: this
-       !! input features has dimensions (feature, vertex, batch_size)
-       type(feature_type), dimension(this%batch_size), intent(in) :: input
-       type(graph_type), dimension(this%batch_size), intent(in) :: graph
-     end subroutine message_update
-
-     pure function get_message_differential(this, input, graph) result(output)
[…]
