### 2024-04-24 9ab439a
Add initialisation handling
