+++ src/lib/mod_mpnn_layer.f90
@@ -7,34 +7,45 @@
 module mpnn_layer
   use constants, only: real12
   use custom_types, only: graph_type
+  use base_layer, only: base_layer_type
   implicit none
   
 
   private
 
-  public :: mpnn_layer_type, feature_type
+  public :: mpnn_layer_type, feature_type, mpnn_method_type
   public :: state_method_type, message_method_type, readout_method_type
   public :: state_update, get_state_differential
   public :: message_update, get_message_differential
   public :: get_readout_output, get_readout_differential
 
 
-  type :: mpnn_layer_type
-     integer :: num_features
-     integer :: num_time_steps
-     integer :: num_outputs
-     integer :: batch_size
-     !! state and message dimension is (time_step)
-     class(message_method_type), dimension(:), allocatable :: message
-     class(state_method_type), dimension(:), allocatable :: state
-     class(readout_method_type), allocatable :: readout
+  type, extends(base_layer_type) :: mpnn_layer_type
+     type(graph_type), dimension(:), allocatable :: graph
+     class(mpnn_method_type), allocatable :: method
      real(real12), dimension(:,:), allocatable :: output
      !real(real12), dimension(:,:,:), allocatable :: di
    contains
-     procedure, pass(this) :: forward
-     procedure, pass(this) :: backward
+     procedure, pass(this) :: init => init_mpnn
+     procedure, pass(this) :: get_output => get_output_mpnn
+     procedure, pass(this) :: set_batch_size => set_batch_size_base_mpnn
+     procedure, pass(this) :: set_graph
+     procedure, pass(this) :: forward => forward_rank
+     procedure, pass(this) :: backward => backward_rank
   end type mpnn_layer_type
 
+  type, abstract :: mpnn_method_type
+    integer :: num_outputs
+    integer :: num_time_steps
+    !! each dimension of num_features is for vertex and edge
+    integer, dimension(2) :: num_features
+    !! state and message dimension is (time_step)
+    class(message_method_type), dimension(:), allocatable :: message
+    class(state_method_type), dimension(:), allocatable :: state
+    class(readout_method_type), allocatable :: readout
+   contains
+    procedure(init_method), deferred, pass(this) :: init
+  end type mpnn_method_type
 
   type :: feature_type
      real(real12), dimension(:,:), allocatable :: val
@@ -49,6 +60,7 @@ module mpnn_layer
 
    type, abstract :: message_method_type
      integer :: num_features
+     integer :: num_outputs
      integer :: batch_size
      !! feature has dimensions (feature, vertex)
      type(feature_type), dimension(:), allocatable :: feature
@@ -61,6 +73,7 @@ module mpnn_layer
 
   type, abstract :: state_method_type
      integer :: num_features
+     integer :: num_outputs
      integer :: batch_size
      !! feature has dimensions (feature, vertex)
      type(feature_type), dimension(:), allocatable :: feature
@@ -72,8 +85,9 @@ module mpnn_layer
   end type state_method_type
 
   type, abstract :: readout_method_type
-     integer :: batch_size
+     integer :: num_features
      integer :: num_outputs
+     integer :: batch_size
      type(feature_type), dimension(:), allocatable :: di
    contains
      procedure(get_readout_output), deferred, pass(this) :: get_output
@@ -83,7 +97,7 @@ module mpnn_layer
 
 
   abstract interface
-     module subroutine message_update(this, input, graph)
+     pure module subroutine message_update(this, input, graph)
        class(message_method_type), intent(inout) :: this
        !! input features has dimensions (feature, vertex, batch_size)
        type(feature_type), dimension(this%batch_size), intent(in) :: input
@@ -98,7 +112,7 @@ module mpnn_layer
        type(feature_type), dimension(this%batch_size) :: output
[…]

[…]
