### 2024-04-24 f0f0ffb
Remove superfluous allocate
