+++ src/athena.f90
@@ -11,6 +11,7 @@ module athena
   use random, only: random_setup
   use network, only: network_type
   use metrics, only: metric_dict_type
+  use custom_types, only: graph_type
 
   !! accuracy methods
   use accuracy, only: categorical_score, mae_score, mse_score, r2_score
