### 2024-05-02 09ed8ef
Fix MPNN forward
