### 2024-05-02 25b579a
Convert readout output procedure to subroutine
