+++ src/lib/mod_conv_mpnn_layer.f90
@@ -139,6 +139,7 @@ contains
     allocate(state_method%weight(num_vertex_features, num_vertex_features + num_edge_features, max_vertex_degree))
     allocate(state_method%dw(num_vertex_features, num_vertex_features + num_edge_features, max_vertex_degree, batch_size))
     allocate(state_method%z(batch_size))
+    allocate(state_method%di(batch_size))
   
     write(*,*) "setting up transfer function"
     allocate(state_method%transfer, &
@@ -157,9 +158,10 @@ contains
     readout_method%num_outputs = num_outputs
     readout_method%batch_size = batch_size
 
-    allocate(readout_method%weight(num_outputs, num_inputs, num_time_steps))
-    allocate(readout_method%dw(num_outputs, num_time_steps, num_outputs, batch_size))
-    allocate(readout_method%z(num_time_steps, batch_size))
+    allocate(readout_method%weight(num_inputs, num_outputs, num_time_steps+1))
+    allocate(readout_method%dw(num_inputs, num_outputs, num_time_steps+1, batch_size))
+    allocate(readout_method%z(num_time_steps+1, batch_size))
+    allocate(readout_method%di(batch_size))
 
     write(*,*) "setting up transfer function"
     allocate(readout_method%transfer, &
@@ -243,6 +245,7 @@ contains
     type(conv_mpnn_layer_type) :: layer
 
     layer%batch_size = batch_size
+    layer%output_shape = [num_outputs]
 
     layer%method = conv_mpnn_method_type( &
          [num_vertex_features, num_edge_features, num_time_steps], [num_outputs], batch_size &
@@ -313,6 +316,8 @@ contains
     !! delta(l) = differential of activation * error from next layer
 
     do concurrent(s=1:this%batch_size)
+       if(allocated(this%di(s)%val)) deallocate(this%di(s)%val)
+       allocate(this%di(s)%val(size(gradient(s)%val, 1), graph(s)%num_vertex_features))
        !! no message passing transfer function
        this%di(s)%val(:,:) = gradient(s)%val(:,:graph(s)%num_vertex_features)
     end do
@@ -329,13 +334,14 @@ contains
     integer :: s, v
 
     do s = 1, this%batch_size
+       if(allocated(this%z(s)%val)) deallocate(this%z(s)%val)
+       allocate(this%z(s)%val(size(input(s)%val, 1), graph(s)%num_vertices))
        do v = 1, graph(s)%num_vertices
-          this%feature(s)%val(:,v) = this%transfer%activate( &
-                matmul( &
-                     this%weight(:,:,graph(s)%vertex(v)%degree), &
-                     input(s)%val(:,v) &
-                ) &
+          this%z(s)%val(:,v) = matmul( &
+               this%weight(:,:,graph(s)%vertex(v)%degree), &
+               input(s)%val(:,v) &
           )
+          this%feature(s)%val(:,v) = this%transfer%activate( this%z(s)%val(:,v))
        end do
     end do
 
@@ -353,11 +359,7 @@ contains
 
     do s = 1, this%batch_size
        do v = 1, graph(s)%num_vertices
-          output(s)%val(:,v) = this%transfer%differentiate( &
-               matmul( &
-                    this%weight(:,:,graph(s)%vertex(v)%degree), &
-                    input(s)%val(:,v) &
-               ) )
+          output(s)%val(:,v) = this%transfer%differentiate(this%z(s)%val(:,v))
        end do
     end do
 
@@ -382,8 +384,11 @@ contains
 
     do concurrent(s=1:this%batch_size)
        !! no message passing transfer function
-       delta(:,:) = gradient(s)%val(:,:) * &
+       delta = gradient(s)%val(:,:) * &
             this%transfer%differentiate(this%z(s)%val(:,:))
[…]
