### 2024-05-03 34dfbd7
Add parameter and gradient handling procedures
