### 2024-05-03 7c5ffc8
Convert mpnn layer from extend base to learnable
