### 2024-05-03 cc8c59e
Add gradient handling procedures
