+++ src/lib/mod_conv_mpnn_layer.f90
@@ -7,6 +7,7 @@ module conv_mpnn_layer
   use misc, only: outer_product
   use custom_types, only: graph_type, activation_type
   use activation, only: activation_setup
+  use clipper, only: clip_type
   use mpnn_layer, only: &
        mpnn_layer_type, method_container_type, &
        state_method_type, message_method_type, readout_method_type, &
@@ -21,9 +22,9 @@ module conv_mpnn_layer
 
   type, extends(message_method_type) :: convolutional_message_method_type
    contains
-     procedure :: update => convolutional_message_update
-     procedure :: get_differential => convolutional_get_message_differential
-     procedure :: calculate_partials => convolutional_calculate_message_partials
+     procedure :: update => update_message_conv
+     procedure :: get_differential => get_differential_message_conv
+     procedure :: calculate_partials => calculate_partials_message_conv
   end type convolutional_message_method_type
   interface convolutional_message_method_type
     module function message_method_setup( &
@@ -41,9 +42,14 @@ module conv_mpnn_layer
      type(feature_type), dimension(:), allocatable :: z
      class(activation_type), allocatable :: transfer
    contains
-     procedure :: update => convolutional_state_update
-     procedure :: get_differential => convolutional_get_state_differential
-     procedure :: calculate_partials => convolutional_calculate_state_partials
+     procedure :: get_num_params => get_num_params_state_conv
+     procedure :: get_params => get_params_state_conv
+     procedure :: set_params => set_params_state_conv
+     procedure :: get_gradients => get_gradients_state_conv
+     procedure :: set_gradients => set_gradients_state_conv
+     procedure :: update => update_state_conv
+     procedure :: get_differential => get_state_differential_conv
+     procedure :: calculate_partials => calculate_state_partials_conv
   end type convolutional_state_method_type
   interface convolutional_state_method_type
     module function state_method_setup( &
@@ -61,9 +67,14 @@ module conv_mpnn_layer
     type(feature_type), dimension(:,:), allocatable :: z
     class(activation_type), allocatable :: transfer
    contains
-     procedure :: get_output => convolutional_get_readout_output
-     procedure :: get_differential => convolutional_get_readout_differential
-     procedure :: calculate_partials => convolutional_calculate_readout_partials
+     procedure :: get_num_params => get_num_params_readout_conv
+     procedure :: get_params => get_params_readout_conv
+     procedure :: set_params => set_params_readout_conv
+     procedure :: get_gradients => get_gradients_readout_conv
+     procedure :: set_gradients => set_gradients_readout_conv
+     procedure :: get_output => get_output_readout_conv
+     procedure :: get_differential => get_readout_differential_conv
+     procedure :: calculate_partials => calculate_readout_partials_conv
   end type convolutional_readout_method_type
   interface convolutional_readout_method_type
     module function readout_method_setup( &
@@ -111,6 +122,137 @@ module conv_mpnn_layer
 
 contains
 
+!!!#############################################################################
+!!! 
+!!!#############################################################################
+  pure function get_num_params_state_conv(this) result(num_params)
+    implicit none
+    class(convolutional_state_method_type), intent(in) :: this
+    integer :: num_params
+
+    num_params = size(this%weight)
+  end function get_num_params_state_conv
+  
+  pure function get_num_params_readout_conv(this) result(num_params)
+    implicit none
+    class(convolutional_readout_method_type), intent(in) :: this
+    integer :: num_params
+
+    num_params = size(this%weight)
+  end function get_num_params_readout_conv
+!!!#############################################################################
+
+
[…]
