### 2024-05-20 nedtaylor/athena#23 @nedtaylor
I have opened up an issue relating to required references neural-fortran within the code, #24. I think I have captured all parts, @milancurcic.

Further discussions relating to that topic can be handled on that thread.
