### 2024-05-21 7612617
Add set_shape procedure and fix sizes
