+++ src/lib/mod_conv_mpnn_layer.f90
@@ -17,25 +17,27 @@ module conv_mpnn_layer
 
   private
   public :: conv_mpnn_layer_type, conv_method_container_type
+  public :: conv_readout_method_type
 
 
 
-  type, extends(message_method_type) :: convolutional_message_method_type
+  type, extends(message_method_type) :: conv_message_method_type
    contains
      procedure :: update => update_message_conv
      procedure :: get_differential => get_differential_message_conv
      procedure :: calculate_partials => calculate_partials_message_conv
-  end type convolutional_message_method_type
-  interface convolutional_message_method_type
+     procedure :: set_shape => set_shape_message_conv
+  end type conv_message_method_type
+  interface conv_message_method_type
     module function message_method_setup( &
          num_vertex_features, num_edge_features, batch_size ) result(message_method)
       integer, intent(in) :: num_vertex_features, num_edge_features, batch_size
-      type(convolutional_message_method_type) :: message_method
+      type(conv_message_method_type) :: message_method
     end function message_method_setup
-  end interface convolutional_message_method_type
+  end interface conv_message_method_type
 
 
-  type, extends(state_method_type) :: convolutional_state_method_type
+  type, extends(state_method_type) :: conv_state_method_type
      !! weight has dimensions (feature_out, feature_in, vertex_degree, batch_size)
      real(real12), dimension(:,:,:), allocatable :: weight
      real(real12), dimension(:,:,:,:), allocatable :: dw
@@ -47,20 +49,21 @@ module conv_mpnn_layer
      procedure :: set_params => set_params_state_conv
      procedure :: get_gradients => get_gradients_state_conv
      procedure :: set_gradients => set_gradients_state_conv
+     procedure :: set_shape => set_shape_state_conv
      procedure :: update => update_state_conv
      procedure :: get_differential => get_state_differential_conv
      procedure :: calculate_partials => calculate_state_partials_conv
-  end type convolutional_state_method_type
-  interface convolutional_state_method_type
+  end type conv_state_method_type
+  interface conv_state_method_type
     module function state_method_setup( &
          num_vertex_features, num_edge_features, max_vertex_degree, batch_size ) result(state_method)
       integer, intent(in) :: num_vertex_features, num_edge_features, max_vertex_degree, batch_size
-      type(convolutional_state_method_type) :: state_method
+      type(conv_state_method_type) :: state_method
     end function state_method_setup
-  end interface convolutional_state_method_type
+  end interface conv_state_method_type
 
 
-  type, extends(readout_method_type) :: convolutional_readout_method_type
+  type, extends(readout_method_type) :: conv_readout_method_type
     integer :: num_time_steps
     real(real12), dimension(:,:,:), allocatable :: weight
     real(real12), dimension(:,:,:,:), allocatable :: dw
@@ -72,17 +75,18 @@ module conv_mpnn_layer
      procedure :: set_params => set_params_readout_conv
      procedure :: get_gradients => get_gradients_readout_conv
      procedure :: set_gradients => set_gradients_readout_conv
+     procedure :: set_shape => set_shape_readout_conv
      procedure :: get_output => get_output_readout_conv
      procedure :: get_differential => get_readout_differential_conv
      procedure :: calculate_partials => calculate_readout_partials_conv
-  end type convolutional_readout_method_type
-  interface convolutional_readout_method_type
+  end type conv_readout_method_type
+  interface conv_readout_method_type
     module function readout_method_setup( &
[…]
