### 2024-05-21 5f9d932
Remove get_differential procedure
