+++ src/lib/mod_mpnn_layer.f90
@@ -16,9 +16,9 @@ module mpnn_layer
 
   public :: mpnn_layer_type, feature_type, method_container_type
   public :: state_method_type, message_method_type, readout_method_type
-  public :: update_state, get_differential_state
-  public :: update_message, get_differential_message
-  public :: get_output_readout, get_differential_readout
+  public :: update_state
+  public :: update_message
+  public :: get_output_readout
 
 
   type, extends(learnable_layer_type) :: mpnn_layer_type
@@ -89,21 +89,18 @@ module mpnn_layer
   type, extends(base_method_type), abstract :: message_method_type
    contains
      procedure(update_message), deferred, pass(this) :: update
-     procedure(get_differential_message), deferred, pass(this) :: get_differential
      procedure(calculate_partials_message), deferred, pass(this) :: calculate_partials
   end type message_method_type
 
   type, extends(base_method_type), abstract :: state_method_type
    contains
      procedure(update_state), deferred, pass(this) :: update
-     procedure(get_differential_state), deferred, pass(this) :: get_differential
      procedure(calculate_partials_state), deferred, pass(this) :: calculate_partials
   end type state_method_type
 
   type, extends(base_method_type), abstract :: readout_method_type
    contains
      procedure(get_output_readout), deferred, pass(this) :: get_output
-     procedure(get_differential_readout), deferred, pass(this) :: get_differential
      procedure(calculate_partials_readout), deferred, pass(this) :: calculate_partials
   end type readout_method_type
 
@@ -115,14 +112,6 @@ module mpnn_layer
        type(feature_type), dimension(this%batch_size), intent(in) :: input
        type(graph_type), dimension(this%batch_size), intent(in) :: graph
      end subroutine update_message
-
-     pure module function get_differential_message(this, input, graph) result(output)
-       class(message_method_type), intent(in) :: this
-       !! input features has dimensions (feature, vertex, batch_size)
-       type(feature_type), dimension(this%batch_size), intent(in) :: input
-       type(graph_type), dimension(this%batch_size), intent(in) :: graph
-       type(feature_type), dimension(this%batch_size) :: output
-     end function get_differential_message
 
      pure module subroutine calculate_partials_message(this, input, gradient, graph)
        class(message_method_type), intent(inout) :: this
@@ -140,14 +129,6 @@ module mpnn_layer
        type(graph_type), dimension(this%batch_size), intent(in) :: graph
      end subroutine update_state
 
-     pure module function get_differential_state(this, input, graph) result(output)
-       class(state_method_type), intent(in) :: this
-       !! input has dimensions (feature, vertex, batch_size)
-       type(feature_type), dimension(this%batch_size), intent(in) :: input
-       type(graph_type), dimension(this%batch_size), intent(in) :: graph
-       type(feature_type), dimension(this%batch_size) :: output
-     end function get_differential_state
-
      pure module subroutine calculate_partials_state(this, input, gradient, graph)
        class(state_method_type), intent(inout) :: this
        !! hidden features has dimensions (feature, vertex, batch_size)
@@ -162,12 +143,6 @@ module mpnn_layer
        class(state_method_type), dimension(:), intent(in) :: input
        real(real12), dimension(this%num_outputs, this%batch_size), intent(out) :: output
      end subroutine get_output_readout
-
-     pure module function get_differential_readout(this, input) result(output)
-       class(readout_method_type), intent(in) :: this
-       class(state_method_type), dimension(:), intent(in) :: input
-       type(feature_type), dimension(this%batch_size) :: output
-     end function get_differential_readout
 
      pure module subroutine calculate_partials_readout(this, input, gradient)
        class(readout_method_type), intent(inout) :: this
