### 2024-05-21 bb8060c
Fix degree and update message
