### 2024-05-21 3881145
Remove redundant reallocation
