### 2024-05-24 #6492 @awvwgk
I finished my review, the paper and the repo look all good for JOSS.
