### 2024-05-24 #6492 @nedtaylor
> @nedtaylor could you take a look at the issues raised by the reviewers and provide an ETA?

Hi @HaoZeke. From what I can see, the three issues opened are:
- https://github.com/nedtaylor/athena/issues/20
- https://github.com/nedtaylor/athena/issues/21
- https://github.com/nedtaylor/athena/issues/23

From what I can see, these have been resolved. All issues have been kept open as requested for the review process until approved for the new paper branch version.
- Issue 20 is for bookkeeping (i.e. gfortran 11 does not work with the library, which is already documented in the README)
- Issue 21 has been fixed
- I believe I have addressed issue 23 and am awaiting a response from @milancurcic regarding a related issue (issue 24).

If I haven't addressed these issues properly and there is more to do on my end, then please let me know. I'm eager to do whatever I can to continue the process.

If there are other issues I have missed, please point me to them.
