### 2024-05-25 bf347c2
21 ifx compilation error (#25)
* Fix abstract interface

* Fix optimiser module variable name conflict

* Fix activation module variable name conflict

* Fix initialiser module variable name conflict

* Remove coverage compiler flag option for intel
