+++ src/lib/mod_base_layer.f90
@@ -116,7 +116,7 @@ module base_layer
   end interface
 
 
-  abstract interface
+  interface
      !!--------------------------------------------------------------------------
      !! initialise layer
      !!--------------------------------------------------------------------------
@@ -144,7 +144,7 @@ module base_layer
      end subroutine set_batch_size
   end interface
 
-  abstract interface
+  interface
      !!--------------------------------------------------------------------------
      !! get number of parameters in layer
      !! procedure modified from neural-fortran library
