### 2024-05-25 nedtaylor/athena#20 @nedtaylor
> @nedtaylor Can you put an update to the README.md that the library works with v12.3?

@milancurcic Reference added in commit 5882155ff5e11d12cf041e5d097823e37de6323f.
