### 2024-05-25 nedtaylor/athena#20 @nedtaylor
Leaving this issue open for now as I will look into why it fails for v11.4 and whether this can be fixed at some point.
