### 2024-05-25 nedtaylor/athena#23 @nedtaylor
Merged into main. Closing.

Please reopen if more discussions are required regarding this.
