### 2024-05-25 45203d1
Convert mpnn_layer to 1D input data
