### 2024-05-27 487920f
Add mpnn_layer_type to network and container
