### 2024-05-28 74e988b
Add mpnn example
