+++ example/mpnn_chemical/src/mod_read_chemical_graphs.f90
@@ -1,0 +1,619 @@
+module read_chemical_graphs
+  use constants_mnist, only: real12
+  use misc_linalg, only: modu
+  use rw_geom, only: bas_type, geom_read, igeom_input
+  use athena, only: graph_type, edge_type
+  implicit none
+
+  private
+
+  public :: read_mp_db, read_extxyz_db
+
+contains
+
+!!!#############################################################################
+!!! get line of unknown length
+!!!#############################################################################
+!!! read line of unknown length from file
+!!! implemented as written by IanH in stackoverflow:
+!!! https://stackoverflow.com/questions/34746461/how-to-read-text-files-with-possibly-very-long-lines-in-fortran
+subroutine get_line(lun, line, iostat, iomsg)
+   integer, intent(in)           :: lun
+   character(len=:), intent(out), allocatable :: line
+   integer, intent(out)          :: iostat
+   character(*), intent(inout)   :: iomsg
+ 
+   integer, parameter            :: buffer_len = 1024
+   character(len=buffer_len)     :: buffer
+   integer                       :: size_read
+ 
+   line = ''
+   do
+     read ( lun, '(A)',  &
+         iostat = iostat,  &
+         iomsg = iomsg,  &
+         advance = 'no',  &
+         size = size_read ) buffer
+     if (is_iostat_eor(iostat)) then
+       line = line // buffer(:size_read)
+       iostat = 0
+       exit
+     else if (iostat == 0) then
+       line = line // buffer
+     else
+       exit
+     end if
+     write(*,*) line
+   end do
+ end subroutine get_line
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! read mnist dataset
+!!!#############################################################################
+  subroutine read_mp_db(file,graphs,labels)
+    use misc_mnist, only: icount
+    implicit none
+    integer ::v, e, pos_i, pos_f, length, Reason, unit
+    character(:), allocatable :: line
+    real(real12) :: label
+    type(graph_type) :: graph
+    character(100) :: iomsg
+
+    integer :: num_samples
+    character(1024), intent(in) :: file
+    type(graph_type), allocatable, dimension(:), intent(out) :: graphs
+    real(real12), allocatable, dimension(:), intent(out) :: labels
+
+
+    open(newunit=unit, file=file, status='old', action='read', iostat=Reason)
+
+    write(*,*) "Reading data from file: ", trim(file)
+    call get_line(unit, line, Reason, iomsg)
+    write(*,*) "Line read"
+
+    num_samples = 0
+    pos_i = 1
+    graph%num_vertex_features = 1
+    graph%num_edge_features = 1
+    length = len(line)
+    do
+       if (index(line(pos_i:), 'graph') .eq. 0) exit
+       if(num_samples.gt.10) exit
+       write(*,*) "pos_i: ", pos_i
+
+       !! get graph name
+       pos_i = pos_i + index(line(pos_i:), '"material_id": "') - 1 + 16
+       pos_f = pos_i + index(line(pos_i:), '"') - 2
+       graph%name = line(pos_i:pos_i+pos_f-1)
+       pos_i = pos_f
+       
+       pos_i = pos_i + index(line(pos_i:), 'graph') - 1
+
+       pos_i = pos_i + index(line(pos_i:), '"atom": [') - 1 + 9
+       pos_f = pos_i + index(line(pos_i:), ']') - 2
+       graph%num_vertices = icount(line(pos_i:pos_f), ',')
+       allocate(graph%vertex(graph%num_vertices))
+       do v = 1, graph%num_vertices
+          allocate(graph%vertex(v)%feature(graph%num_vertex_features))
+       end do
[…]
