+++ example/mpnn_chemical/src/mod_rw_geom.f90
@@ -1,0 +1,1112 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor and Francis Huw Davies
+!!! Code part of the ARTEMIS group (Hepplestone research group).
+!!! Think Hepplestone, think HRG.
+!!!#############################################################################
+!!! Module made to read and write structure input files
+!!! Currently supports:
+!!!    -VASP
+!!!    -Quantum Espresso
+!!!    -CASTEP
+!!!    -xyz (read only)
+!!!#############################################################################
+module rw_geom
+  use constants_mnist, only: pi,real12
+  use misc_mnist, only: to_upper,jump,Icount
+  use misc_linalg, only: LUinv,modu
+  implicit none
+
+  private
+
+  integer :: igeom_input=1,igeom_output=1
+  real(real12), dimension(3,3) :: lattice
+
+  type spec_type
+     real(real12), allocatable ,dimension(:,:) :: atom
+     real(real12) :: mass
+     real(real12) :: charge
+     character(len=3) :: name
+     integer :: num
+  end type spec_type
+  type bas_type
+     type(spec_type), allocatable, dimension(:) :: spec
+     integer :: nspec
+     integer :: natom
+     real(real12) :: energy
+     logical :: lcart=.false.
+     character(len=1024) :: sysname
+  end type bas_type
+  type(bas_type) :: basis
+
+
+  public :: igeom_input,igeom_output
+  public :: bas_type
+  public :: clone_bas
+  public :: convert_bas
+  public :: geom_read,geom_write
+  
+
+!!!updated  2020/02/06
+
+
+contains
+!!!#############################################################################
+!!! sets up the name of output files and subroutines to read files
+!!!#############################################################################
+  subroutine geom_read(UNIT,lat,bas,length)
+    implicit none
+    integer :: UNIT,dim,i
+    type(bas_type) :: bas
+    real(real12), dimension(3,3) :: lat
+    integer, optional, intent(in) :: length
+
+    lattice=0._real12
+    dim=3
+    if(present(length)) dim=length
+
+    select case(igeom_input)
+    case(1)
+       call VASP_geom_read(UNIT,dim)
+    case(2)
+       call CASTEP_geom_read(UNIT,dim)
+    case(3)
+       call QE_geom_read(UNIT,dim)
+    case(4)
+       !call err_abort('ERROR: ARTEMIS not yet set up for CRYSTAL')
+       write(0,'("ERROR: ARTEMIS not yet set up for CRYSTAL")')
+       stop
+    case(5)
+       call XYZ_geom_read(UNIT,dim)
+       write(0,'("WARNING: XYZ file format does not contain lattice data")')
+    case(6)
+       call extXYZ_geom_read(UNIT,dim)
+    end select
+    call clone_bas(basis,bas,lattice,lat)
+    deallocate(basis%spec)
+    if(dim.eq.4)then
+       do i=1,bas%nspec
+          bas%spec(i)%atom(:,4)=1._real12
+       end do
+    end if
+
+
+  end subroutine geom_read
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! sets up the name of output files and subroutines to read files
+!!!#############################################################################
+  subroutine geom_write(UNIT,lat,bas)
+    implicit none
+    integer :: UNIT
+    type(bas_type) :: bas
+    real(real12), dimension(3,3) :: lat
+
+!!! MAKE IT CHANGE HERE IF USER SPECIFIES LCART OR NOT
+!!! AND GIVE IT THE CASTEP AND QE OPTION OF LABC !!!
+
+    select case(igeom_output)
+    case(1)
[…]

[…]
