### 2024-05-28 7a66db5
Add lr_decay initialisation
