### 2024-05-28 4144b65
Move graph_type to separate module
