+++ src/lib/mod_conv_mpnn_layer.f90
@@ -5,63 +5,52 @@
 module conv_mpnn_layer
   use constants, only: real12
   use misc, only: outer_product
-  use custom_types, only: graph_type, activation_type
+  use custom_types, only: activation_type
+  use graph_structure, only: graph_type
   use activation, only: activation_setup
   use clipper, only: clip_type
   use mpnn_layer, only: &
        mpnn_layer_type, method_container_type, &
-       state_method_type, message_method_type, readout_method_type, &
+       message_phase_type, readout_phase_type, &
        feature_type
   implicit none
   
 
   private
-  public :: conv_mpnn_layer_type, conv_method_container_type
-  public :: conv_readout_method_type, conv_state_method_type
+  public :: conv_mpnn_layer_type
 
 
-
-  type, extends(message_method_type) :: conv_message_method_type
-   contains
-     procedure :: update => update_message_conv
-     procedure :: calculate_partials => calculate_partials_message_conv
-     procedure :: set_shape => set_shape_message_conv
-  end type conv_message_method_type
-  interface conv_message_method_type
-    module function message_method_setup( &
-         num_vertex_features, num_edge_features, batch_size ) result(message_method)
-      integer, intent(in) :: num_vertex_features, num_edge_features, batch_size
-      type(conv_message_method_type) :: message_method
-    end function message_method_setup
-  end interface conv_message_method_type
-
-
-  type, extends(state_method_type) :: conv_state_method_type
-     !! weight has dimensions (feature_out, feature_in, vertex_degree, batch_size)
+  type, extends(message_phase_type) :: conv_message_phase_type
+     !! weight dimensions (feature_out, feature_in, vertex_degree, batch_size)
      real(real12), dimension(:,:,:), allocatable :: weight
      real(real12), dimension(:,:,:,:), allocatable :: dw
      type(feature_type), dimension(:), allocatable :: z
      class(activation_type), allocatable :: transfer
    contains
-     procedure :: get_num_params => get_num_params_state_conv
-     procedure :: get_params => get_params_state_conv
-     procedure :: set_params => set_params_state_conv
-     procedure :: get_gradients => get_gradients_state_conv
-     procedure :: set_gradients => set_gradients_state_conv
-     procedure :: set_shape => set_shape_state_conv
-     procedure :: update => update_state_conv
-     procedure :: calculate_partials => calculate_state_partials_conv
-  end type conv_state_method_type
-  interface conv_state_method_type
-    module function state_method_setup( &
-         num_vertex_features, num_edge_features, max_vertex_degree, batch_size ) result(state_method)
-      integer, intent(in) :: num_vertex_features, num_edge_features, max_vertex_degree, batch_size
-      type(conv_state_method_type) :: state_method
-    end function state_method_setup
-  end interface conv_state_method_type
+     procedure :: get_num_params => get_num_params_message_conv
+     procedure :: get_params => get_params_message_conv
+     procedure :: set_params => set_params_message_conv
+
+     procedure :: get_gradients => get_gradients_message_conv
+     procedure :: set_gradients => set_gradients_message_conv
+
+     procedure :: set_shape => set_shape_message_conv
+
+     procedure :: update => update_message_conv
+     procedure :: calculate_partials => calculate_partials_message_conv
+  end type conv_message_phase_type
+  interface conv_message_phase_type
+    module function message_phase_setup( &
+         num_vertex_features, num_edge_features, &
+         max_vertex_degree, batch_size ) result(message_phase)
+      integer, intent(in) :: num_vertex_features, num_edge_features, &
+           max_vertex_degree, batch_size
+      type(conv_message_phase_type) :: message_phase
[…]
