+++ src/lib/mod_mpnn_layer.f90
@@ -6,7 +6,7 @@
 !!!#############################################################################
 module mpnn_layer
   use constants, only: real12
-  use custom_types, only: graph_type
+  use graph_structure, only: graph_type
   use base_layer, only: learnable_layer_type
   use clipper, only: clip_type
   implicit none
@@ -15,8 +15,7 @@ module mpnn_layer
   private
 
   public :: mpnn_layer_type, feature_type, method_container_type
-  public :: state_method_type, message_method_type, readout_method_type
-  public :: update_state
+  public :: message_phase_type, readout_phase_type
   public :: update_message
   public :: get_output_readout
 
@@ -51,10 +50,9 @@ module mpnn_layer
     integer :: num_time_steps
     !! each dimension of num_features is for vertex and edge
     integer, dimension(2) :: num_features
-    !! state and message dimension is (time_step)
-    class(message_method_type), dimension(:), allocatable :: message
-    class(state_method_type), dimension(:), allocatable :: state
-    class(readout_method_type), allocatable :: readout
+    !! message dimension is (time_step)
+    class(message_phase_type), dimension(:), allocatable :: message
+    class(readout_phase_type), allocatable :: readout
    contains
     procedure(init_method), deferred, pass(this) :: init
   end type method_container_type
@@ -70,7 +68,7 @@ module mpnn_layer
   end type feature_type
 
 
-  type, abstract :: base_method_type
+  type, abstract :: base_phase_type
      !!! HAVE LOGICAL THAT STATES WHETHER IT IS LEARNABLE ???
      integer :: num_inputs
      integer :: num_outputs
@@ -79,44 +77,41 @@ module mpnn_layer
      type(feature_type), dimension(:), allocatable :: feature
      type(feature_type), dimension(:), allocatable :: di
    contains
-     procedure, pass(this) :: get_num_params => get_method_num_params
-     procedure, pass(this) :: get_params => get_method_params
-     procedure, pass(this) :: set_params => set_method_params
-     procedure, pass(this) :: get_gradients => get_method_gradients
-     procedure, pass(this) :: set_gradients => set_method_gradients
-     procedure, pass(this) :: set_shape => set_method_shape
-  end type base_method_type
+     procedure, pass(this) :: get_num_params => get_phase_num_params
+     procedure, pass(this) :: get_params => get_phase_params
+     procedure, pass(this) :: set_params => set_phase_params
+     procedure, pass(this) :: get_gradients => get_phase_gradients
+     procedure, pass(this) :: set_gradients => set_phase_gradients
+     procedure, pass(this) :: set_shape => set_phase_shape
+  end type base_phase_type
 
 
-  type, extends(base_method_type), abstract :: message_method_type
+  type, extends(base_phase_type), abstract :: message_phase_type
+     integer :: num_message_features
+     logical :: use_message = .true.
+     type(feature_type), dimension(:), allocatable :: message
    contains
      procedure(update_message), deferred, pass(this) :: update
      procedure(calculate_partials_message), deferred, pass(this) :: calculate_partials
-  end type message_method_type
+  end type message_phase_type
 
-  type, extends(base_method_type), abstract :: state_method_type
-   contains
-     procedure(update_state), deferred, pass(this) :: update
-     procedure(calculate_partials_state), deferred, pass(this) :: calculate_partials
-  end type state_method_type
-
-  type, extends(base_method_type), abstract :: readout_method_type
+  type, extends(base_phase_type), abstract :: readout_phase_type
    contains
      procedure(get_output_readout), deferred, pass(this) :: get_output
      procedure(calculate_partials_readout), deferred, pass(this) :: calculate_partials
-  end type readout_method_type
+  end type readout_phase_type
 
 
   abstract interface
      pure module subroutine update_message(this, input, graph)
[…]

[…]
