### 2024-05-28 181ff78
Add max_vertex_degree reference
