+++ src/lib/mod_conv_mpnn_layer.f90
@@ -2,6 +2,11 @@
 !!! Code written by Ned Thaddeus Taylor
 !!! Code part of the ATHENA library - a feedforward neural network library
 !!!#############################################################################
+!!! module contains implementation of a ...
+!!! ... convolutional message passing neural network layer
+!!! Original work by Duvenaud et al. (2015)
+!!! https://doi.org/10.48550/arXiv.1509.09292
+!!!#############################################################################
 module conv_mpnn_layer
   use constants, only: real12
   use misc, only: outer_product
@@ -20,6 +25,9 @@ module conv_mpnn_layer
   public :: conv_mpnn_layer_type
 
 
+!!!-----------------------------------------------------------------------------
+!!! convolutional message passing phase
+!!!-----------------------------------------------------------------------------
   type, extends(message_phase_type) :: conv_message_phase_type
      !! weight dimensions (feature_out, feature_in, vertex_degree, batch_size)
      real(real12), dimension(:,:,:), allocatable :: weight
@@ -39,6 +47,9 @@ module conv_mpnn_layer
      procedure :: update => update_message_conv
      procedure :: calculate_partials => calculate_partials_message_conv
   end type conv_message_phase_type
+
+  !! interface for the convolutional message passing phase
+  !!----------------------------------------------------------------------------
   interface conv_message_phase_type
     module function message_phase_setup( &
          num_vertex_features, num_edge_features, &
@@ -50,6 +61,9 @@ module conv_mpnn_layer
   end interface conv_message_phase_type
 
 
+!!!-----------------------------------------------------------------------------
+!!! convolutional readout passing phase
+!!!-----------------------------------------------------------------------------
   type, extends(readout_phase_type) :: conv_readout_phase_type
     integer :: num_time_steps
     real(real12), dimension(:,:,:), allocatable :: weight
@@ -69,6 +83,9 @@ module conv_mpnn_layer
      procedure :: get_output => get_output_readout_conv
      procedure :: calculate_partials => calculate_partials_readout_conv
   end type conv_readout_phase_type
+
+  !! interface for the convolutional readout passing phase
+  !!----------------------------------------------------------------------------
   interface conv_readout_phase_type
     module function readout_phase_setup( &
          num_time_steps, num_inputs, num_outputs, batch_size ) &
@@ -79,31 +96,17 @@ module conv_mpnn_layer
   end interface conv_readout_phase_type
 
 
-
-
-  type, extends(mpnn_layer_type) :: conv_mpnn_layer_type
-  end type conv_mpnn_layer_type
-
+!!!-----------------------------------------------------------------------------
+!!! convolutional MPNN method container
+!!!-----------------------------------------------------------------------------
   type, extends(method_container_type) :: conv_method_container_type
     integer :: max_vertex_degree = 6
    contains
     procedure, pass(this) :: init => init_conv_mpnn_method
   end type conv_method_container_type
 
-
-  interface conv_mpnn_layer_type
-    module function layer_setup( &
-           num_time_steps, &
-           num_vertex_features, num_edge_features, &
-           num_outputs, batch_size ) result(layer)
-      integer, intent(in) :: num_time_steps, num_vertex_features, &
-           num_edge_features, num_outputs, batch_size
-      type(conv_mpnn_layer_type) :: layer
-    end function layer_setup
-  end interface conv_mpnn_layer_type
-
-
-
+  !! interface for the convolutional MPNN method container
+  !!----------------------------------------------------------------------------
   interface conv_method_container_type
     module function method_setup( &
          num_vertex_features, num_edge_features, num_time_steps, &
@@ -120,6 +123,26 @@ module conv_mpnn_layer
   end interface conv_method_container_type
 
 
+!!!-----------------------------------------------------------------------------
+!!! convolutional MPNN layer
[…]
