### 2024-05-28 0d4a31c
Add di readout timestep implementation
