### 2024-05-28 nedtaylor/athena#28 @milancurcic
FWIW, all tests seem to pass with gfortran-12 in both debug and release modes.
