### 2024-05-28 #6492 @milancurcic
Now commenting on some parts of the paper. In the Statement of Need section, L26-31 states:

> While existing Fortran-based libraries, such as neural-fortran (Curcic (2019)), address many aspects of neural networks, the focus on convolutional neural networks is drastically reduced. ATHENA is developed to handle both fully-connected and convolutional layers, including the ability to handle 3D data for convolutional layers (a domain sometimes underappreciated in comparison to its 3D counterpart). The ATHENA library is developed to handle diverse layer types, including fully-connected, Dropout, pooling, and convolution.

I think "the focus on convolutional neural networks is drastically reduced" in neural-fortran is too strong of a statement. NF supports 2-d conv layers. What ATHENA brings new to the table here regarding convolutional networks are 3-d conv layers and average-pooling layers, which NF doesn't have. That said, 3-d conv layers and average-pooling layers are much less commonly used in practice than the 2-d conv + maxpool layers, so I think that the lack of 3-d conv layers in existing libraries motivating a whole new library is not convincing (to me!).

Rather, you can expand on what exactly is new. Before ATHENA, to the best of my knowledge, NF had the broadest set of features for neural networks in Fortran. So if we compare ATHENA features against NF, what stands out to me are:

* 3-d conv layers;
* average pooling layers
* 1, 2, 3-d batchnorm layers
* regularization layers (dropout and dropblock)

NF doesn't have any of these. I know that you mention these toward the end, but I suggest that you bring them up to the front and emphasize all the great and new things about ATHENA.

And then, there should also be a paragraph about why a new library is needed. IMO you could be contributing these new features to NF which has a track record (peer-reviewed publications) of use in the community. So, perhaps this is where you can tie in ATHENA's stakeholders and explain why a new and different API and/or software design was needed.

I apologize for the somewhat gatekeepy response. I have a personal interest in NF and I wish these features were pushed to NF rather than a new library (and more importantly, if you were another NF user). However, the more ML libraries for Fortran the better for Fortran overall, and I want you to succeed and have a strong paper describing ATHENA.
