### 2024-05-28 #6492 @milancurcic
I think we're very close. With a few more things ironed out on both the repo and paper sides, I'll be happy to thumbs up this paper.
