### 2024-05-28 nedtaylor/athena#28 @nedtaylor
Resolved in commit a8f0d619b0cdef6f3b1064367b400f7b16ab6772
