### 2024-05-28 7e83309
ifort compiler issue (#30)
* Fix unallocated, undefined and ifort array issues

* Fix temporary array warnings

* Fix unmatched array shapes

* Update CHANGELOG
