+++ src/lib/mod_optimiser.f90
@@ -226,6 +226,8 @@ contains
     if(present(lr_decay)) then
        if(allocated(optimiser%lr_decay)) deallocate(optimiser%lr_decay)
        allocate(optimiser%lr_decay, source = lr_decay)
+    else
+       allocate(optimiser%lr_decay, source = base_lr_decay_type())
     end if
 
     !! initialise gradients
@@ -350,6 +352,8 @@ contains
     if(present(lr_decay)) then
        if(allocated(optimiser%lr_decay)) deallocate(optimiser%lr_decay)
        allocate(optimiser%lr_decay, source = lr_decay)
+    else
+       allocate(optimiser%lr_decay, source = base_lr_decay_type())
     end if
 
      !! initialise nesterov boolean
@@ -467,6 +471,8 @@ contains
     if(present(lr_decay)) then
        if(allocated(optimiser%lr_decay)) deallocate(optimiser%lr_decay)
        allocate(optimiser%lr_decay, source = lr_decay)
+    else
+       allocate(optimiser%lr_decay, source = base_lr_decay_type())
     end if
 
     !! initialise adam parameters
@@ -577,6 +583,8 @@ contains
     if(present(lr_decay)) then
        if(allocated(optimiser%lr_decay)) deallocate(optimiser%lr_decay)
        allocate(optimiser%lr_decay, source = lr_decay)
+    else
+       allocate(optimiser%lr_decay, source = base_lr_decay_type())
     end if
 
     !! initialise adam parameters
@@ -685,6 +693,8 @@ contains
     if(present(lr_decay)) then
        if(allocated(optimiser%lr_decay)) deallocate(optimiser%lr_decay)
        allocate(optimiser%lr_decay, source = lr_decay)
+    else
+       allocate(optimiser%lr_decay, source = base_lr_decay_type())
     end if
 
     !! initialise adam parameters
