### 2024-05-28 88602d5
Specify release profile in fpm test and fpm run (#29)
* Specify release profile in fpm test and fpm run

* Update CHANGELOG

---------

Co-authored-by: Ned Taylor <71959356+nedtaylor@users.noreply.github.com>
