### 2024-05-28 #6492 @nedtaylor
@milancurcic I have addressed your current logged issues and pull requests. They are now all pushed to the `main` branch (no new release has been made in case new issues are found).

Response to paper discussion (where should I put this? In the paper branch, or just write it here?):

I mostly agree with your first point. I think, at the time, neural-fortran had no implementation for backward propagation of convolutional layers and, as far as I am aware no Fortran library other than athena currently handles other convolution features such as  different padding types (e.g. periodic), and data stride (hoping to get dilation implemented also, but that's a future plan).

My new paragraph for this is:
> While existing `Fortran-based` libraries, such as neural-fortran (Curcic (2019)), address many aspects of neural networks, ATHENA provides implementation of some well-known features not currently available within other libraries; these features include batchnormalisation, regularisation layers (such as dropout and dropblock), and average pooling layers. Additionally, the library provides support for 1, 2, and 3D input data for all features currently implemented; this includes 1, 2, and 3D data for convolutional layers. Finally, more features convolutional techiques are supported in the ATHENA library, including various data padding types, and stride.

For your second point, here is my update (I could go into more detail on reasons why a new API is being developed, but much of it would be speculative as it is not actualised in this current implementation and only found in development branches of the code so far):

> Notably, discussions with a spectrum of stakeholders have significantly influenced the development of ATHENA, placing paramount importance on accessibility and usability. This user-centric approach ensures that `ATHENA` is not just a library but a tool that seamlessly integrates with the evolving needs of the neural network community. One of the primary intended applications of ATHENA is in materials science, which heavily utilises convolutional and graph neural networks for learning based on charge densities and atomic structures. Given the unique data structure of atomic configurations, specifically their graph-based nature, a specialised API must be developed to accommodate these needs.

Let me know if you feel more detail and clarity is needed on why ATHENA should be a distinct library.

Finally, no need to apologise for your response, I completely understand and appreciate what you are doing. It is what the review process is for and I value the input you and the others are providing. This process is here to help us determine whether the library is of sufficient scientific/software quality and should be separate from NF and other existing Fortran NN libraries or should be implemented into an existing one. :)
