### 2024-05-28 #6492 @milancurcic
@nedtaylor I think your re-write is excellent and I have no objections with it!
