### 2024-05-28 #6492 @milancurcic
>  I think, at the time, neural-fortran had no implementation for backward propagation of convolutional layers

Indeed, that was confusing. For a while I actually thought the backward implementation wasn't working, but adding more tests for it in https://github.com/modern-fortran/neural-fortran/pull/174 convinced me that it does. The bug in the MNIST CNN example seems to be elsewhere but I haven't revisited it since.

> and, as far as I am aware no Fortran library other than athena currently handles other convolution features such as different padding types (e.g. periodic), and data stride

Yes, NF doesn't implement these. It may actually be worth it to mention padding and strides features in your statement of need.
