### 2024-05-29 1d1125d
Add conv1d_layer_type
