+++ src/lib/mod_container_layer_sub.f90
@@ -13,6 +13,7 @@ submodule(container_layer) container_layer_submodule
   use batchnorm1d_layer, only: batchnorm1d_layer_type
   use batchnorm2d_layer, only: batchnorm2d_layer_type
   use batchnorm3d_layer, only: batchnorm3d_layer_type
+  use conv1d_layer, only: conv1d_layer_type
   use conv2d_layer, only: conv2d_layer_type
   use conv3d_layer, only: conv3d_layer_type
   use dropout_layer, only: dropout_layer_type
@@ -46,6 +47,8 @@ contains
     type is(batchnorm3d_layer_type)
        call this%layer%forward(previous%output)
 
+    type is(conv1d_layer_type)
+       call this%layer%forward(previous%output)
     type is(conv2d_layer_type)
        call this%layer%forward(previous%output)
     type is(conv3d_layer_type)
@@ -98,6 +101,8 @@ contains
     type is(batchnorm3d_layer_type)
        call this%layer%backward(previous%output, gradient)
 
+    type is(conv1d_layer_type)
+       call this%layer%backward(previous%output, gradient)
     type is(conv2d_layer_type)
        call this%layer%backward(previous%output, gradient)
     type is(conv3d_layer_type)
