+++ src/lib/mod_network_sub.f90
@@ -34,6 +34,7 @@ submodule(network) network_submodule
   use batchnorm3d_layer, only: batchnorm3d_layer_type, read_batchnorm3d_layer
 
   !! convolution layer types
+  use conv1d_layer,    only: conv1d_layer_type, read_conv1d_layer
   use conv2d_layer,    only: conv2d_layer_type, read_conv2d_layer
   use conv3d_layer,    only: conv3d_layer_type, read_conv3d_layer
 
@@ -171,6 +172,8 @@ contains
          call this%add(read_batchnorm2d_layer(unit))
       case("BATCHNORM3D")
          call this%add(read_batchnorm3d_layer(unit))
+      case("CONV1D")
+         call this%add(read_conv1d_layer(unit))
       case("CONV2D")
          call this%add(read_conv2d_layer(unit))
       case("CONV3D")
@@ -490,6 +493,18 @@ contains
          case(1)
             t_input_layer = input1d_layer_type(input_shape = next%input_shape)
             allocate(this%model(1)%layer, source = t_input_layer)
+         case(2)
+            select type(next)
+            type is(conv1d_layer_type)
+               t_input_layer = input1d_layer_type(&
+                    input_shape = next%input_shape + &
+                    [2*next%pad,0])
+               allocate(this%model(1)%layer, source = t_input_layer)
+            class default
+               t_input_layer = input1d_layer_type(&
+                    input_shape = next%input_shape)
+               allocate(this%model(1)%layer, source = t_input_layer)
+            end select
          case(3)
             select type(next)
             type is(conv2d_layer_type)
@@ -524,9 +539,7 @@ contains
 !!! ignore calcuation of input gradients for 1st non-input layer
 !!!-----------------------------------------------------------------------------
     select type(second => this%model(2)%layer)
-    type is(conv2d_layer_type)
-       second%calc_input_gradients = .false.
-    type is(conv3d_layer_type)
+    class is(conv_layer_type)
        second%calc_input_gradients = .false.
     end select
 
@@ -941,6 +954,8 @@ end function get_gradients
        type is(batchnorm3d_layer_type)
           call this%model(i)%backward(this%model(i-1),next%di)
 
+       type is(conv1d_layer_type)
+          call this%model(i)%backward(this%model(i-1),next%di)
        type is(conv2d_layer_type)
           call this%model(i)%backward(this%model(i-1),next%di)
        type is(conv3d_layer_type)
