### 2024-05-29 0afea09
Add conv1d_layer_type
