+++ src/lib/mod_conv1d_layer.f90
@@ -1,0 +1,959 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains implementation of a 1D convolutional layer
+!!!#############################################################################
+!!! Attribution statement:
+!!! The following procedures are based on code from the neural-fortran library
+!!! https://github.com/modern-fortran/neural-fortran/blob/main/src/nf/nf_layer.f90
+!!! procedures:
+!!! - get_params*
+!!! - set_params*
+!!! - get_gradients*
+!!! - set_gradients*
+!!!#############################################################################
+module conv1d_layer
+  use constants, only: real12
+  use base_layer, only: learnable_layer_type, conv_layer_type
+  use custom_types, only: initialiser_type
+  implicit none
+  
+  
+  type, extends(conv_layer_type) :: conv1d_layer_type
+     real(real12), allocatable, dimension(:,:,:) :: weight
+     real(real12), allocatable, dimension(:,:,:,:) :: dw ! weight gradient
+     real(real12), allocatable, dimension(:,:,:) :: output, z
+     real(real12), allocatable, dimension(:,:,:) :: di ! input gradient
+   contains
+     procedure, pass(this) :: get_params => get_params_conv1d
+     procedure, pass(this) :: set_params => set_params_conv1d
+     procedure, pass(this) :: get_gradients => get_gradients_conv1d
+     procedure, pass(this) :: set_gradients => set_gradients_conv1d
+     procedure, pass(this) :: get_output => get_output_conv1d
+
+     procedure, pass(this) :: init => init_conv1d
+     procedure, pass(this) :: set_batch_size => set_batch_size_conv1d
+     procedure, pass(this) :: print => print_conv1d
+
+     procedure, pass(this) :: forward  => forward_rank
+     procedure, pass(this) :: backward => backward_rank
+     procedure, private, pass(this) :: forward_3d
+     procedure, private, pass(this) :: backward_3d
+
+     procedure, pass(this) :: reduce => layer_reduction
+     procedure, pass(this) :: merge => layer_merge
+     procedure :: add_t_t => layer_add  !t = type, r = real, i = int
+     generic :: operator(+) => add_t_t !, public
+  end type conv1d_layer_type
+
+  
+!!!-----------------------------------------------------------------------------
+!!! interface for layer set up
+!!!-----------------------------------------------------------------------------
+  interface conv1d_layer_type
+     module function layer_setup( &
+          input_shape, batch_size, &
+          num_filters, kernel_size, stride, padding, &
+          activation_function, activation_scale, &
+          kernel_initialiser, bias_initialiser, &
+          calc_input_gradients) result(layer)
+       integer, dimension(:), optional, intent(in) :: input_shape
+       integer, optional, intent(in) :: batch_size
+       integer, optional, intent(in) :: num_filters
+       integer, dimension(..), optional, intent(in) :: kernel_size
+       integer, dimension(..), optional, intent(in) :: stride
+       real(real12), optional, intent(in) :: activation_scale
+       character(*), optional, intent(in) :: activation_function, &
+            kernel_initialiser, bias_initialiser, padding
+       logical, optional, intent(in) :: calc_input_gradients
+       type(conv1d_layer_type) :: layer
+     end function layer_setup
+  end interface conv1d_layer_type
+
+
+  private
+  public :: conv1d_layer_type
+  public :: read_conv1d_layer
+
+
+contains
+
+!!!#############################################################################
+!!! layer reduction
+!!!#############################################################################
[…]
