### 2024-05-29 5ff26f0
Add conv1d_layer_type
