### 2024-05-29 612ab7e
Add avgpool1d_layer_type
