+++ src/lib/mod_container_layer_sub.f90
@@ -19,6 +19,7 @@ submodule(container_layer) container_layer_submodule
   use dropout_layer, only: dropout_layer_type
   use dropblock2d_layer, only: dropblock2d_layer_type
   use dropblock3d_layer, only: dropblock3d_layer_type
+  use avgpool1d_layer, only: avgpool1d_layer_type
   use avgpool2d_layer, only: avgpool2d_layer_type
   use avgpool3d_layer, only: avgpool3d_layer_type
   use maxpool2d_layer, only: maxpool2d_layer_type
@@ -61,6 +62,8 @@ contains
     type is(dropblock3d_layer_type)
        call this%layer%forward(previous%output)
 
+    type is(avgpool1d_layer_type)
+       call this%layer%forward(previous%output)
     type is(avgpool2d_layer_type)
        call this%layer%forward(previous%output)
     type is(avgpool3d_layer_type)
@@ -115,6 +118,8 @@ contains
     type is(dropblock3d_layer_type)
        call this%layer%backward(previous%output, gradient)
 
+    type is(avgpool1d_layer_type)
+       call this%layer%backward(previous%output, gradient)
     type is(avgpool2d_layer_type)
        call this%layer%backward(previous%output, gradient)
     type is(avgpool3d_layer_type)
