### 2024-05-29 1e7ceeb
Add avgpool1d_layer_type
