+++ src/lib/mod_avgpool1d_layer.f90
@@ -1,0 +1,463 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains implementation of a 2D average pooling layer
+!!!#############################################################################
+module avgpool1d_layer
+  use constants, only: real12
+  use base_layer, only: pool_layer_type
+  implicit none
+  
+  
+  type, extends(pool_layer_type) :: avgpool1d_layer_type
+     real(real12), allocatable, dimension(:,:,:) :: output
+     real(real12), allocatable, dimension(:,:,:) :: di ! gradient of input (i.e. delta)
+   contains
+     procedure, pass(this) :: get_output => get_output_avgpool1d
+     procedure, pass(this) :: init => init_avgpool1d
+     procedure, pass(this) :: set_batch_size => set_batch_size_avgpool1d
+     procedure, pass(this) :: print => print_avgpool1d
+     procedure, pass(this) :: forward  => forward_rank
+     procedure, pass(this) :: backward => backward_rank
+     procedure, private, pass(this) :: forward_3d
+     procedure, private, pass(this) :: backward_3d
+  end type avgpool1d_layer_type
+
+  
+  interface avgpool1d_layer_type
+     module function layer_setup( &
+          input_shape, batch_size, &
+          pool_size, stride) result(layer)
+       integer, dimension(:), optional, intent(in) :: input_shape
+       integer, optional, intent(in) :: batch_size 
+       integer, dimension(..), optional, intent(in) :: pool_size
+       integer, dimension(..), optional, intent(in) :: stride
+       type(avgpool1d_layer_type) :: layer
+     end function layer_setup
+  end interface avgpool1d_layer_type
+
+
+  private
+  public :: avgpool1d_layer_type
+  public :: read_avgpool1d_layer
+
+
+contains
+
+!!!#############################################################################
+!!! get layer outputs
+!!!#############################################################################
+  pure subroutine get_output_avgpool1d(this, output)
+    implicit none
+    class(avgpool1d_layer_type), intent(in) :: this
+    real(real12), allocatable, dimension(..), intent(out) :: output
+  
+    select rank(output)
+    rank(1)
+       output = reshape(this%output, [size(this%output)])
+    rank(2)
+       output = &
+            reshape(this%output, [product(this%output_shape),this%batch_size])
+    rank(3)
+       output = this%output
+    end select
+  
+  end subroutine get_output_avgpool1d
+!!!#############################################################################
+
+
+!!!##########################################################################!!!
+!!! * * * * * * * * * * * * * * * * * *  * * * * * * * * * * * * * * * * * * !!!
+!!!##########################################################################!!!
+
+
+!!!#############################################################################
+!!! forward propagation assumed rank handler
+!!!#############################################################################
+  pure subroutine forward_rank(this, input)
+    implicit none
+    class(avgpool1d_layer_type), intent(inout) :: this
+    real(real12), dimension(..), intent(in) :: input
+
+    select rank(input); rank(3)
+       call forward_3d(this, input)
+    end select
+  end subroutine forward_rank
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! backward propagation assumed rank handler
+!!!#############################################################################
+  pure subroutine backward_rank(this, input, gradient)
+    implicit none
+    class(avgpool1d_layer_type), intent(inout) :: this
+    real(real12), dimension(..), intent(in) :: input
+    real(real12), dimension(..), intent(in) :: gradient
+
+    select rank(input); rank(3)
+    select rank(gradient); rank(3)
[…]
