### 2024-05-29 d7645d0
Add avgpool1d_layer_type
