+++ CMakeLists.txt
@@ -103,6 +103,7 @@ set(LIB_FILES
     mod_dropout_layer.f90
     mod_dropblock2d_layer.f90
     mod_dropblock3d_layer.f90
+    mod_avgpool1d_layer.f90
     mod_avgpool2d_layer.f90
     mod_avgpool3d_layer.f90
     mod_maxpool2d_layer.f90
