### 2024-05-29 71e562d
Add maxpool1d_layer_type
