### 2024-05-29 b72f790
Add maxpool1d_layer_type
