### 2024-05-29 170d3d1
Add maxpool1d_layer_type
