### 2024-05-29 c0ac0de
Add maxpool1d_layer_type
