+++ src/lib/mod_maxpool1d_layer.f90
@@ -2,30 +2,30 @@
 !!! Code written by Ned Thaddeus Taylor
 !!! Code part of the ATHENA library - a feedforward neural network library
 !!!#############################################################################
-!!! module contains implementation of a 1D average pooling layer
+!!! module contains implementation of a 1D maxpooling layer
 !!!#############################################################################
-module avgpool1d_layer
+module maxpool1d_layer
   use constants, only: real12
   use base_layer, only: pool_layer_type
   implicit none
   
   
-  type, extends(pool_layer_type) :: avgpool1d_layer_type
+  type, extends(pool_layer_type) :: maxpool1d_layer_type
      real(real12), allocatable, dimension(:,:,:) :: output
      real(real12), allocatable, dimension(:,:,:) :: di ! gradient of input (i.e. delta)
    contains
-     procedure, pass(this) :: get_output => get_output_avgpool1d
-     procedure, pass(this) :: init => init_avgpool1d
-     procedure, pass(this) :: set_batch_size => set_batch_size_avgpool1d
-     procedure, pass(this) :: print => print_avgpool1d
+     procedure, pass(this) :: get_output => get_output_maxpool1d
+     procedure, pass(this) :: init => init_maxpool1d
+     procedure, pass(this) :: set_batch_size => set_batch_size_maxpool1d
+     procedure, pass(this) :: print => print_maxpool1d
      procedure, pass(this) :: forward  => forward_rank
      procedure, pass(this) :: backward => backward_rank
      procedure, private, pass(this) :: forward_3d
      procedure, private, pass(this) :: backward_3d
-  end type avgpool1d_layer_type
+  end type maxpool1d_layer_type
 
   
-  interface avgpool1d_layer_type
+  interface maxpool1d_layer_type
      module function layer_setup( &
           input_shape, batch_size, &
           pool_size, stride) result(layer)
@@ -33,14 +33,14 @@ module avgpool1d_layer
        integer, optional, intent(in) :: batch_size 
        integer, dimension(..), optional, intent(in) :: pool_size
        integer, dimension(..), optional, intent(in) :: stride
-       type(avgpool1d_layer_type) :: layer
+       type(maxpool1d_layer_type) :: layer
      end function layer_setup
-  end interface avgpool1d_layer_type
+  end interface maxpool1d_layer_type
 
 
   private
-  public :: avgpool1d_layer_type
-  public :: read_avgpool1d_layer
+  public :: maxpool1d_layer_type
+  public :: read_maxpool1d_layer
 
 
 contains
@@ -48,9 +48,9 @@ contains
 !!!#############################################################################
 !!! get layer outputs
 !!!#############################################################################
-  pure subroutine get_output_avgpool1d(this, output)
+  pure subroutine get_output_maxpool1d(this, output)
     implicit none
-    class(avgpool1d_layer_type), intent(in) :: this
+    class(maxpool1d_layer_type), intent(in) :: this
     real(real12), allocatable, dimension(..), intent(out) :: output
   
     select rank(output)
@@ -63,7 +63,7 @@ contains
        output = this%output
     end select
   
-  end subroutine get_output_avgpool1d
+  end subroutine get_output_maxpool1d
 !!!#############################################################################
 
 
@@ -77,7 +77,7 @@ contains
 !!!#############################################################################
   pure subroutine forward_rank(this, input)
     implicit none
-    class(avgpool1d_layer_type), intent(inout) :: this
+    class(maxpool1d_layer_type), intent(inout) :: this
     real(real12), dimension(..), intent(in) :: input
 
     select rank(input); rank(3)
@@ -92,7 +92,7 @@ contains
[…]
