### 2024-05-29 bf27900
Add input2d_layer_type
