### 2024-05-29 eba8917
Add input2d_layer_type
