### 2024-05-29 12761f7
Add input2d_layer_type
