+++ src/lib/mod_input2d_layer.f90
@@ -1,0 +1,256 @@
+!!!#############################################################################
+!!! Code written by Ned Thaddeus Taylor
+!!! Code part of the ATHENA library - a feedforward neural network library
+!!!#############################################################################
+!!! module contains implementation of a 2D input layer
+!!!#############################################################################
+module input2d_layer
+  use constants, only: real12
+  use base_layer, only: input_layer_type
+  implicit none
+  
+  
+  type, extends(input_layer_type) :: input2d_layer_type
+     real(real12), allocatable, dimension(:,:,:) :: output
+   contains
+     procedure, pass(this) :: get_output => get_output_input2d
+     procedure, pass(this) :: init => init_input2d
+     procedure, pass(this) :: set_batch_size => set_batch_size_input2d
+     procedure, pass(this) :: forward  => forward_rank
+     procedure, pass(this) :: backward => backward_rank
+     procedure, pass(this) :: set => set_input2d
+  end type input2d_layer_type
+
+  interface input2d_layer_type
+     module function layer_setup(input_shape, batch_size) result(layer)
+       integer, dimension(:), optional, intent(in) :: input_shape
+       integer, optional, intent(in) :: batch_size 
+       type(input2d_layer_type) :: layer
+     end function layer_setup
+  end interface input2d_layer_type
+
+  
+  private
+  public :: input2d_layer_type
+
+
+contains
+
+!!!#############################################################################
+!!! get layer outputs
+!!!#############################################################################
+  pure subroutine get_output_input2d(this, output)
+    implicit none
+    class(input2d_layer_type), intent(in) :: this
+    real(real12), allocatable, dimension(..), intent(out) :: output
+  
+    select rank(output)
+    rank(1)
+       output = reshape(this%output, [size(this%output)])
+    rank(2)
+       output = &
+            reshape(this%output, [product(this%output_shape),this%batch_size])
+    rank(3)
+       output = this%output
+    end select
+  
+  end subroutine get_output_input2d
+!!!#############################################################################
+
+
+!!!##########################################################################!!!
+!!! * * * * * * * * * * * * * * * * * *  * * * * * * * * * * * * * * * * * * !!!
+!!!##########################################################################!!!
+
+
+!!!#############################################################################
+!!! forward propagation assumed rank handler
+!!! placeholder to satisfy deferred
+!!!#############################################################################
+  pure subroutine forward_rank(this, input)
+    implicit none
+    class(input2d_layer_type), intent(inout) :: this
+    real(real12), dimension(..), intent(in) :: input
+
+    select rank(input)
+    rank(1)
+       this%output = reshape(input, shape=shape(this%output))
+    rank(2)
+       this%output = reshape(input, shape=shape(this%output))
+    rank(3)
+       this%output = reshape(input, shape=shape(this%output))
+    rank(4)
+       this%output = reshape(input, shape=shape(this%output))
+    rank(5)
+       this%output = reshape(input, shape=shape(this%output))
+    rank(6)
+       this%output = reshape(input, shape=shape(this%output))
+    end select
+  end subroutine forward_rank
+!!!#############################################################################
+
+
+!!!#############################################################################
+!!! backward propagation assumed rank handler
+!!! placeholder to satisfy deferred
+!!!#############################################################################
+  pure subroutine backward_rank(this, input, gradient)
+    implicit none
+    class(input2d_layer_type), intent(inout) :: this
+    real(real12), dimension(..), intent(in) :: input
+    real(real12), dimension(..), intent(in) :: gradient
+    return
+  end subroutine backward_rank
+!!!#############################################################################
+
+
+!!!##########################################################################!!!
+!!! * * * * * * * * * * * * * * * * * *  * * * * * * * * * * * * * * * * * * !!!
+!!!##########################################################################!!!
+
+
[…]
