### 2024-05-29 91de914
Add input2d_layer_type test
