+++ test/test_input_layer.f90
@@ -1,6 +1,7 @@
 program test_input_layer
   use athena, only: &
        input1d_layer_type, &
+       input2d_layer_type, &
        input3d_layer_type, &
        input4d_layer_type, &
        base_layer_type
@@ -78,6 +79,67 @@ program test_input_layer
      call input_layer%forward(input_6d)
      if(any(abs(input_layer%output-6.E0).gt.1.E-6))then
         write(0,*) 'input1d_layer forward 4d failed'
+        write(*,*) input_layer%output
+        success = .false.
+     end if
+     call input_layer%backward([1.E0], [1.E0])
+  end select
+
+
+!!!-----------------------------------------------------------------------------
+!!! set up input2d layer
+!!!-----------------------------------------------------------------------------
+  deallocate(input_layer)
+  input_layer = input2d_layer_type( &
+       input_shape=shape(input_3d), &
+       batch_size=batch_size)
+  if(input_layer%batch_size.ne.batch_size)then
+     write(0,*) 'input2d_layer batch_size failed'
+     success = .false.
+  end if
+  select type(input_layer)
+  type is (input2d_layer_type)
+     call input_layer%set(input_1d)
+     call input_layer%forward(input_1d)
+     call input_layer%get_output(output_1d)
+     if(any(abs(output_1d-1.E0).gt.1.E-6))then
+        write(0,*) 'input2d_layer forward 1d failed'
+        write(*,*) input_layer%output
+        success = .false.
+     end if
+     call input_layer%set(input_2d)
+     call input_layer%forward(input_2d)
+     call input_layer%get_output(output_2d)
+     if(any(abs(output_2d-2.E0).gt.1.E-6))then
+        write(0,*) 'input2d_layer forward 2d failed'
+        write(*,*) input_layer%output
+        success = .false.
+     end if
+     call input_layer%set(input_3d)
+     call input_layer%forward(input_3d)
+     if(any(abs(input_layer%output-3.E0).gt.1.E-6))then
+        write(0,*) 'input2d_layer forward 3d failed'
+        write(*,*) input_layer%output
+        success = .false.
+     end if
+     call input_layer%set(input_4d)
+     call input_layer%forward(input_4d)
+     if(any(abs(input_layer%output-4.E0).gt.1.E-6))then
+        write(0,*) 'input2d_layer forward 4d failed'
+        write(*,*) input_layer%output
+        success = .false.
+     end if
+     call input_layer%set(input_5d)
+     call input_layer%forward(input_5d)
+     if(any(abs(input_layer%output-5.E0).gt.1.E-6))then
+        write(0,*) 'input2d_layer forward 5d failed'
+        write(*,*) input_layer%output
+        success = .false.
+     end if
+     call input_layer%set(input_6d)
+     call input_layer%forward(input_6d)
+     if(any(abs(input_layer%output-6.E0).gt.1.E-6))then
+        write(0,*) 'input2d_layer forward 6d failed'
         write(*,*) input_layer%output
         success = .false.
      end if
