### 2024-05-29 baf3cf3
Add conv1d_layer_type test
