### 2024-05-29 4208eb2
Add input2d_layer_type reference
