### 2024-05-29 b3b9e2b
Fix output shape initialisation
