### 2024-05-29 0cf1fad
Fix output shape initialisation
