### 2024-05-29 #6492 @HaoZeke
> @milancurcic and @HaoZeke. Unfortunately, I noticed an error with the updated text I provided for the Statement of Need regarding support for 1, 2, and 3D data for all features. 1D support is a new feature I have implemented since submitting the paper and it has not been in this review process. It is currently available in the `development` branch, so can be pushed up to `main` for testing if reviewers want, or I can amend the Statement of Need to remove reference to 1D.
>
> The added features are:
>
>     * 1D convolution layer
>
>     * 2D input layer (i.e. for handling 1D image data with 1 channel dimension)
>
>     * 1D average pooling layer
>
>     * 1D maxpooling layer
>
>
> These features all have their associated test codes developed (basically carbon copies of their 2D counterparts) and the only changes are in the following files:
>
>     * src/lib/mod_avgpool1d_layer.f90 (new file)
>
>     * src/lib/mod_maxpool1d_layer.f90 (new file)
>
>     * src/lib/mod_conv1d_layer.f90 (new file)
>
>     * src/lib/mod_input2d_layer.f90 (new file)
>
>     * test/test_avgpool1d_layer.f90 (new file)
>
>     * test/test_maxpool1d_layer.f90 (new file)
>
>     * test/test_conv1d_layer.f90 (new file)
>
>     * test/test_conv1d_network.f90 (new file)
>
>     * test/test_input_layer.f90 (74 lines changed to include input2d test)
>
>     * src/lib/mod_network_sub.f90 (32 lines changed to include reference to new layer types)
>
>     * src/lib/mod_container_layer_sub.f90 (20 lines changed to include reference to new layer types)
>
>     * src/athena.f90 (4 lines changed to include reference to new layer types)
>
>     * CHANGELOG (4 lines changed to reference changes)
>
>     * CMakeLists.txt (4 lines to reference new files)
>
>     * test/CMakeLists.txt (4 lines to reference new files)
>
>
> From my checks, all tests have passed cmake and fpm (build and release, intel and gcc compiler), but would, of course, need checking by reviewers.
>
> **The three options I can see are (I would prefer option 3, but understand if this is too difficult/frustrating/time consuming):**
>
>     1. **exclude these 1D features** from the paper release and remove reference to it from the paper (include 1D feature in the next code release after the paper)
>
>     2. **exclude these 1D features** from the paper release **but refer to the 1D features as a planned future release** in the paper
>
>     3. **Include 1D features** and extend review process accordingly (hopefully not long as all features are just copies of their higher-dimension counterparts)

I think its alright to have 1D support merged if @milancurcic @awvwgk and @jrybarczyk don't mind taking a look at the changed files.
