### 2024-05-29 #6492 @nedtaylor
> Agree, please go ahead and add the 1D features.

Thanks. 🙂

@jrybarczyk, @milancurcic, and @awvwgk. The features are now implemented into the `main` branch, ready for review.
