### 2024-05-30 169215f
Add clip handling to get_gradients
