### 2024-05-30 198ca03
Add weight initialiser
