### 2024-05-30 47a0a9c
Add clipping method
