### 2024-06-28 nedtaylor/athena#39 @milancurcic
Statement of need in the docs
Following the checklist, can you point me to, or add if not present, the statement of need to the docs (ideally near the top of the README.md):

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Part of https://github.com/openjournals/joss-reviews/issues/6492
