### 2024-06-28 nedtaylor/athena#39 @nedtaylor
Would this be the same as the Statement of Need provided in the paper.md (in the `paper` branch)? If so, I'll paste that into the README.md near the top.
