### 2024-06-28 #6492 @milancurcic
@nedtaylor Where can I find the latest article proof? The one linked in the PR is outdated.
