### 2024-06-28 nedtaylor/athena#39 @milancurcic
I don't think it needs to be but I also think it can be.
