### 2024-06-28 #6492 @nedtaylor
> @nedtaylor Where can I find the latest article proof? The one linked in the PR is outdated.

@milancurcic, not sure where I am able to upload it. I have attached it to this comment, but I don't know if that is what people would want. It can also be found as an artifact to this
 GitHub workflow action: https://github.com/nedtaylor/athena/actions/runs/9712052018

[paper.pdf](https://github.com/user-attachments/files/16028730/paper.pdf)
