### 2024-06-28 nedtaylor/athena#39 @nedtaylor
Added in commit 0b67b82 to main branch.

Slightly modified and cutdown version of what is in the paper.
