### 2024-06-28 #6492 @nedtaylor
> > I think all reviews have been addressed and this paper can go forward. Thanks all.
>
> Thanks everyone for taking the time to review and oversee this, I greatly appreciate it!
>
> I have made a new release with these latest fixes (1.3.2).
>
> I have a pull request reviewed and tested and ready to merge the fixes into the `paper` branch. Let me know if you want me to go ahead with this. An additional commit will then be needed to update the version number in the `codemeta.json` file.

Ive completed the pull request into the `paper` branch.
