### 2024-07-03 4c81041
Update LICENSE year and version number
