+++ fpm.toml
@@ -1,5 +1,5 @@
 name = "athena"
-version = "1.3.2"
+version = "1.3.3"
 license = "MIT"
 author = "Ned Thaddeus Taylor"
 maintainer = "n.t.taylor@exeter.ac.uk"
