### 2024-07-03 #6492 @nedtaylor
@HaoZeke, three questions:
1. Do I need to update any of the dates on the paper.md or codemeta.json files?
2. Do I need to change something to compile the paper in non-draft format, or is that something that is handled by someone else?
3. Does the figshare release just need to be a zip archived there, or is anything else needed also (and should it be the `paper` or `main` branch)?

## Tasks completed by author
- [ ] Check authors and affiliations (including ORCIDs)
- [ ] Make a release of the software with latest changes from review and post it here
- [ ] Archive of figshare and post DOI here
- [ ] Make sure the title and author list in the archive match those in the JOSS paper
- [ ] Make sure the license listed for the archive is the same as the software license
