### 2024-07-03 #6492 @nedtaylor
> > @HaoZeke, three questions:
> > ```
> 1. Yup, the submitted date on the `paper.pdf` in the earlier comment says "Submitted: 01 January 1970" which should be modified, and the code-meta should be regenerated after the new version
> 2. AFAIK the paper options do not need to be modified, it will be switched away from the draft format automatically on acceptance
> 3. The figshare should be the `main` branch, since the `paper` branch artifact is effectively the article itself

@HaoZeke

1. My `paper.md` already has the following line in it, but doesn't seem to get used when compiling the `paper.pf`. I've followed the template provided [here](https://github.com/openjournals/docs/blob/master//submitting.md), so I wonder whether this is just how the draft is formatted.
`date: 18 February 2024`
I have updated the codemeta.json `dataModified` line to:
`  "dateModified": "2024-07-03",`
but I don't think I should be changing the `"datePublished"` line as that is not in my control to choose when it gets published.
2. Cool.
3. Done.

Latest release of software with changes: version 1.3.3 (only change fom 1.3.2 is updated license year range)
https://github.com/nedtaylor/athena/releases/tag/1.3.3

figshare DOI: [10.6084/m9.figshare.26158630](https://doi.org/10.6084/m9.figshare.26158630.v1)

I think I have completed all the tasks on my end now. Let me know if there is more I need to do. :)
