### 2024-07-03 #6492 @nedtaylor
@HaoZeke

These are ACM Digital Library DOIs (association for Computing Machinery) and appear to be the only places these works are published, from what I can find. The links are:

DropBlock: https://dl.acm.org/doi/10.1145/3474085.3475302
Batch normalisation: https://dl.acm.org/doi/10.5555/3045118.3045167
Dropout: https://dl.acm.org/doi/10.5555/2627435.2670313

Are these sufficient, or do I need to find a different publication for these to reference to?
