### 2024-07-04 5269230
Fix reference to number of pooling dimensions
