### 2024-07-04 9abca1e
Fix broken DOIs
