+++ paper.bib
@@ -28,7 +28,7 @@
       pages = {10750–10760},
       numpages = {11},
       location = {Montr\'{e}al, Canada},
-      doi = {10.5555/3327546.3327732},
+      url = {https://dl.acm.org/doi/10.5555/3327546.3327732},
       series = {NIPS'18}
 }
 
@@ -41,7 +41,7 @@
       booktitle = {Proceedings of the 32nd International Conference on International Conference on Machine Learning - Volume 37},
       pages = {448–456},
       numpages = {9},
-      doi = {10.5555/3045118.3045167},
+      url = {https://dl.acm.org/doi/10.5555/3045118.3045167},
       location = {Lille, France},
       series = {ICML'15}
 }
@@ -60,7 +60,7 @@
       month = {jan},
       pages = {1929–1958},
       numpages = {30},
-      doi = {10.5555/2627435.2670313},
+      url = {https://dl.acm.org/doi/10.5555/2627435.2670313},
       keywords = {deep learning, model combination, neural networks, regularization}
 }
 
