### 2024-07-04 #6492 @nedtaylor
> @nedtaylor thanks for looking into this. AFAIK these should be inproceedings, i.e. something like this:
>
> ```
> @inproceedings{bengioInputOutputHMM1994,
>   title = {An {{Input Output HMM Architecture}}},
>   booktitle = {Advances in {{Neural Information Processing Systems}}},
>   author = {Bengio, Yoshua and Frasconi, Paolo},
>   date = {1994},
>   volume = {7},
>   publisher = {MIT Press},
>   url = {https://proceedings.neurips.cc/paper/1994/hash/8065d07da4a77621450aa84fee5656d9-Abstract.html},
>   urldate = {2023-03-12},
>   abstract = {We  introduce  a  recurrent  architecture  having  a  modular structure  and we formulate a training procedure based on the EM  algorithm.  The resulting model has similarities to hidden  Markov models, but  supports  recurrent  networks  processing style and allows  to exploit  the supervised  learning paradigm while using maximum likelihood  estimation.},
> }
> ```

@HaoZeke, thanks for the response. I have made the required changes to `inproceedings` (the Dropout paper seems to be an article and not an inproceedings, so I have not changed it from an `article` to `inproceedings`) and have used the `url` tag instead of `doi` data where appropriate. I have now pushed this to the paper branch. When a new paper is compiled, it will probably come up with saying that their DOIs are missing though, so please be aware of that.

Please let me know if this resolves the problem.
