### 2024-07-04 #6492 @HaoZeke
@nedtaylor it seems like all the NeurIPS ones are to be coded as `inproceedings` as I understand, since they don't have a usable DOI, but I'll ping a few more editors, maybe @crvernon has an opinion as well?
