### 2024-07-04 #6492 @nedtaylor
> @nedtaylor it seems like all the NeurIPS ones are to be coded as `inproceedings` as I understand, since they don't have a usable DOI, but I'll ping a few more editors, maybe @crvernon has an opinion as well?
>
> Also the Missing ones are a suggestion, so I can also just let is slide as is.

@HaoZeke

Maybe this information will help for clarity and in determining how to handle these.

I've set both DropBlock and Batchnorm papers as `inproceedings` (they've been that for a whiles) and the editor bot is still bringing them up as "missing DOIs". I'm happy to change Dropout to an `inproceedings`, but that's not what it says it is published in and their website when exporting that specific citation returns a bib `article` format (unlike the previous two, which export as `inproceedings`.
