### 2024-07-04 #6492 @crvernon
@nedtaylor and @HaoZeke - it is fine to have a justification here for the missing DOI.  If there is none available for the record, regardless of the specification, then we will be OK with publishing it as such.  The more critical notices are those that get labeled as invalid, which it looks like you don't have any of those in your listing.  Thanks!
