### 2024-07-04 #6492 @nedtaylor
> @editorialbot recommend-accept
>
> Thanks for the additional context and quick turnaround @nedtaylor and @crvernon!! I'm happy to recommend accepting this, after all the excellent effort put in by the reviewers (@awvwgk, @milancurcic and @jrybarczyk) and author (@nedtaylor).

I've had a quick look through the final proof and am happy with it.

Thanks very much, @HaoZeke and @crvernon for the fantastic editorial work! 😄

And thanks to the reviewers, (@awvwgk, @jrybarczyk, and @milancurcic ) for the massive amount of work and dedication put towards reviewing this paper and the code! I greatly appreciate it! 😄
