### 2024-07-04 #6492 @crvernon
:wave: @nedtaylor - You are almost there before we are ready to accept this submission for publication.  Please correct the following (no need to make a new release for these):

In the paper:
- [ ] LINE 26: "neural-fortran (Curcic (2019)), address" should read "neural-fortran (Curcic, 2019), address" please fix the formatting of this citation in the paper file.
- [ ] LINE 28: "batchnormalisation" needs a space between "batch" and "normalisation"
- [ ] LINES 31-32:  "Finally, more features convolutional techniques are supported in the ATHENA library, including various data padding types, and stride."  This sentence needs some work.  Perhaps the following, "Finally, more features for convolutional techniques are supported in the ATHENA library, including various data padding types and stride."  Note the addition of "for" and the removal of the comma before "and stride".
- [ ] LINE 45:  The citation "Ioffe & Szegedy (2015))" should appear as "Ioffe & Szegedy, 2015)" in this case.  Please fix the formatting in the paper.
- [ ] LINE 46:  The citation "Ghiasi et al. (2018))" should appear as "Ghiasi et al., 2018)" in the paper.
- [ ] LINE 54:  Preserve capitalization for the "F" in "fortran".  This can be done in your bib file using curly brackets around how you want the case to appear in the paper.
- [ ] LINE 73:  the "F" in "fortran" should be capitalized.

Let me know when these have been addressed.  Thanks!
