+++ paper.bib
@@ -1,6 +1,6 @@
 @article{curcic2019parallel,
       author = {Curcic, Milan},
-      title = {A parallel Fortran framework for neural networks and deep learning},
+      title = {A parallel {F}ortran framework for neural networks and deep learning},
       year = {2019},
       issue_date = {December 2018},
       publisher = {Association for Computing Machinery},
@@ -10,8 +10,8 @@
       issn = {1061-7264},
       url = {https://doi.org/10.1145/3323057.3323059},
       doi = {10.1145/3323057.3323059},
-      abstract = {This paper describes neural-fortran, a parallel Fortran framework for neural networks and deep learning. It features a simple interface to construct feed-forward neural networks of arbitrary structure and size, several activation functions, and stochastic gradient descent as the default optimization algorithm. Neural-fortran also leverages the Fortran 2018 standard collective subroutines to achieve data-based parallelism on shared- or distributed-memory machines. First, I describe the implementation of neural networks with Fortran derived types, whole-array arithmetic, and collective sum and broadcast operations to achieve parallelism. Second, I demonstrate the use of neural-fortran in an example of recognizing hand-written digits from images. Finally, I evaluate the computational performance in both serial and parallel modes. Ease of use and computational performance are similar to an existing popular machine learning framework, making neural-fortran a viable candidate for further development and use in production.},
-      journal = {SIGPLAN Fortran Forum},
+      abstract = {This paper describes neural-fortran, a parallel {F}ortran framework for neural networks and deep learning. It features a simple interface to construct feed-forward neural networks of arbitrary structure and size, several activation functions, and stochastic gradient descent as the default optimization algorithm. Neural-fortran also leverages the {F}ortran 2018 standard collective subroutines to achieve data-based parallelism on shared- or distributed-memory machines. First, I describe the implementation of neural networks with {F}ortran derived types, whole-array arithmetic, and collective sum and broadcast operations to achieve parallelism. Second, I demonstrate the use of neural-fortran in an example of recognizing hand-written digits from images. Finally, I evaluate the computational performance in both serial and parallel modes. Ease of use and computational performance are similar to an existing popular machine learning framework, making neural-fortran a viable candidate for further development and use in production.},
+      journal = {SIGPLAN {F}ortran Forum},
       month = {mar},
       pages = {4–21},
       numpages = {18}
@@ -66,7 +66,7 @@
 
 @article{reid2018new,
       author = {Reid, John},
-      title = {The new features of Fortran 2018},
+      title = {The new features of {F}ortran 2018},
       year = {2018},
       issue_date = {April 2018},
       publisher = {Association for Computing Machinery},
@@ -76,7 +76,7 @@
       issn = {1061-7264},
       url = {https://doi.org/10.1145/3206214.3206215},
       doi = {10.1145/3206214.3206215},
-      journal = {SIGPLAN Fortran Forum},
+      journal = {SIGPLAN {F}ortran Forum},
       month = {apr},
       pages = {5–43},
       numpages = {39}
