### 2024-07-04 #6492 @nedtaylor
@crvernon I’m having issue finding documentation on how to remove brackets around citations inside bracketed statements. Do you know how to tackle that? If not, I might just rephrase it to move it outside of brackets instead.

Also, just to confirm, I think the second to last issue is line 64, not 54. Am I correct?

I cannot tick off the list, so I’ll make my own comment with them.
