### 2024-07-04 20d0908
Try to fix reference brackets
