### 2024-07-04 #6492 @nedtaylor
@crvernon I have addressed the errors in the paper and the corrections can be found in the `paper` branch.
