### 2024-07-04 3fee86b
Create CITATION.cff
