### 2024-07-05 #6492 @crvernon
🥳 Congratulations on your new publication @nedtaylor! Many thanks to @HaoZeke for editing and @milancurcic, @awvwgk, and @jrybarczyk  for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
